% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Biting.R
\name{add2Q_Bites}
\alias{add2Q_Bites}
\title{MACRO \code{Human} Method: addBites2Q}
\usage{
add2Q_Bites(tEvent, PAR)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me! a method for \code{\link{Human}}
}
\examples{
some_function()
}
