% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{add2Q_endPfMOI}
\alias{add2Q_endPfMOI}
\title{PfMOI \code{Human} Event: Add PfMOI Clearance Event to Event Queue}
\usage{
add2Q_endPfMOI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{time of infection}

\item{PAR}{named list
\itemize{
\item PfID: PfID of the infection to end
}}
}
\description{
Add PfMOI infection clearance event to the event queue.
This method is called from \code{\link{infectHumanPfMOI}}
This method adds event \code{\link{event_endPfMOI}} to the event queue.
This method is bound to \code{Human$add2Q_endPfMOI()}
}
