% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{add2Q_endprophylaxisPfSI}
\alias{add2Q_endprophylaxisPfSI}
\title{PfSI \code{Human} Event: Add PfSI End of Chemoprophylaxis Event to Event Queue}
\usage{
add2Q_endprophylaxisPfSI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{time of event}

\item{PAR}{\code{NULL}}
}
\description{
Add PfSI end of chemoprophylaxis event to the event queue.
This method is called from \code{\link{treatPfSI}}
This method adds event \code{\link{event_endprophylaxisPfSI}} to the event queue.
This method is bound to \code{Human$add2Q_endprophylaxisPfSI()}
}
