% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{add2Q_feverPfMOI}
\alias{add2Q_feverPfMOI}
\title{PfMOI \code{Human} Event: Add PfMOI Fever Event to Event Queue}
\usage{
add2Q_feverPfMOI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{time of fever}

\item{PAR}{\code{NULL}}
}
\description{
Add PfMOI fever event to the event queue.
This method is called from \code{\link{infectHumanPfMOI}}
This method adds event \code{\link{event_feverPfMOI}} to the event queue.
This method is bound to \code{Human$add2Q_feverPfMOI()}
}
