% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{add2Q_gsvaccinatePfMOI}
\alias{add2Q_gsvaccinatePfMOI}
\title{PfMOI \code{Human} Event: Add PfMOI GS Vaccination Event to Event Queue}
\usage{
add2Q_gsvaccinatePfMOI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{time of vaccination}

\item{PAR}{\code{NULL}}
}
\description{
Add PfMOI GS vaccination event to the event queue.
This method is called from \code{\link{queueVaccination_SimBitePfMOI}}
This method adds event \code{\link{event_gsvaccinatePfMOI}} to the event queue.
This method is bound to \code{Human$add2Q_gsvaccinatePfMOI()}
}
