% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{add2Q_gsvaccinatePfSI}
\alias{add2Q_gsvaccinatePfSI}
\title{PfSI \code{Human} Event: Add PfSI GS Vaccination Event to Event Queue}
\usage{
add2Q_gsvaccinatePfSI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{time of vaccination}

\item{PAR}{\code{NULL}}
}
\description{
Add PfSI GS vaccination event to the event queue.
This method is called from \code{\link{queueVaccination_SimBitePfSI}}
This method adds event \code{\link{event_gsvaccinatePfSI}} to the event queue.
This method is bound to \code{Human$add2Q_gsvaccinatePfSI()}
}
