% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{add2Q_gswanePfSI}
\alias{add2Q_gswanePfSI}
\title{PfSI \code{Human} Event: Add PfSI GS Waning Protection Event to Event Queue}
\usage{
add2Q_gswanePfSI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{time of vaccination}

\item{PAR}{\code{NULL}}
}
\description{
Add PfSI GS waning protection event to the event queue.
This method is called from \code{\link{gsvaccinatePfSI}}
This method adds event \code{\link{event_gswanePfSI}} to the event queue.
This method is bound to \code{Human$add2Q_gswanePfSI()}
}
