% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{add2Q_infectHumanPfMOI}
\alias{add2Q_infectHumanPfMOI}
\title{PfMOI \code{Human} Event: Add PfMOI Infection Event to Event Queue}
\usage{
add2Q_infectHumanPfMOI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{time of infection}

\item{PAR}{single clonal variant returned from \code{mosquitoPfMOI$get_InfectionIx()}}
}
\description{
Add a PfMOI infection to the event queue.
This method is called from \code{\link{infectiousBite_PfMOI}}
This method adds event \code{\link{event_infectHumanPfMOI}} to the event queue.
This method is bound to \code{Human$add2Q_infectHumanPfMOI()}
}
