% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{add2Q_pevaccinatePfMOI}
\alias{add2Q_pevaccinatePfMOI}
\title{PfMOI \code{Human} Event: Add PfMOI PE Vaccination Event to Event Queue}
\usage{
add2Q_pevaccinatePfMOI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{time of vaccination}

\item{PAR}{\code{NULL}}
}
\description{
Add PfMOI PE vaccination event to the event queue.
This method is called from \code{\link{queueVaccination_SimBitePfMOI}}
This method adds event \code{\link{event_pevaccinatePfMOI}} to the event queue.
This method is bound to \code{Human$add2Q_pevaccinatePfMOI()}
}
