% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{add2Q_pevaccinatePfSI}
\alias{add2Q_pevaccinatePfSI}
\title{PfSI \code{Human} Event: Add PfSI PE Vaccination Event to Event Queue}
\usage{
add2Q_pevaccinatePfSI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{time of vaccination}

\item{PAR}{\code{NULL}}
}
\description{
Add PfSI PE vaccination event to the event queue.
This method is called from \code{\link{queueVaccination_SimBitePfSI}}
This method adds event \code{\link{event_pevaccinatePfSI}} to the event queue.
This method is bound to \code{Human$add2Q_pevaccinatePfSI()}
}
