% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Movement.R
\name{add2Q_returnHome}
\alias{add2Q_returnHome}
\title{add2Q_returnHome}
\usage{
add2Q_returnHome(ixH, t)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
some_function()
}
