% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-SimBite.R
\name{add2Q_SimBitePfMOI}
\alias{add2Q_SimBitePfMOI}
\title{PfMOI SimBite \code{Human} Event: Add PfMOI Simulated Bite Event to Event Queue}
\usage{
add2Q_SimBitePfMOI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{time of bite}

\item{PAR}{write me!}
}
\description{
Simulate a PfMOI infectious bite.
This method is bound to \code{Human$add2Q_SimBitePfMOI()}
This method adds event \code{\link{event_SimBitePfMOI}} to the event queue.
This method is bound to \code{Human$add2Q_SimBitePfMOI()}
}
