% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PFSI-PfSI.R
\name{add2Q_simbitePfSI}
\alias{add2Q_simbitePfSI}
\title{\code{add2Q_XX} Family: Add Simulated Bite to EventQ}
\usage{
add2Q_simbitePfSI(ixH, t, PAR = NULL)
}
\arguments{
\item{ixH}{index of human}

\item{t}{time the event will occur}

\item{PAR}{NULL; the function will generate a default pathogen object with \code{\link{makePf0}}}
}
\description{
\code{add2Q_simbitePfSI} adds a simulated bite to human ixH's eventQ.
For more information on how this family of functions works, see \url{http://smitdave.github.io/MASH-Development/docHumans.html#add2q_xx}
}
\examples{
add2Q_simbitePfSI(ixH, t, PAR = NULL)
}
