% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{add2Q_treatPfMOI}
\alias{add2Q_treatPfMOI}
\title{PfMOI \code{Human} Event: Add PfMOI Treatment Event to Event Queue}
\usage{
add2Q_treatPfMOI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{time of fever}

\item{PAR}{\code{NULL}}
}
\description{
Add PfMOI treatment event to the event queue.
This method is called from \code{\link{feverPfMOI}}
This method adds event \code{\link{event_treatPfMOI}} to the event queue.
This method is bound to \code{Human$add2Q_treatPfMOI()}
}
