% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{add2Q_treatPfSI}
\alias{add2Q_treatPfSI}
\title{PfSI \code{Human} Event: Add PfSI Treatment Event to Event Queue}
\usage{
add2Q_treatPfSI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{time of fever}

\item{PAR}{\code{NULL}}
}
\description{
Add PfSI treatment event to the event queue.
This method is called from \code{\link{feverPfSI}}
This method adds event \code{\link{event_treatPfSI}} to the event queue.
This method is bound to \code{Human$add2Q_treatPfSI()}
}
