% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-Queues.R
\name{add_MicroEggQ}
\alias{add_MicroEggQ}
\title{MICRO \code{AquaticSite} Method: Add a batch to the EggQ}
\usage{
add_MicroEggQ(eggBatch)
}
\arguments{
\item{eggBatch}{a single egg batch, see \code{\link{eggBatch}} for the egg batch structure}
}
\description{
This method adds an egg batch to an EggQ, see \code{\link{eggBatch}} for the egg batch structure; it uses \code{\link{empty_MicroEggQ}} to find empty indices and
calls \code{\link{extend_MicroEggQ}} if no empty slots are found.
This method is bound to \code{AquaticSite$add_EggQ()}.
}
