% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-Queues.R
\name{add_MicroImagoQ}
\alias{add_MicroImagoQ}
\title{MICRO \code{\link{AquaticSite}} Method: Add emerging adults to the ImagoQ}
\usage{
add_MicroImagoQ(newImago)
}
\arguments{
\item{newImago}{a single emerging adults structure, see \code{\link{newImago}} for the emerging adults structure}
}
\description{
This method adds emerging adults to an ImagoQ, see \code{\link{newImago}} for the emerging adults structure; it uses \code{\link{empty_MicroImagoQ}} to find empty indices and
calls \code{\link{extend_MicroImagoQ}} if no empty slots are found.
This method is bound to \code{AquaticSite$add_ImagoQ()}.
}
