% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Site-RiskList.R
\name{add_riskList}
\alias{add_riskList}
\title{MICRO \code{FeedingSite} Method: Add a Person's Risk to a Risk List}
\usage{
add_riskList(who, pTm, w)
}
\arguments{
\item{who}{integer index of this \code{\link{Human}} myID slot}

\item{pTm}{this human's time at risk}

\item{w}{this human's biting weight}
}
\description{
Add risk information for a single human to a site's risk list. This function is a method for \code{\link{FeedingSite}}.
}
