% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUATIC-ECOLOGY.R
\name{addAdult}
\alias{addAdult}
\title{Insert Adults into ImagoQ}
\usage{
addAdult(ixQ, ix, N, tm, dam, sire)
}
\arguments{
\item{ixQ}{index of position in ImagoQ to zero out}

\item{ix}{index of aquatic habitat of emergence}

\item{N}{number of emerging adults from this element of queue}

\item{tm}{time of emergence}

\item{dam}{id of mother}

\item{sire}{id of father}

\item{ix}{index of aquatic habitat}
}
\value{
modifies global LANDSCAPE$aquaSites[[ix]]$ImagoQ[[ixQ]]
}
\description{
This function is a low-level utility to add a cohort of adults to ImagoQ. It is called by addAdults2Q.
}
\examples{
addAdult(ixQ,N,tm,ix,dam,sire)
}
