% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-EL4P-Methods.R
\name{addCohort_MicroEL4P}
\alias{addCohort_MicroEL4P}
\title{MICRO \code{\link{Landscape}} Method: Get Emerging Adults from ImagoQ and Zero out ImagoQ}
\usage{
addCohort_MicroEL4P()
}
\description{
Grab emerging adult batches where tEmerge <= tNow and add to the \code{\link{MicroMosquitoPopFemale}}.
This function fills a generic for \code{\link{simMICRO_oneStep}}.
\itemize{
\item This method is bound to \code{Landscape$addCohort()}
}
}
