% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-EL4P-Methods.R
\name{addCohort_MicroEL4PSite}
\alias{addCohort_MicroEL4PSite}
\title{MICRO \code{\link{Landscape}} Method: Get Emerging Adults from ImagoQ and Zero out ImagoQ}
\usage{
addCohort_MicroEL4PSite(tNow)
}
\arguments{
\item{tNow}{current global time of \code{\link{MicroTile}}}
}
\description{
Grab emerging adult batches where tEmerge <= tNow and add to the \code{\link{MicroMosquitoPopFemale}}.
This is a helper function for \code{\link{addCohort_MicroEmerge}}.
\itemize{
\item This method is bound to \code{AquaticSite$addCohort_MicroEL4PSite()}.
}
}
