% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-Emerge-Methods.R
\name{addCohort_MicroEmergeSite}
\alias{addCohort_MicroEmergeSite}
\title{MICRO \code{\link{Landscape}} Method: Get Emerging Adults from ImagoQ and Zero out ImagoQ}
\usage{
addCohort_MicroEmergeSite(tNow)
}
\description{
Grab emerging adult batches where tEmerge <= tNow and add to the \code{\link{MicroMosquitoPopFemale}}.
This is a helper function for \code{\link{addCohort_MicroEmerge}}.
\itemize{
\item This method is bound to \code{AquaticSite$addCohort_MicroEmergeSite()}.
}
}
