% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Site-AquaticEcology.R
\name{allocEggQ}
\alias{allocEggQ}
\title{Allocate EggQ}
\usage{
allocEggQ(N)
}
\arguments{
\item{N}{size of empty queue}
}
\value{
empty EggQ
}
\description{
This function allocates an empty EggQ to an object of class \code{\link{AquaticSite}}
}
\examples{
allocEggQ(N=N)
}
