% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Structures.R
\name{allocImagoQ}
\alias{allocImagoQ}
\title{Allocate ImagoQ}
\usage{
allocImagoQ(N)
}
\arguments{
\item{N}{size of empty queue}
}
\value{
empty ImagoQ
}
\description{
This function allocates an empty ImagoQ
}
\examples{
allocImagoQ(N=N)
}
