% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Emerge-Main.R
\name{aquaEmerge_makeLambda}
\alias{aquaEmerge_makeLambda}
\title{Generate Seasonal Emergence (Lambda) for \code{Emerge} model of Aquatic Ecology}
\usage{
aquaEmerge_makeLambda(lambda, lambdaWeight = NULL, offset = NULL)
}
\arguments{
\item{lambda}{vector of length equal to number of sites/patches where each element is the number of emerging adult females per human per day averaged over one year}

\item{lambdaWeight}{vector of weights applied to each site}

\item{offset}{seasonal offset in peak emergence}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
aquaEmerge_makeLambda(lambda = c(2,3,4))
}
