% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Utilities.R
\name{AquaEmergeLambdaPlot_utility}
\alias{AquaEmergeLambdaPlot_utility}
\title{Plot Lambda for Emerge Module}
\usage{
AquaEmergeLambdaPlot_utility(lambda)
}
\arguments{
\item{lambda}{a list of vectors}
}
\description{
Given a list of vectors for lambda across sites, plot emergence.
}
\examples{
AquaEmergeLambdaPlot_utility(lambda = makeLambda_Macro(aquaPars = list(N = 10, lambda = 10)))
}
