% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Site-Class.R
\docType{class}
\name{AquaticSite}
\alias{AquaticSite}
\title{MICRO-Site Aquatic Habitat Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
AquaticSite
}
\description{
This is a generic site blah blah ...
 This class inherits from \code{\link{Site}} class.
 Defines the fitting algorithm for a regression model \code{A[j] ~ W + ...}.
}
\details{
\itemize{
\item{\code{reg}} - .
\item{\code{outvar}} - .
\item{\code{levels}} - .
\item{\code{nbins}} - .
}
}
\section{Methods}{

\describe{
  \item{\code{new(reg, DataStorageClass.g0, ...)}}{...}
  \item{\code{fit(data)}}{...}
  \item{\code{predict(newdata)}}{...}
  \item{\code{predictAeqa(newdata)}}{...}
}
}

\section{Active Bindings}{

\describe{
  \item{\code{cats}}{...}
}
}

\keyword{R6}
\keyword{class}
