% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Landscape-Parameters.R
\name{betaRootA}
\alias{betaRootA}
\title{Find shape1 (alpha) Parameter of Beta Distribution for Given Mean}
\usage{
betaRootA(mean, beta = 1)
}
\arguments{
\item{mean}{the mean of the beta distribution}

\item{beta}{beta parameter of the beta distribution}
}
\value{
numeric value
}
\description{
Given a user-specified mean value, \code{betaRootA} uses \code{\link{uniroot}} to find the shape1 (alpha)
parameter of the distribution that will give that mean. Parameter shape2 may also be given as user input, but has
a default value of 1.
}
\examples{
betaRootA(mean = 0.95, beta = 20)
}
