% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Landscape-Parameters.R
\name{betaRootB}
\alias{betaRootB}
\title{Find shape2 (beta) Parameter of Beta Distribution for Given Mean}
\usage{
betaRootB(mean, alpha = 1)
}
\arguments{
\item{mean}{the mean of the beta distribution}

\item{alpha}{= 1 alpha parameter of beta distribution}
}
\value{
numeric value
}
\description{
Given a user-specified mean value, \code{betaRootB} uses \code{\link{uniroot}} to find the shape2 (beta)
parameter of the distribution that will give that mean. Parameter shape1 may also be given as user input, but has
a default value of 1.
}
\examples{
betaRootB(mean = 0.05, alpha = 1)
}
