% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Calibration.R
\name{burnin_GEL4P}
\alias{burnin_GEL4P}
\title{Run a Single Aquatic Population through Burnin}
\usage{
burnin_GEL4P(ix, psi, pop, PAR, tMax = 800)
}
\arguments{
\item{ix}{index of this aquatic population}

\item{psi}{value of psi}

\item{pop}{EL4P aquatic population object (see \code{link\{EL4P})}

\item{PAR}{aquatic parameters passed from \code{\link{makePAR_EL4P}}}

\item{tMax}{800; time to run the model through burnin period.}
}
\value{
the aquatic population
}
\description{
Run a single aquatic population through a burnin period (until initial oscillations dampen).
}
