% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Utility.R
\name{calculateBionomics_utility}
\alias{calculateBionomics_utility}
\title{M-BITES Utility: Calculate Ross-MacDonald Bionomics from Mosquito Histories}
\usage{
calculateBionomics_utility(history)
}
\arguments{
\item{history}{list of mosquito histories, if data exists in JSON format see \code{\link{importMicroMosquitoHistory_utility}} to import}
}
\description{
From \code{MicroTile$simCohort} or microsimulation runs, calculate basic mosquito bionomics.
}
