% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Calibration.R
\name{checkDX_GEL4P}
\alias{checkDX_GEL4P}
\title{Run a Single Aquatic Population with Simulated Adult Dynamics and Output Daily Lambda}
\usage{
checkDX_GEL4P(ix, psi, pop, PAR, tMax = 800)
}
\arguments{
\item{ix}{index of this aquatic population}

\item{psi}{value of psi}

\item{pop}{EL4P aquatic population object (see \code{link\{EL4P})}

\item{PAR}{aquatic parameters passed from \code{\link{makePAR_EL4P}}}

\item{tMax}{800; time to run the model through burnin period.}
}
\value{
vector of lambda values over time \code{tMax}
}
\description{
Run a single aquatic population's dynamics using simulated adult dynamics and egg laying from derived Ross-MacDonald parameters (see \code{\link{cohortBionomics}}).
Adult dynamics follow the simple equation \eqn{( e^{-1/lifespan} *M[ix]) + \lambda} where lambda parameter is from the aquatic population being simulated.
This function will output the record of daily lambda emergence, used to check variance in \code{\link{run2Eq_GEL4P}}.
}
