% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-Queues.R
\name{clear_MicroEggQ}
\alias{clear_MicroEggQ}
\title{MICRO \code{\link{AquaticSite}} Method: Clear the EggQ}
\usage{
clear_MicroEggQ()
}
\description{
Clear out all populated slots in an EggQ for the EL4P module of Aquatic Ecology; populated slots are found by calling \code{\link{full_MicroEggQ}}.
This method should be called after moving batches from the EggQ to EL4P object (whatever function does this should only move those batches that are ready to go, time-wise). WHEN YOU WRITE THESE FUNCTIONS UPDATE THESE DOCS!
This method is bound to \code{AquaticSite$clear_EggQ()}.
}
