% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-Queues.R
\name{clear_MicroImagoQ}
\alias{clear_MicroImagoQ}
\title{MICRO \code{\link{AquaticSite}} Method: Clear the ImagoQ}
\usage{
clear_MicroImagoQ()
}
\description{
Clear out all populated slots in an ImagoQ for the EL4P module of Aquatic Ecology; populated slots are found by calling \code{\link{full_MicroImagoQ}}.
This method should be called after calling adult emergence from the ImagoQ to \code{\link{MicroMosquitoPopFemale}} or \code{\link{MicroMosquitoPopMale}} object (whatever function does this should only move those adults that are ready to go, time-wise). WHEN YOU WRITE THESE FUNCTIONS UPDATE THESE DOCS!
This method is bound to \code{AquaticSite$clear_ImagoQ()}.
}
