% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Landscape-Utilities.R
\name{colLuminosity_utility}
\alias{colLuminosity_utility}
\title{Brighten or Darken Colors}
\usage{
colLuminosity_utility(color, factor, bright, alpha = NULL)
}
\arguments{
\item{color}{vector of hcl colors}

\item{factor}{factor to brighten or darken colors}

\item{bright}{logical variable to brighten or darken}

\item{alpha}{opacity}
}
\value{
a vector of colors in hex format
}
\description{
With input of hcl colors (hex code), brighten or darken by a factor
}
\examples{
colLuminosity_utility(color=MASH::ggCol_utility(n=5), factor = 1.15, bright = TRUE)
}
