% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILITIES-Human.R
\name{convertHumanEvent_PfSI}
\alias{convertHumanEvent_PfSI}
\title{Extract PfSI History from Human Object}
\usage{
convertHumanEvent_PfSI(HUMANS)
}
\arguments{
\item{HUMANS}{the \code{HUMANS} object}
}
\value{
list of histories
\itemize{
\item eventT: times events occur
\item events: events
}
}
\description{
Essentially go through the process of exporting via \code{\link{writeHumanEvent_PfSI}} and importing via \code{\link{importHumanEvent_PfSI}}
without exporting data to .json to return formatted human event histories.
}
\examples{
writeHumanEvent(directory, fileName)
}
