% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Site-AquaticEcology.R
\name{eggBatch}
\alias{eggBatch}
\title{Empty Egg Batch Slot for EggQ}
\usage{
eggBatch(N = 0L, tm = 0, ix = 0L, dam = 0L, sire = 0L)
}
\arguments{
\item{N}{number of eggs in this batch}

\item{tm}{time of oviposition}

\item{ix}{index of aquatic habitat}

\item{dam}{id of mother}

\item{sire}{id of father}
}
\value{
a list of named elements
}
\description{
This function is a low-level utility to generate a null placeholder value for an egg batch in EggQ.
It is called by \code{\link{allocEggQ}} and \code{\link{extendEggQ}}.
}
\examples{
eggBatch(N=0,tm=0,ix=0,dam=0,sire=0)
}
