% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{EggQ}
\alias{EggQ}
\title{MICRO Egg Queue}
\usage{
EggQ()
}
\description{
EggQ
}
\section{Methods}{

\itemize{
\item clear_EggQ:
\item clear_EggQTime:
\item add_EggQ:
\item track_EggQ:
\item get_N:
\item set_N:
\item get_EggQ:
\item get_EggQTime:
}
}

\section{Fields}{

\itemize{
\item a field:
}
}

