% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-EL4P-Calibrate.R
\name{EL4P.Landscape.Fit}
\alias{EL4P.Landscape.Fit}
\title{Fit EL4P Aquatic Ecology Module to \code{\link{Landscape}}}
\usage{
EL4P.Landscape.Fit()
}
\description{
Fit the EL4P Aquatic Ecology module on the exact LANDSCAPE to match a desired level of daily emergence at equilibrium.
This means that the site-specific density dependent mortality parameter \code{psi} will be fit
to K for each aquatic habitat on the landscape, and then each site will be run to equilibrium. Compare with \code{\link{setupAquaPop_EL4PsamplePoints}} which will
fit \code{psi} based on a sampling grid of K.
}
