% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-EL4P-Calibrate.R
\name{EL4P.Mesh.Fit}
\alias{EL4P.Mesh.Fit}
\title{Fit EL4P Aquatic Ecology Module to Mesh of K}
\usage{
EL4P.Mesh.Fit(mesh_N, EL4P_PAR, var_tol = 0.1, plot = FALSE)
}
\arguments{
\item{mesh_N}{number of sample points of K (density of mesh of K values used to fit psi)}

\item{EL4P_PAR}{parameters from \code{\link{EL4P.Parameters}}}

\item{var_tol}{target minimum variance in lambda for emergence for equilibrium to be assumed}

\item{plot}{produce diagnostic plots of fitting algorithm}
}
\value{
named list
\itemize{
\item equilibriumPops: equilibrium aquatic populations
\item psi: fitted values of psi
\item alpha: sampled values of alpha
\item K: sampled values of K
\item meshK: sampled values of K
\item psi2K_cf: slope coefficient between 1/psi and K
}
}
\description{
Fit the EL4P Aquatic Ecology module on a sampling grid of K values to match a desired level of daily emergence at equilibrium.
This means that the site-specific density dependent mortality parameter \code{psi} will be fit
to K based on a sampling grid of values for K in log-space. If \code{plot = TRUE}, the linear regression of \code{psi} against logged values of K
should show exact linear dependence, indicating fitted \code{psi} will produce desired level of lambda at equilibrium. Compare with \code{\link{setupAquaPop_EL4Pexact}} which will
fit \code{psi} based on an exact LANDSCAPE. This will also return the coefficients of a linear regression of K on psi (see \code{\link{psi2K_cf}}) to give the functional relationship between K and psi.
}
