% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-EL4P-Calibrate.R
\name{EL4P.Parameters}
\alias{EL4P.Parameters}
\title{Initialize EL4P Aquatic Ecoogy Module Parameters}
\usage{
EL4P.Parameters(nAqua, nHumans, R0, eqAqua, EIP, lifespan, G, nu, S,
  alpha_a = 0.8, alpha_sd = 0.004, p = 0.9, K_a = 1, K_b = 1,
  b = 0.55, c = 0.15, r = 1/38)
}
\arguments{
\item{nAqua}{number of aquatic habitats on landscape}

\item{nHumans}{number of humans on landscape}

\item{R0}{desired pathogen transmission level at equilibrium}

\item{eqAqua}{vector of probability of oviposition at each aquatic habitat}

\item{EIP}{length of the entomological incubation period}

\item{lifespan}{average lifespan of mosquito}

\item{G}{mean total lifetime egg production of adult female}

\item{nu}{mean egg batch size for single oviposition}

\item{S}{stability index; number of bites on humans over the average mosquito lifespan}

\item{alpha_a}{density independent survival parameter \code{alpha} from pupae to adult \deqn{-log(alpha_a^{\frac{1-p}{5}})}}

\item{alpha_sd}{standard deviation of density independent survival parameter \code{alpha}}

\item{p}{expected fraction of cohort that advances to next life stage (1/p is expected time spent in each stage L1,L2,L3,L4,P)}

\item{K_a}{shape parameter for gamma distributed weights on K}

\item{K_b}{scale parameter for gamma distributed weights on K}

\item{b}{mosquito to human transmission efficiency}

\item{c}{human to mosquito transmission efficiency}

\item{r}{rate of recovery in humans (1/average length of infection)}
}
\description{
Generate named list of EL4P parameters, and calculates lambda (see \code{\link{calcLambda_MicroEL4P}} for details).
}
