% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{EL4P}
\alias{EL4P}
\title{Aquatic Ecology EL4P Class}
\usage{
EL4P(numGenotypes, psi_new, alpha_new, p_new)
}
\description{
Initialize an EL4P pool inside an aquatic site or for a well-mixed macrosimulation tile.
This class is nested inside \code{\link{AquaticSite}} or \code{\link{MacroTile}}.
Note that genotype corresponding to index 0 is always the wild type.
}
\section{Methods}{

\itemize{
\item get_N:
\item set_N:
\item get_who:
\item push_who:
\item get_pTm:
\item set_pTm:
\item get_w:
\item push_w:
\item set_nOther:
\item get_nOther:
\item add_OtherHost:
\item get_OtherHost:
}
}

\section{Fields}{

\itemize{
\item a field:
}
}

