% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Macro-Emerge.R
\name{emergingAdults}
\alias{emergingAdults}
\title{MACRO: Calculate Emerging Adults from ImagoQ for \code{MacroPatch}}
\usage{
emergingAdults()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me! does this for all patches
}
\examples{
some_function()
}
