% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Macro-Emerge.R
\name{emergingAdults_MACRO}
\alias{emergingAdults_MACRO}
\title{MACRO: Calculate Emerging Adults from ImagoQ for \code{MacroPatch}}
\usage{
emergingAdults_MACRO()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me! does this for all patches. Goes from ImagoQ to a vector of adults to be born, and zeros out
}
\examples{
some_function()
}
