% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-Emerge.R
\name{emergingAdults_MicroEmerge}
\alias{emergingAdults_MicroEmerge}
\title{MICRO \code{\link{Landscape}} Method: Get Emerging Adults from ImagoQ and Zero out ImagoQ}
\usage{
emergingAdults_MicroEmerge()
}
\value{
does stuff
}
\description{
This method is bound to \code{Landscape$emergingAdults_Emerge()}
}
\examples{
some_function()
}
