% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-Queues.R
\name{empty_MicroEggQ}
\alias{empty_MicroEggQ}
\title{MICRO \code{\link{AquaticSite}} Method: Get indices of empty EggQ slots}
\usage{
empty_MicroEggQ()
}
\description{
This method finds empty slots in the EggQ; if all slots are full it returns 0. It is the complement of \code{\link{full_MicroEggQ}}
This method is bound to \code{AquaticSite$empty_EggQ()}.
}
