% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-Queues.R
\name{empty_MicroImagoQ}
\alias{empty_MicroImagoQ}
\title{MICRO \code{\link{AquaticSite}} Method: Get indices of empty ImagoQ slots}
\usage{
empty_MicroImagoQ()
}
\description{
This method finds empty slots in the ImagoQ; if all slots are full it returns 0. It is the complement of \code{\link{full_MicroImagoQ}}
This method is bound to \code{AquaticSite$empty_ImagoQ()}.
}
