% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{endprophylaxisPfMOI}
\alias{endprophylaxisPfMOI}
\title{PfMOI \code{Human} Event: PfMOI End of Chemoprophylaxis Event}
\usage{
endprophylaxisPfMOI(tEvent, PAR)
}
\arguments{
\item{tEvent}{time to end chemoprophylaxis}

\item{PAR}{\code{NULL}}
}
\description{
End PfMOI chemoprophylaxis protection.
This method is bound to \code{Human$endprophylaxisPfMOI()}
}
