% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{endprophylaxisPfSI}
\alias{endprophylaxisPfSI}
\title{PfSI \code{Human} Event: PfSI End of Chemoprophylaxis Event}
\usage{
endprophylaxisPfSI(tEvent, PAR)
}
\arguments{
\item{tEvent}{time to end chemoprophlaxis}

\item{PAR}{\code{NULL}}
}
\description{
End PfSI chemoprophlaxis protection.
This method is bound to \code{Human$endprophylaxisPfSI()}
}
