% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{event_endPfMOI}
\alias{event_endPfMOI}
\title{PfMOI \code{Human} Event: Generate PfMOI Clearance Event}
\usage{
event_endPfMOI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{time of clearance}

\item{PAR}{named list
\itemize{
\item PfID: PfID of the infection to end
}}
}
\description{
Generate PfMOI clearance event to place in event queue.
This method is called from \code{\link{add2Q_endPfMOI}}
This method is bound to \code{Human$event_endPfMOI()}
\itemize{
\item tag: \code{\link{endPfMOI}}
\item tEvent: treatment time is calculated as tTreat = tEvent + \code{\link{PfMOI_ttClearPf}}
}
}
