% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{event_endprophylaxisPfMOI}
\alias{event_endprophylaxisPfMOI}
\title{PfMOI \code{Human} Event: Generate PfMOI End of Chemoprophylaxis Event}
\usage{
event_endprophylaxisPfMOI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{time to end chemoprophylaxis}

\item{PAR}{\code{NULL}}
}
\description{
Generate PfMOI end of chemoprophylaxis event to place in event queue.
This method is called from \code{\link{add2Q_endprophylaxisPfMOI}}
This method is bound to \code{Human$event_endprophylaxisPfMOI()}
\itemize{
\item tag: \code{\link{endprophylaxisPfMOI}}
\item tEvent: treatment time is calculated as tSusceptible = tEvent + \code{\link{PfMOI_ttSusceptiblePf}}
}
}
