% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{event_endprophylaxisPfSI}
\alias{event_endprophylaxisPfSI}
\title{PfSI \code{Human} Event: Generate PfSI End of Chemoprophylaxis Event}
\usage{
event_endprophylaxisPfSI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{time to end chemoprophylaxis}

\item{PAR}{\code{NULL}}
}
\description{
Generate PfSI end of chemoprophylaxis event to place in event queue.
This method is called from \code{\link{add2Q_endprophylaxisPfSI}}
This method is bound to \code{Human$event_endprophylaxisPfSI()}
\itemize{
\item tag: \code{\link{endprophylaxisPfSI}}
\item tEvent: treatment time is calculated as tSusceptible = tEvent + \code{\link{PfSI_ttSusceptiblePf}}
}
}
