% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{event_feverPfMOI}
\alias{event_feverPfMOI}
\title{PfMOI \code{Human} Event: Generate PfMOI Fever Event}
\usage{
event_feverPfMOI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{time of fever}

\item{PAR}{\code{NULL}}
}
\description{
Generate PfMOI fever event to place in event queue.
This method is called from \code{\link{add2Q_feverPfMOI}}
This method is bound to \code{Human$event_feverPfMOI()}
\itemize{
\item tag: \code{\link{feverPfMOI}}
\item tEvent: treatment time is calculated as tTreat = tEvent + \code{\link{PfMOI_ttFeverPf}}
}
}
