% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{event_gsvaccinatePfMOI}
\alias{event_gsvaccinatePfMOI}
\title{PfMOI \code{Human} Event: Generate PfMOI GS Vaccination Event}
\usage{
event_gsvaccinatePfMOI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{begin gs protection}

\item{PAR}{\code{NULL}}
}
\description{
Generate PfMOI GS vaccination event to place in event queue.
This method is called from \code{\link{add2Q_gsvaccinatePfMOI}}
This method is bound to \code{Human$event_gsvaccinatePfMOI()}
\itemize{
\item tag: \code{\link{gsvaccinatePfMOI}}
}
}
