% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{event_gswanePfMOI}
\alias{event_gswanePfMOI}
\title{PfMOI \code{Human} Event: Generate PfMOI GS Waning Protection Event}
\usage{
event_gswanePfMOI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{end gs protection}

\item{PAR}{\code{NULL}}
}
\description{
Generate PfMOI GS waning protection event to place in event queue.
This method is called from \code{\link{add2Q_gsvaccinatePfMOI}}
This method is bound to \code{Human$event_gsvaccinatePfMOI()}
\itemize{
\item tag: \code{\link{gsvaccinatePfMOI}}
\item tEvent: loss of efficacy is calculated as tWane = tEvent + \code{\link{PfMOI_ttGSWanePf}}
}
}
