% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{event_infectHumanPfMOI}
\alias{event_infectHumanPfMOI}
\title{PfMOI \code{Human} Event: Generate PfMOI Infection Event}
\usage{
event_infectHumanPfMOI(tEvent, PAR)
}
\arguments{
\item{tEvent}{time of infection}

\item{PAR}{single clonal variant returned from \code{mosquitoPfMOI$get_clone()}}
}
\description{
Generate PfMOI infection event to place in event queue.
This method is called from \code{\link{add2Q_infectHumanPfMOI}}
This method is bound to \code{Human$event_infectHumanPfMOI()}
\itemize{
\item tag: \code{\link{infectHumanPfMOI}}
}
}
