% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{event_infectHumanPfSI}
\alias{event_infectHumanPfSI}
\title{PfSI \code{Human} Event: Generate PfSI Infection Event}
\usage{
event_infectHumanPfSI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{time of infection}

\item{PAR}{\code{NULL}}
}
\description{
Generate PfSI infection event to place in event queue.
This method is called from \code{\link{add2Q_infectHumanPfSI}}
This method is bound to \code{Human$event_infectHumanPfSI()}
\itemize{
\item tag: \code{\link{infectHumanPfSI}}
}
}
