% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{event_pevaccinatePfMOI}
\alias{event_pevaccinatePfMOI}
\title{PfMOI \code{Human} Event: Generate PfMOI PE Vaccination Event}
\usage{
event_pevaccinatePfMOI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{begin PE protection}

\item{PAR}{\code{NULL}}
}
\description{
Generate PfMOI PE vaccination event to place in event queue.
This method is called from \code{\link{add2Q_pevaccinatePfMOI}}
This method is bound to \code{Human$event_pevaccinatePfMOI()}
\itemize{
\item tag: \code{\link{pevaccinatePfMOI}}
}
}
