% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{event_pevaccinatePfSI}
\alias{event_pevaccinatePfSI}
\title{PfSI \code{Human} Event: Generate PfSI PE Vaccination Event}
\usage{
event_pevaccinatePfSI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{begin PE protection}

\item{PAR}{\code{NULL}}
}
\description{
Generate PfSI PE vaccination event to place in event queue.
This method is called from \code{\link{add2Q_pevaccinatePfSI}}
This method is bound to \code{Human$event_pevaccinatePfSI()}
\itemize{
\item tag: \code{\link{pevaccinatePfSI}}
}
}
