% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{event_pewanePfMOI}
\alias{event_pewanePfMOI}
\title{PfMOI \code{Human} Event: Generate PfMOI PE Waning Protection Event}
\usage{
event_pewanePfMOI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{end PE protection}

\item{PAR}{\code{NULL}}
}
\description{
Generate PfMOI PE waning protection event to place in event queue.
This method is called from \code{\link{add2Q_pevaccinatePfMOI}}
This method is bound to \code{Human$event_pevaccinatePfMOI()}
\itemize{
\item tag: \code{\link{pevaccinatePfMOI}}
\item tEvent: loss of efficacy is calculated as tWane = tEvent + \code{\link{PfMOI_ttPEWanePf}}
}
}
