% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-SimBite.R
\name{event_SimBitePfMOI}
\alias{event_SimBitePfMOI}
\title{PfMOI SimBite \code{Human} Event: Generate PfMOI Simulated Bite Event}
\usage{
event_SimBitePfMOI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{time of bite}

\item{PAR}{named list
\itemize{
\item mosquitoPfSI: object of class \code{\link{mosquitoPfMOI}}
}}
}
\description{
Generate PfMOI Simulated Bite event to place in event queue.
This method is called from \code{\link{add2Q_SimBitePfMOI}}
This method is bound to \code{Human$event_SimBitePfMOI()}
\itemize{
\item tag: \code{\link{SimBitePfMOI}}
}
}
