% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PFSI-PfSI.R
\name{event_simbitePfSI}
\alias{event_simbitePfSI}
\title{\code{event_XX} Family: Simulated Bite Event in EventQ}
\usage{
event_simbitePfSI(t, PAR)
}
\arguments{
\item{t}{time the event will occur}

\item{PAR}{passed from Pf0}

\item{ixH}{index of human}
}
\description{
This creates an event for the simulated bite in a human's eventQ and is called by \code{\link{add2Q_simbitePfSI}}.
For more information on how this family of functions works, see \url{http://smitdave.github.io/MASH-Development/docHumans.html#add2q_xx}
}
\examples{
getPfParent_SI(ixH)
}
