% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{event_treatPfMOI}
\alias{event_treatPfMOI}
\title{PfMOI \code{Human} Event: Generate PfMOI Treatment Event}
\usage{
event_treatPfMOI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{time of treatment}

\item{PAR}{\code{NULL}}
}
\description{
Generate PfMOI treatment event to place in event queue.
This method is called from \code{\link{add2Q_treatPfMOI}}
This method is bound to \code{Human$event_treatPfMOI()}
\itemize{
\item tag: \code{\link{treatPfMOI}}
\item tEvent: treatment time is calculated as tTreat = tEvent + \code{\link{PfMOI_ttTreatPf}}
}
}
