% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{event_treatPfSI}
\alias{event_treatPfSI}
\title{PfSI \code{Human} Event: Generate PfSI Treatment Event}
\usage{
event_treatPfSI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{time of treatment}

\item{PAR}{\code{NULL}}
}
\description{
Generate PfSI treatment event to place in event queue.
This method is called from \code{\link{add2Q_treatPfSI}}
This method is bound to \code{Human$event_treatPfSI()}
\itemize{
\item tag: \code{\link{treatPfSI}}
\item tEvent: treatment time is calculated as tTreat = tEvent + \code{\link{PfSI_ttTreatPf}}
}
}
