% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Biting.R
\name{expectedBites}
\alias{expectedBites}
\title{MACRO \code{Human} Method: expectedBites}
\usage{
expectedBites()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me! a method for \code{\link{Human}}
}
\examples{
some_function()
}
