% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-Queues.R
\name{extend_MicroImagoQ}
\alias{extend_MicroImagoQ}
\title{MICRO \code{\link{AquaticSite}} Method: Extend the ImagoQ}
\usage{
extend_MicroImagoQ()
}
\description{
This method extends the ImagoQ by 2 times its current length of empty emerging adult objects, see \code{\link{newImago}} for the emerging adult structure.
This method is bound to \code{AquaticSite$extend_ImagoQ()}.
}
