% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Macro-EL4P.R
\name{extend_PatchesImagoQ}
\alias{extend_PatchesImagoQ}
\title{Extend \code{MacroPatch} ImagoQ (Emerging Adults Queue)}
\usage{
extend_PatchesImagoQ(ixP, N = 10L)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Extend the ImagoQ for patch \code{ixP}
}
\examples{
some_function()
}
