% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Site-RiskList.R
\name{extend_riskList}
\alias{extend_riskList}
\title{MICRO \code{FeedingSite} Method: Extend a Risk List}
\usage{
extend_riskList()
}
\description{
Extend a site's risk list by one slot. This function is a method for \code{\link{FeedingSite}}.
}
