% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILITIES-Pathogen.R
\name{extractChain}
\alias{extractChain}
\title{Extract a Transmission Chain from PfTransmission}
\usage{
extractChain(ix, PfTransmission)
}
\arguments{
\item{ix}{index of mosquito to human transmission}
}
\description{
Extract a transmission chain generated by \code{\link{trackPfTransmission}}
}
