% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{feeding_PfMOI}
\alias{feeding_PfMOI}
\title{PfMOI \code{\link{MicroMosquitoFemale}} Method: Host Feeding}
\usage{
feeding_PfMOI()
}
\description{
The mosquito feeds on the host. Feeding occurs during code{MicroMosquitoFemale$humanEncounter()}.
Human to mosquito pathogen transmission occurs during host feeding.
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$feeding()}
}
}
