% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Site-Class.R
\docType{class}
\name{FeedingSite}
\alias{FeedingSite}
\title{MICRO Feeding Site Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
FeedingSite
}
\description{
This is a generic site blah blah ...
}
\details{
\itemize{
\item some deets
}
}
\section{Methods}{

\itemize{
\item a method
\item another method
\item even more methods
\item \code{\link{init_riskList}}: initialize a risk list for this site.
\item \code{\link{extend_riskList}}: extend a risk list for this site.
\item \code{\link{add_riskList}}: add to a risk list for this site.
}
}

\section{Active Bindings}{

\describe{
\item{\code{cats}}{...}
}
}

\keyword{R6}
\keyword{class}
