% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{feverPfMOI}
\alias{feverPfMOI}
\title{PfMOI \code{Human} Event: PfMOI Infection Event}
\usage{
feverPfMOI(tEvent, PAR)
}
\arguments{
\item{tEvent}{time of clearance}

\item{PAR}{\code{NULL}}
}
\description{
Clear a PfMOI infection corresponding to the given PfID.
This method is bound to \code{Human$endPfMOI()}
}
