% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-Queues.R
\name{full_MicroImagoQ}
\alias{full_MicroImagoQ}
\title{MICRO \code{\link{AquaticSite}} Method: Get indices of full ImagoQ slots}
\usage{
full_MicroImagoQ()
}
\description{
This method finds filled slots in the ImagoQ; if all slots are empty it returns \code{NULL}. It is the complement of \code{\link{empty_MicroImagoQ}}
This method is bound to \code{AquaticSite$full_ImagoQ()}.
}
