% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUEUEING-ParametricDistribution.R
\name{gammaRates}
\alias{gammaRates}
\title{Gamma Distributed Biting Rates}
\usage{
gammaRates(N, mean = 5, shape = 10)
}
\arguments{
\item{N}{number of humans}

\item{mean}{mean value}

\item{shape}{controls variance in gamma distributed rates; larger values produce distributions with lower variance.}
}
\value{
distribution of biting rates
}
\description{
Produce gamma-distributed rates for each person, with a chosen mean value of Poisson process parameterized by gamma rates, and selected shape (controls variance). This function calls \code{\link{rgamma}} to sample random numbers.
}
