% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMANS-Pf-Methods.R
\name{get_biteWeightPf}
\alias{get_biteWeightPf}
\title{Pf: Get \code{Human} Personal Biting Propensity}
\usage{
get_biteWeightPf()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
get_biteWeight()
}
