% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Macro-Emerge.R
\name{get_ImagoQ}
\alias{get_ImagoQ}
\title{Get \code{MacroPatch} ImagoQ (Emerging Adults Queue)}
\usage{
get_ImagoQ(ix = NULL)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
some_function()
}
