% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Methods.R
\name{get_location}
\alias{get_location}
\title{MACRO: Get \code{Human} location}
\usage{
get_location()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
some_function()
}
