% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-Queues.R
\name{get_MicroEggQ}
\alias{get_MicroEggQ}
\title{MICRO \code{AquaticSite} Method: Get the EggQ}
\usage{
get_MicroEggQ(ixQ = NULL)
}
\arguments{
\item{ixQ}{if \code{NULL} return the entire EggQ, else, return the slot \code{ixQ}}
}
\description{
Get either a single slot or the entire EggQ.
This method is bound to \code{AquaticSite$get_EggQ()}.
}
