% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Human-ActivitySpace.R
\name{get_MicroHuman_ActivitySpace}
\alias{get_MicroHuman_ActivitySpace}
\title{MICRO: Get \code{\link{Human}} ActivitySpace}
\usage{
get_MicroHuman_ActivitySpace()
}
\description{
This function is bound to \code{Human$get_ActivitySpace()}
}
