% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Human-Setup.R
\name{get_MicroHuman_FemalePopPointer}
\alias{get_MicroHuman_FemalePopPointer}
\title{MICRO: Get \code{\link{Human}} \code{\link{MicroMosquitoPopFemale}} Pointer}
\usage{
get_MicroHuman_FemalePopPointer()
}
\description{
This function is bound to \code{Human$get_FemalePopPointer()}
}
