% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Human-Setup.R
\name{get_MicroHuman_LandscapePointer}
\alias{get_MicroHuman_LandscapePointer}
\title{MICRO: Get \code{\link{Human}} \code{\link{Landscape}} Pointer}
\usage{
get_MicroHuman_LandscapePointer()
}
\description{
This function is bound to \code{Human$get_LandscapePointer()}
}
