% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Human-Setup.R
\name{get_MicroHuman_TilePointer}
\alias{get_MicroHuman_TilePointer}
\title{MICRO: Get \code{\link{Human}} \code{\link{MicroTile}} Pointer}
\usage{
get_MicroHuman_TilePointer()
}
\description{
This function is bound to \code{Human$get_TilePointer()}
}
