% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Human-Setup.R
\name{get_MicroHumanPop_FemalePopPointer}
\alias{get_MicroHumanPop_FemalePopPointer}
\title{MICRO: Get \code{\link{HumanPop}} \code{\link{MicroMosquitoPopFemale}} Pointer}
\usage{
get_MicroHumanPop_FemalePopPointer()
}
\description{
This function is bound to \code{HumanPop$get_FemalePopPointer()}
}
