% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Human-Setup.R
\name{get_MicroHumanPop_LandscapePointer}
\alias{get_MicroHumanPop_LandscapePointer}
\title{MICRO: Get \code{\link{HumanPop}} \code{\link{Landscape}} Pointer}
\usage{
get_MicroHumanPop_LandscapePointer()
}
\description{
This function is bound to \code{HumanPop$get_LandscapePointer()}
}
