% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Human-Setup.R
\name{get_MicroHumanPop_TilePointer}
\alias{get_MicroHumanPop_TilePointer}
\title{MICRO: Get \code{\link{HumanPop}} \code{\link{MicroTile}} Pointer}
\usage{
get_MicroHumanPop_TilePointer()
}
\description{
This function is bound to \code{HumanPop$get_TilePointer()}
}
