% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-Queues.R
\name{get_MicroImagoQ}
\alias{get_MicroImagoQ}
\title{MICRO \code{\link{AquaticSite}} Method: Get the ImagoQ}
\usage{
get_MicroImagoQ(ixQ = NULL)
}
\arguments{
\item{ixQ}{if \code{NULL} return the entire ImagoQ, else, return the slot \code{ixQ}}
}
\description{
Get either a single slot or the entire ImagoQ.
This method is bound to \code{AquaticSite$get_ImagoQ()}.
}
