% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Search-MicroKernel.R
\name{get_MicroKernel_movement}
\alias{get_MicroKernel_movement}
\title{MICRO Search Kernels: \code{\link{MicroMosquitoPop}} Access MvAll Object}
\usage{
get_MicroKernel_movement(ixS, state, inPointSet)
}
\description{
Replace generic \code{MicroMosquitoPop$get_movement()} method for MicroKernel module; it will be bound to \code{MicroMosquitoPop$get_movement()}
}
