% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEARCH-MicroKernel-Methods.R
\name{get_MicroKernel_movement_BRO}
\alias{get_MicroKernel_movement_BRO}
\title{MICRO Search Kernels: \code{\link{MicroTile}} Access MvAll Object for Female M-BITES BRO Lifecycle Model}
\usage{
get_MicroKernel_movement_BRO(ixS, state, inPointSet)
}
\description{
Replace generic \code{MicroTile$get_movement()} method for MicroKernel module, return a MvOb from \code{movementFemale} field of a microsimulation tile.
\itemize{
\item This method bound to \code{MicroTile$get_movement}
}
}
