% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Search-MicroKernel.R
\name{get_MicroKernel_movement_Male}
\alias{get_MicroKernel_movement_Male}
\title{MICRO Search Kernels: \code{\link{MicroMosquitoPopMale}} Access MvAll Object for M-BITES BROM Lifecycle Model}
\usage{
get_MicroKernel_movement_Male(ixS, state, inPointSet)
}
\description{
Replace generic \code{MicroMosquitoPopMale$get_movement()} method for MicroKernel module; it will be bound to \code{MicroMosquitoPopMale$get_movement()}
}
