% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-Emerge.R
\name{get_MicroLambda}
\alias{get_MicroLambda}
\title{MICRO \code{\link{AquaticSite}} Method: Get Lambda}
\usage{
get_MicroLambda(ix = NULL)
}
\arguments{
\item{ixQ}{if \code{NULL} return the entire vector of lambda, else, return the value corresponding to day \code{ix}}
}
\description{
Get either a single day lambda or entire vector
This method is bound to \code{AquaticSite$get_lambda()}.
}
