% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Methods.R
\name{get_MosquitoPointer}
\alias{get_MosquitoPointer}
\title{Get \code{Human} \code{MacroMosquitoPop} Pointer}
\usage{
get_MosquitoPointer()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
some_function()
}
