% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Macro-Emerge.R
\name{get_PatchesEggQ}
\alias{get_PatchesEggQ}
\title{Get \code{MacroPatch} EggQ (Egg Batches)}
\usage{
get_PatchesEggQ(ixP = NULL)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
some_function()
}
