% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-Macro-PfMethods.R
\name{get_patchPf}
\alias{get_patchPf}
\title{Pf: Get \code{MacroPatch} Patch Pf Population}
\usage{
get_patchPf(ix)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
get_patchPf()
}
