% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Macro-Emerge.R
\name{get_season}
\alias{get_season}
\title{Get \code{MacroPatch} Seasonal Emergence}
\usage{
get_season(ixP = NULL)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
some_function()
}
