% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Setup.R
\name{get_TilePointer}
\alias{get_TilePointer}
\title{Get \code{Human} \code{MacroTile} Pointer}
\usage{
get_TilePointer()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
some_function()
}
