% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Movement.R
\name{get_travel}
\alias{get_travel}
\title{Get \code{Human} travel}
\usage{
get_travel()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
some_function()
}
