% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Movement.R
\name{get_travelHistoryHumanPop}
\alias{get_travelHistoryHumanPop}
\title{Return \code{HumanPop} Travel History}
\usage{
get_travelHistoryHumanPop()
}
\value{
a list
}
\description{
Write me! Called as \code{self$get_travelHistory()} Defined in MACRO-Human-Movement.R
}
\examples{
some_function()
}
