% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-MosquitoPop-Methods.R
\name{getEIP_MacroMosquitoPop}
\alias{getEIP_MacroMosquitoPop}
\title{Set \code{MacroMosquitoPop} getEIP}
\usage{
getEIP_MacroMosquitoPop(tNow)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
get time dependent EIP
}
\examples{
some_function()
}
