% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Calibration.R
\name{getLambda}
\alias{getLambda}
\title{Calculate Lambda (Equilibrium Rate of Emergence)}
\usage{
getLambda(R0, summary, nH, P)
}
\arguments{
\item{R0}{desired intensity of pathogen transmission}

\item{summary}{output from \code{\link{cohortBionomics}} which can be calculated from a run of MASH, or, less laboriously, from \code{\link{MBITES.basic}}}

\item{nH}{number of humans on landscape (ie; number of humans in the MICRO patch)}

\item{P}{list of parameters from \code{\link{MBITES.PAR}}}
}
\value{
numeric value of lambda
}
\description{
Calculate parameter \code{lambda} (daily adult female emergence over entire landscape) required to sustain \emph{Plasmodium falciparum} transmission at given value of \emph{R0} based on
classical Ross-MacDonald assumptions.
}
\examples{
getLambda(R0, summary, nH, P)
}
