% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfPedigree.R
\name{getPfParent_SI}
\alias{getPfParent_SI}
\title{Return a pfid During Human to Mosquito Transmission}
\usage{
getPfParent_SI(ixH)
}
\arguments{
\item{ixH}{index of human}
}
\value{
a single pfid (integer)
}
\description{
For PfSI module this will only return a single pfid but in general the damID and sireID could be a vector of male and female gametocyte IDs
ingested by the mosquito during host to vector transmission during a blood meal.
}
\examples{
getPfParent_SI(ixH)
}
