% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Landscape-Utilities.R
\name{ggCol_utility}
\alias{ggCol_utility}
\title{Sample Equally from Color Space}
\usage{
ggCol_utility(n, alpha = 1)
}
\arguments{
\item{n}{number of colors to sample}

\item{alpha}{opacity}
}
\value{
a vector of colors in hex format
}
\description{
This function is a low-level utility to sample at equal points from the color wheel to produce ggplot2 color scheme.
}
\examples{
ggCol_utility(n=10, alpha = 0.5)
}
