% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Landscape-Utilities.R
\name{ggColOffset_utility}
\alias{ggColOffset_utility}
\title{Sample Equally from Color Space with Offset}
\usage{
ggColOffset_utility(n, offset, alpha = 1)
}
\arguments{
\item{n}{number of colors to sample}

\item{offset}{offset from 1}

\item{alpha}{opacity}
}
\value{
a vector of colors in hex format
}
\description{
This function is a low-level utility to sample at equal points from the color wheel to produce ggplot2 color scheme.
This is a modification of \code{\link{ggCol_utility}} with a simple offset where sampling begins at 1 + offset.
}
\examples{
ggColOffset_utility(n=10, offset = 2, alpha = 0.5)
}
