% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Travel.R
\name{go_Patch}
\alias{go_Patch}
\title{Set \code{Human} go_Patch}
\usage{
go_Patch(old, new)
}
\arguments{
\item{old}{old patch (integer index)}

\item{new}{new patch (integer index)}
}
\value{
does stuff
}
\description{
Update \code{\link{MacroPatch}} list of humanIDs when a human leaves a site \code{old} and goes to a site \code{new}
}
\examples{
some_function()
}
