% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{gsvaccinatePfMOI}
\alias{gsvaccinatePfMOI}
\title{PfMOI \code{Human} Event: PfMOI GS Vaccination Event}
\usage{
gsvaccinatePfMOI(tEvent, PAR)
}
\arguments{
\item{tEvent}{begin gs protection}

\item{PAR}{\code{NULL}}
}
\description{
Begin PfMOI GS vaccination protection.
This method is bound to \code{Human$endprophylaxisPfMOI()}
\itemize{
\item protection: infected human to mosquito transmission efficiency is modified by \code{gsBlockPf}, see \code{\link{PfMOI.Parameters}}
\item waning efficacy: queue \code{\link{add2Q_gswanePfMOI}}
}
}
