% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{gsvaccinatePfSI}
\alias{gsvaccinatePfSI}
\title{PfSI \code{Human} Event: PfSI GS Vaccination Event}
\usage{
gsvaccinatePfSI(tEvent, PAR)
}
\arguments{
\item{tEvent}{begin gs protection}

\item{PAR}{\code{NULL}}
}
\description{
Begin PfSI GS vaccination protection.
This method is bound to \code{Human$endprophylaxisPfSI()}
\itemize{
\item protection: infected human to mosquito transmission efficiency is modified by \code{gsBlockPf}, see \code{\link{PfSI.Parameters}}
\item waning efficacy: queue \code{\link{add2Q_gswanePfSI}}
}
}
