% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{gswanePfMOI}
\alias{gswanePfMOI}
\title{PfMOI \code{Human} Event: PfMOI GS Waning Protection Event}
\usage{
gswanePfMOI(tEvent, PAR)
}
\arguments{
\item{tEvent}{end GS protection}

\item{PAR}{\code{NULL}}
}
\description{
End PfMOI GS protection.
This method is bound to \code{Human$gswanePfMOI()}
\itemize{
\item protection: infected human to mosquito transmission efficiency is set back to \code{Pf_c}, see \code{\link{PfMOI.Parameters}}
}
}
