% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMANS-Human-Class.R
\docType{class}
\name{Human}
\alias{Human}
\title{Human Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
Human
}
\description{
This is a generic human being blah blah ...
Each instance of a \code{Human} lives in a \code{\link{HumanPop}}
}
\details{
\itemize{
\item{\code{reg}} - .
\item{\code{outvar}} - .
\item{\code{levels}} - .
\item{\code{nbins}} - .
}
}
\section{Methods}{

\describe{
  \item{\code{new(reg, DataStorageClass.g0, ...)}}{...}
  \item{\code{fit(data)}}{...}
  \item{\code{predict(newdata)}}{...}
  \item{\code{predictAeqa(newdata)}}{...}
}
}

\section{Active Bindings}{

\describe{
  \item{\code{cats}}{...}
}
}

\keyword{R6}
\keyword{class}
