% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{Human_get_humanPfMOI}
\alias{Human_get_humanPfMOI}
\title{PfMOI \code{Human} Method: Get Human-stage PfMOI Object}
\usage{
Human_get_humanPfMOI()
}
\description{
Get the \code{\link{humanPfMOI}} object in a human.
This method is bound to \code{Human$get_humanPfMOI()}
}
