% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{Human_get_humanPfSI}
\alias{Human_get_humanPfSI}
\title{PfSI \code{Human} Method: Get Human-stage PfSI Object}
\usage{
Human_get_humanPfSI()
}
\description{
Get the \code{\link{humanPfSI}} object in a human.
This method is bound to \code{Human$get_humanPfSI()}
}
