% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{Human_get_PfMOI_PAR}
\alias{Human_get_PfMOI_PAR}
\title{PfMOI \code{Human} Method: Get PfMOI_PAR}
\usage{
Human_get_PfMOI_PAR(ix = NULL)
}
\description{
Get either entire list or one named element of PfMOI_PAR. See \code{\link{PfMOI.Parameters}} for definitions.
This method is bound to \code{Human$get_PfMOI_PAR_PAR()}
}
