% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{Human_get_PfSI_PAR}
\alias{Human_get_PfSI_PAR}
\title{PfSI \code{Human} Method: Get PfSI_PAR}
\usage{
Human_get_PfSI_PAR(ix = NULL)
}
\description{
Get either entire list or one named element of PfSI_PAR. See \code{\link{PfSI.Parameters}} for definitions.
This method is bound to \code{Human$get_PfSI_PAR()}
}
