% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{Human_set_humanPfMOI}
\alias{Human_set_humanPfMOI}
\title{PfMOI \code{Human} Method: Set Human-stage PfMOI Object}
\usage{
Human_set_humanPfMOI(PfID, tInf = NULL, MOI = 0L, b = 0.55, c = 0.15,
  damID = NULL, sireID = NULL)
}
\description{
Set the \code{\link{humanPfMOI}} object in a human.
This method is bound to \code{Human$set_humanPfMOI()}
}
