% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{Human_set_humanPfSI}
\alias{Human_set_humanPfSI}
\title{PfSI \code{Human} Method: Set Human-stage PfSI Object}
\usage{
Human_set_humanPfSI(PfID, tInf = -1L, b = 0.55, c = 0.15, damID = -1L,
  sireID = -1L, infected = FALSE, chemoprophylaxis = FALSE)
}
\description{
Set the \code{\link{humanPfSI}} object in a human.
This method is bound to \code{Human$set_humanPfSI()}
}
