% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{Human_set_PfMOI_PAR}
\alias{Human_set_PfMOI_PAR}
\title{PfMOI \code{Human} Method: Set PfMOI_PAR}
\usage{
Human_set_PfMOI_PAR(PfMOI_PAR)
}
\description{
Set either entire list of PfMOI_PAR. See \code{\link{PfMOI.Parameters}} for definitions.
This method is bound to \code{Human$set_PfMOI_PAR_PAR()}
}
