% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{Human_set_PfSI_PAR}
\alias{Human_set_PfSI_PAR}
\title{PfSI \code{Human} Method: Set PfSI_PAR}
\usage{
Human_set_PfSI_PAR(PfSI_PAR)
}
\description{
Set either entire list of PfSI_PAR. See \code{\link{PfSI.Parameters}} for definitions.
This method is bound to \code{Human$set_PfSI_PAR()}
}
