% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMANS-Pf-Methods.R
\name{HumanPf.Setup}
\alias{HumanPf.Setup}
\title{Initialize Additional Methods & Fields in \code{Human} and \code{HumanPop} for P.falciparum PATHOGEN component}
\usage{
HumanPf.Setup()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
HumanPf.Setup()
}
