% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{humanPfSI}
\alias{humanPfSI}
\title{PfSI Human Stage Pathogen Class}
\usage{
humanPfSI(PfID_init, tInf_init = -1, b_init = 0.55, c_init = 0.15,
  damID_init = -1, sireID_init = -1, infected_init = FALSE,
  chemoprophylaxis_init = FALSE)
}
\arguments{
\item{PfID_init}{PfID derived from the tile level PfPedigree}

\item{tInf_init}{time of current infection (successful mosquito to human transmission event)}

\item{b_init}{infected mosquito to human transmission efficiency}

\item{c_init}{infected human to mosquito transmission efficiency}

\item{damID_init}{female gametocyte 'mother'}

\item{sireID_init}{male gametocyte 'father'}

\item{infected_init}{infection status}

\item{chemoprophylaxis_init}{chemoprophylaxis protection status}
}
\description{
A class for storing parameters and state variables of human-stage PfSI infections.
}
\section{Methods}{

\itemize{
\item get_PfID:
\item push_PfID:
\item get_tInf:
\item push_tInf:
\item get_b:
\item set_b:
\item get_c:
\item set_c:
\item get_damID:
\item push_damID:
\item get_sireID:
\item push_sireID:
\item get_infected:
\item set_infected:
\item get_chemoprophylaxis:
\item set_chemoprophylaxis:
}
}

\section{Fields}{

\itemize{
\item a field: talk about me
}
}

