% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{humanPfSIcpp}
\alias{humanPfSIcpp}
\title{PfSI Human Stage Pathogen Class}
\usage{
humanPfSIcpp(PfID_init, tInf_init = -1, b_init = 0.55, c_init = 0.15,
  damID_init = -1, sireID_init = -1, infected_init = FALSE,
  chemoprophylaxis_init = FALSE)
}
\arguments{
\item{PfID_init}{PfID derived from the tile level PfPedigree}

\item{tInf_init}{time of current infection (successful mosquito to human transmission event)}

\item{b_init}{infected mosquito to human transmission efficiency}

\item{c_init}{infected human to mosquito transmission efficiency}

\item{damID_init}{female gametocyte 'mother'}

\item{sireID_init}{male gametocyte 'father'}

\item{infected_init}{infection status}

\item{chemoprophylaxis_init}{chemoprophylaxis protection status}
}
\description{
A class for storing parameters and state variables of human-stage PfSI infections.
}
\section{Methods}{

\itemize{
\item firstEvent: return first event in queue as named R list
\item firstTime: return 'tEvent' slot in first event in queue
\item rmFirstEventFromQ: remove first event in queue and update \code{queueN}
\item rmTagFromQ(const std::string &tag): remove all events with certain tag from queue and update \code{queueN}
\item get_queueN: return \code{queueN}
\item get_EventQ: return entire \code{EventQ} as R list
\item addEvent2Q(Rcpp::List event): add a list to the \code{EventQ} in correct time order and update \code{queueN}
}
}

\section{Fields}{

\itemize{
\item EventQ: \code{std::vector<Rcpp::List>}
\item queueN: integer value of size of queue.
}
}

