% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{HumanPop_get_PfMOI_history}
\alias{HumanPop_get_PfMOI_history}
\title{PfMOI \code{HumanPop} Method: Get PfMOI Histories}
\usage{
HumanPop_get_PfMOI_history()
}
\description{
Get all PfMOI histories, stored in class \code{\link{humanPfMOI}}
This function is bound to \code{HumanPop$get_PfMOI_history()}
}
