% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{HumanPop_get_PfSI_history}
\alias{HumanPop_get_PfSI_history}
\title{PfSI \code{HumanPop} Method: Get PfSI Histories}
\usage{
HumanPop_get_PfSI_history()
}
\description{
Get all PfSI histories, stored in class \code{\link{humanPfSI}}
This function is bound to \code{HumanPop$get_PfSI_history()}
}
