% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{HumanPop_get_PfSI_PAR}
\alias{HumanPop_get_PfSI_PAR}
\title{PfSI \code{HumanPop} Method: Get PfSI_PAR}
\usage{
HumanPop_get_PfSI_PAR(ix = NULL)
}
\arguments{
\item{PfSI_PAR}{new parameter list, see \code{\link{PfSI.Parameters}}}
}
\description{
Get either entire list or one named element of PfSI_PAR. See \code{\link{PfSI.Parameters}} for definitions.
This method is bound to \code{HumanPop$get_PfSI_PAR()}
}
