% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{HumanPop_set_humanPfMOI}
\alias{HumanPop_set_humanPfMOI}
\title{PfMOI \code{HumanPop} Method: Set Human-stage PfMOI Object}
\usage{
HumanPop_set_humanPfMOI(b = NULL, c = NULL)
}
\arguments{
\item{b}{infected mosquito to human transmission efficiency}

\item{c}{infected human to mosquito transmission efficiency}
}
\description{
Set the \code{\link{humanPfMOI}} object in a human poulation.
This method is bound to \code{HumanPop$set_humanPfMOI()}
}
