% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{HumanPop_set_humanPfSI}
\alias{HumanPop_set_humanPfSI}
\title{PfSI \code{HumanPop} Method: Set Human-stage PfSI Object}
\usage{
HumanPop_set_humanPfSI(b = NULL, c = NULL)
}
\arguments{
\item{b}{infected mosquito to human transmission efficiency}

\item{c}{infected human to mosquito transmission efficiency}
}
\description{
Set the \code{\link{humanPfSI}} object in a human poulation; this does not initialize infection, only sets the object.
This method is bound to \code{HumanPop$set_humanPfSI()}
}
