% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{HumanPop_set_PfMOI_PAR}
\alias{HumanPop_set_PfMOI_PAR}
\title{PfMOI \code{HumanPop} Method: Set PfMOI Paramters for a \code{\link{HumanPop}}}
\usage{
HumanPop_set_PfMOI_PAR(PfMOI_PAR)
}
\arguments{
\item{PfMOI_PAR}{new parameter list, see \code{\link{PfSI.Parameters}}}
}
\description{
Set PfMOI_PAR for a HumanPop; this is useful for simulating multiple populations with different parameter values.
This method is bound to \code{HumanPop$set_PfMOI_PAR()}
}
