% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{HumanPop_set_PfSI_PAR}
\alias{HumanPop_set_PfSI_PAR}
\title{PfSI \code{HumanPop} Method: Set PfSI Paramters for a \code{\link{HumanPop}}}
\usage{
HumanPop_set_PfSI_PAR(PfSI_PAR)
}
\arguments{
\item{PfSI_PAR}{new parameter list, see \code{\link{PfSI.Parameters}}}
}
\description{
Set PfSI_PAR for a HumanPop; this is useful for simulating multiple populations with different parameter values.
This method is bound to \code{HumanPop$set_PfSI_PAR()}
}
