% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PFSI-Auxiliary.R
\name{Humans_collectOutput}
\alias{Humans_collectOutput}
\title{Return Human State Histories as a data.table}
\usage{
Humans_collectOutput(history, day_integers)
}
\arguments{
\item{history}{a human history object}

\item{day_integers}{integer vector of days}
}
\value{
data.table of states
}
\description{
Given all human histories (which can be extracted via \code{\link{convertHumanEvent_PfSI}} or \code{\link{importHumanEvent_PfSI}}),
retrieve the cohort state histories as a \code{\link{data.table}}.
}
\examples{
Humans_getInfoDay(history = humanHistories,day = -1L:tMax)
}
