% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PFSI-Auxiliary.R
\name{Humans_getInfoDay}
\alias{Humans_getInfoDay}
\title{Find State Status for Specific Human on Specific Day}
\usage{
Humans_getInfoDay(human, day)
}
\arguments{
\item{human}{a single human history}

\item{day}{what day}
}
\value{
character vector of state
}
\description{
Given a single human history (which can be extracted via \code{\link{convertHumanEvent_PfSI}} or \code{\link{importHumanEvent_PfSI}}),
retrieve their state on that day.
}
\examples{
Humans_getInfoDay(human = humanHistories[[1]],day = 1)
}
