% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{ImagoQ}
\alias{ImagoQ}
\title{MICRO Imago Queue}
\usage{
ImagoQ()
}
\description{
ImagoQ
}
\section{Methods}{

\itemize{
\item clear_ImagoQ:
\item clear_ImagoQTime:
\item add_ImagoQ:
\item track_ImagoQ:
\item get_N:
\item set_N:
\item get_ImagoQ:
\item get_ImagoQTime:
}
}

\section{Fields}{

\itemize{
\item a field:
}
}

