% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUATIC-ECOLOGY.R
\name{importEggQ}
\alias{importEggQ}
\title{Import Logged EggQ Data from .csv}
\usage{
importEggQ(directory, fileName)
}
\arguments{
\item{directory}{directory that OUTPUT/.. is in}

\item{fileName}{file name in OUTPUT/..}
}
\value{
data frame
}
\description{
Generate daily output of EggQ tracking to .csv file through text connection created with \code{trackEggQ_init()}.
}
\examples{
importimportEggQEL4P(directory, fileName)
}
