% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILITIES-Human.R
\name{importHumanEvent_PfSI}
\alias{importHumanEvent_PfSI}
\title{Reimport Human eventT and events from .json}
\usage{
importHumanEvent_PfSI(directory)
}
\arguments{
\item{directory}{directory; files are in directory/OUTPUT/..}
}
\value{
list of histories
\itemize{
\item eventT: times events occur
\item events: events
}
}
\description{
Import human histories exported to .json from PfSI module.
}
\examples{
writeHumanEvent(directory, fileName)
}
