% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Utility.R
\name{importMicroMosquitoHistory_utility}
\alias{importMicroMosquitoHistory_utility}
\title{M-BITES Utility: Import JSON Mosquito Histories}
\usage{
importMicroMosquitoHistory_utility(directory, string)
}
\arguments{
\item{directory}{directory in which output was stored (see \code{MicroTile$get_directory})}

\item{string}{search string for \code{\link{grep}} to read files (can be "cohort" or "history")}
}
\description{
From methods which write mosquito bionomics and histories to JSON, reimport as R lists.
}
