% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfPedigree.R
\name{importPfPedigree}
\alias{importPfPedigree}
\title{Import PfPedigree Data from .csv}
\usage{
importPfPedigree(directory, fileName)
}
\arguments{
\item{directory}{directory of output; files are in directory/OUTPUT/..}

\item{fileName}{name of the file; directory/OUTPUT/fileName.csv}
}
\value{
a data frame
}
\description{
Import PfPedigree data written to .csv from \code{\link{addPf2Pedigree}}.
}
\examples{
importPfTransmission(directory, fileName)
}
