% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfPedigree.R
\name{importPfTransmission}
\alias{importPfTransmission}
\title{Import PfTransmission Data from .csv}
\usage{
importPfTransmission(directory, fileName)
}
\arguments{
\item{directory}{directory of output; files are in directory/OUTPUT/..}

\item{fileName}{name of the file; directory/OUTPUT/fileName.csv}
}
\value{
a data frame
}
\description{
Import PfTransmission data written to .csv from \code{\link{trackPfTransmission}}.
}
\examples{
importPfTransmission(directory, fileName)
}
