% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{infectHumanPfSI}
\alias{infectHumanPfSI}
\title{PfSI \code{Human} Event: PfSI Infection Event}
\usage{
infectHumanPfSI(tEvent, PAR)
}
\arguments{
\item{tEvent}{time of infection}

\item{PAR}{must be a list containing \code{damID} and \code{sireID}}
}
\description{
Simulate a PfSI infection. If the human is not under chemoprophylaxis or already infected, begin an infection.
This method is bound to \code{Human$infectHumanPfSI()}
\itemize{
\item A Bernoulli event is drawn to determine if this infection produces fever; if so \code{\link{add2Q_feverPfSI}} is called.
\item The end of this PfSI infection is queued by \code{\link{add2Q_endPfSI}}
}
}
