% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{infectiousBite_PfMOI}
\alias{infectiousBite_PfMOI}
\title{PfMOI \code{Human} Method: Infectious Bite on Human}
\usage{
infectiousBite_PfMOI(tBite, PAR)
}
\arguments{
\item{tBite}{time of bite}

\item{PAR}{single clonal variant returned from \code{mosquitoPfMOI$get_clone()}}
}
\description{
This method is called from \code{\link{probeHost_PfMOI}}.
If the infectious bite results in successful transmission, this function queues a human infection event, see \code{\link{add2Q_infectHumanPfMOI}}
This method is bound to \code{Human$infectiousBite_PfMOI()}
}
