% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PFSI-PfSI.R
\name{infectMosquito_PfSI}
\alias{infectMosquito_PfSI}
\title{Infectious Bite for Human to Vector Transmission}
\usage{
infectMosquito_PfSI(tBite, ixH, ixS, ixM)
}
\arguments{
\item{ixH}{index of human}

\item{ixS}{index of feeding site}

\item{t}{time of bite (local time of mosquito)}
}
\value{
a named list
\itemize{
\item infected: TRUE or FALSE depending on human infection status and transmisson efficiency
\item PfM: if transmisson successful, the pathogen object passed to the mosquito
}
}
\description{
This function handles human to vector transmission for the PfSI module and is called from \code{getInfected()} defined in MBITES-HostEncounter.R.
If the human has an active infection and there is a successful bloodstream to mosquito transfer of gametocytes a \code{PfM} object is made by \code{makePfM()}
and is passed to the mosquito. If global flag \code{\link{PfTransmission_TRACK}} is true it will call \code{\link{trackPfTransmission}} to track human to vector transmission.
}
