% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Site-AquaticEcology.R
\name{init.AquaticEcology}
\alias{init.AquaticEcology}
\title{Initialize Aquatic Ecology COMPONENT}
\usage{
init.AquaticEcology()
}
\arguments{
\item{hi}{spacecats}
}
\value{
stuff
}
\description{
This function initializes generic methods and fields for the Aquatic Ecology COMPONENT.
It modifies \code{\link{AquaticSite}} and \code{\link{Landscape}} classes.
}
\examples{
init.Emerge()
}
