% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Site-Emerge.R
\name{init.Emerge}
\alias{init.Emerge}
\title{Initialize Emerge MODULE}
\usage{
init.Emerge()
}
\arguments{
\item{dunno}{sdf}
}
\value{
stuff
}
\description{
This function initializes methods and fields for the 'Emerge' MODULE of Aquatic Ecology COMPONENT.
It modifies \code{\link{AquaticSite}} and \code{\link{Landscape}} classes.
}
\examples{
init.Emerge()
}
