% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Pathogen-PfSI.R
\name{init.PfSI}
\alias{init.PfSI}
\title{Initialize PfSI Module (Pathogen)}
\usage{
init.PfSI()
}
\arguments{
\item{write}{me}
}
\value{
write me
}
\description{
Initialize methods in \code{\link{Human}} and \code{\link{HumanPop}} classes for PfSI Pathogen module.
}
\examples{
init.PfSI()
}
