% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{init_MACRO_PfMOI}
\alias{init_MACRO_PfMOI}
\title{PfMOI \code{HumanPop} Method: Initialize MACRO PfMOI Infections}
\usage{
init_MACRO_PfMOI(PfMOI, tStart = 0)
}
\arguments{
\item{PfMOI}{a vector of mean MOI in each patch}

\item{tStart}{time to start simulation}
}
\description{
Initialize PfMOI infections with mean parasite MOI for each patch.
This method is bound to \code{HumanPop$init_MACRO_PfMOI()}
}
