% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{init_MacroTile_PfMOI}
\alias{init_MacroTile_PfMOI}
\title{MACRO PfMOI \code{MacroTile} Method: Initialize MACRO PfMOI Infections}
\usage{
init_MacroTile_PfMOI(PfMOI, tStart = 0)
}
\arguments{
\item{tStart}{time to start simulation}

\item{PfPR}{a vector of parasite prevalence at each patch}
}
\description{
Initialize PfMOI infections with mean parasite MOI for each patch.
Pass \code{PfMOI} and \code{tStart} to function \code{\link{init_MACRO_PfMOI}}
This method is bound to \code{MacroTile$init_PfMOI()}
}
