% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{init_MacroTile_PfSI}
\alias{init_MacroTile_PfSI}
\title{MACRO PfSI \code{MacroTile} Method: Initialize MACRO PfSI Infections}
\usage{
init_MacroTile_PfSI(PfPR, tStart = 0)
}
\arguments{
\item{PfPR}{a vector of parasite prevalence at each patch}

\item{tStart}{time to start simulation}
}
\description{
Initialize PfSI infections with parasite prevalence PfPR for each patch.
Pass \code{PfPR} and \code{tStart} to function \code{\link{init_MACRO_PfSI}}
This method is bound to \code{MacroTile$init_PfSI()}
}
