% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{init_MICRO_PfSI}
\alias{init_MICRO_PfSI}
\title{PfSI \code{HumanPop} Method: Initialize MICRO PfSI Infections}
\usage{
init_MICRO_PfSI(PfPR, tStart = 0)
}
\description{
Initialize PfSI infections with parasite prevalence PfPR for MICRO or SimBitePfSI
This method is bound to \code{HumanPop$init_MICRO_PfSI()}
}
