% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Human-ActivitySpace.R
\name{init_MicroHuman_ActivitySpace}
\alias{init_MicroHuman_ActivitySpace}
\title{MICRO \code{\link{Human}} Method: Initialize Activity Space}
\usage{
init_MicroHuman_ActivitySpace(nDaily)
}
\arguments{
\item{nDaily}{average daily number of other sites visited}
}
\description{
This function is bound to \code{Human$init_ActivitySpace()}
}
