% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Human-ActivitySpace.R
\name{init_MicroHumanPop_ActivitySpace}
\alias{init_MicroHumanPop_ActivitySpace}
\title{MICRO \code{\link{HumanPop}} Method: Initialize Activity Space}
\usage{
init_MicroHumanPop_ActivitySpace(nDaily = 1.4)
}
\arguments{
\item{nDaily}{average daily number of other sites visited}
}
\description{
This function is bound to \code{HumanPop$init_ActivitySpace()}
}
