% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{init_Pathogens_MicroMosquitoFemale_PfMOI}
\alias{init_Pathogens_MicroMosquitoFemale_PfMOI}
\title{PfMOI Helper Code for Pathogen Initialization in \code{\link{MicroMosquitoFemale}}}
\usage{
init_Pathogens_MicroMosquitoFemale_PfMOI()
}
\description{
Initializes an empty PfMOI pathogen in \code{\link{MicroMosquitoFemale}} called during object initialization.
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$init_Pathogens()}
}
}
