% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{init_Pathogens_MicroMosquitoPopFemale_PfMOI}
\alias{init_Pathogens_MicroMosquitoPopFemale_PfMOI}
\title{PfMOI Helper Code for Pathogen Initialization in \code{\link{MicroMosquitoPopFemale}}}
\usage{
init_Pathogens_MicroMosquitoPopFemale_PfMOI()
}
\description{
Initializes empty PfMOI pathogens in \code{\link{MicroMosquitoPopFemale}} called during object initialization.
\itemize{
\item This method is bound to \code{MicroMosquitoPopFemale$init_Pathogens()}
}
}
