% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{init_Pathogens_MicroMosquitoPopFemale_PfSI}
\alias{init_Pathogens_MicroMosquitoPopFemale_PfSI}
\title{PfSI Helper Code for Pathogen Initialization in \code{\link{MicroMosquitoPopFemale}}}
\usage{
init_Pathogens_MicroMosquitoPopFemale_PfSI()
}
\description{
Initializes empty PfSI pathogens in \code{\link{MicroMosquitoPopFemale}} called during object initialization.
\itemize{
\item This method is bound to \code{MicroMosquitoPopFemale$init_Pathogens()}
}
}
