% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Site-RiskList.R
\name{init_riskList}
\alias{init_riskList}
\title{MICRO \code{FeedingSite} Method: Initialize a Risk List}
\usage{
init_riskList(maxH = 20L, zooW = rgamma(1, 1, 1))
}
\arguments{
\item{maxH}{number of human slots to preallocate this site's risk list}

\item{zooW}{biting weight on livestock at this site's risk list}
}
\description{
Initialize a risk list for a feeding site object. This function is a method for \code{\link{FeedingSite}}.
}
\section{Structure}{

\itemize{
\item maxH: maximum number of humans in this risk list (memory allocation)
\item N: current number of humans in this risk list
\item who: index of who is at risk (myID value of the \code{\link{Human}} objects at risk at this site)
\item pTm: person time at risk
\item w: biting weight on hosts at risk
\item nO: number of non-human hosts
\item other: (general list structure for adding non-human hosts)
\itemize{
\item zoo:
\itemize{
\item w: biting weight on livestock
\item typeID: -1 corresponds to livestock host id
}
}
}
}

\examples{
some_function()
}
