% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-Macro-PfMethods.R
\name{initialize_PatchPf}
\alias{initialize_PatchPf}
\title{Pf: Initialize \code{Human}}
\usage{
initialize_PatchPf(MacroPatch_PAR)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
set_biteWeight()
}
