% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMANS-Pf-Methods.R
\name{initialize_Pf}
\alias{initialize_Pf}
\title{Pf: Initialize \code{Human}}
\usage{
initialize_Pf(myID, hhID, bDay, biteWeight)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
set_biteWeight()
}
