% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Movement.R
\name{json_travelHistory}
\alias{json_travelHistory}
\title{Write \code{HumanPop} Travel History to JSON}
\usage{
json_travelHistory(con)
}
\arguments{
\item{con}{a connection open for writing}
}
\value{
none
}
\description{
Write each \code{Human} travel history as a .json object to a connection. Depends on \code{\link{self$get_travelHistory()}} defined in MACRO-Human-Movement.R
}
\examples{
HumanPop$json_travelHistory(con = file(description = paste0(directory, "OUTPUT/humanTravel.json"),open = "wt"))
}
