% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-EL4P-Calibrate.R
\name{K2psi_MicroEL4P}
\alias{K2psi_MicroEL4P}
\title{Convert K and Regression Coefficient into Psi}
\usage{
K2psi_MicroEL4P(K, cf)
}
\arguments{
\item{K}{vector of K values}

\item{cf}{regression coefficients from \code{\link{psi2K_cf_MicroEL4P}}}
}
\value{
psi
}
\description{
Return Psi from K and slope coefficient of linear relationship.
}
