% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Landscape-Parameters.R
\name{Landscape.Aqua.Parameters}
\alias{Landscape.Aqua.Parameters}
\title{MICRO: Generate Parameters for \code{\link{Landscape}} \code{\link{AquaticSite}}}
\usage{
Landscape.Aqua.Parameters(nAqua, siteXY, module, modulePars, searchWt = NULL,
  haz = 0)
}
\arguments{
\item{nAqua}{number of aquatic habitats}

\item{siteXY}{two element list of \code{x} and \code{y} coordinates of aquatic habitats}

\item{module}{character
\itemize{
\item "emerge": initialize parameters for Emerge module of Aquatic Ecology
\item "EL4P": initialize parameters for EL4P module of Aquatic Ecology
}}

\item{modulePars}{additional list of named parameters to be passed to Aquatic Ecology module specific parameter generating functions
\itemize{
\item Emerge: should be named list of parameters for \code{\link{makeLambda_MicroEmerge}}
\item EL4P: NULL
}}

\item{searchWt}{vector of searchWt (if \code{NULL} initialize to Gamma(1,1) distribution)}

\item{haz}{mean value of landing hazards (if \code{!= 0} use \code{\link{betaRootA}} to find alpha parameter of beta distribution to give that mean value and produce Beta distributed hazards)}
}
\value{
return a list
}
\description{
This function generates a named list of parameters to initialize all \code{\link{AquaticSite}} objects on a MICRO \code{\link{Landscape}}.
}
