% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Landscape-Parameters.R
\name{Landscape.Feeding.PAR}
\alias{Landscape.Feeding.PAR}
\title{MICRO: Generate Parameters for \code{Landscape} \code{AquaticSite}}
\usage{
Landscape.Feeding.PAR(nFeed, pointGen = "poisson", searchWt = NULL,
  enterP = 1, ...)
}
\arguments{
\item{nFeed}{number of feeding sites}

\item{pointGen}{character to select spatial point pattern generation function
\itemize{
\item "poisson": \code{\link{pointsPoisson}}
\item "clustered": \code{\link{pointsClustered}}
\item "overdispersed": \code{\link{pointsOverdispersed}}
\item "lattice": \code{\link{pointsLattice}}
}}

\item{searchWt}{vector of searchWt (if \code{NULL} initialize to Gamma(1,1) distribution)}

\item{enterP}{vector of searchWt (if \code{NULL} initialize to Gamma(1,1) distribution)}

\item{...}{additional named arguments to be passed to the pointGen(nFeed, ...) function}

\item{hazV}{mean value for feeding site vegetation landing hazard (if 0 it is set to 0 for all sites)}

\item{hazW}{mean value for feeding site outside wall landing hazard (if 0 it is set to 0 for all sites)}

\item{hazI}{mean value for feeding site indoor wall landing hazard (if 0 it is set to 0 for all sites)}
}
\value{
return a list
}
\description{
This function generates a named list of parameters to initialize all \code{\link{AquaticSite}} objects on a MICRO \code{\link{Landscape}}.
}
