% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Landscape-Parameters.R
\name{landscape.PAR}
\alias{landscape.PAR}
\title{Generate Parameters for Landscape Object}
\usage{
landscape.PAR(nFeed, nAqua, pointGen, hhSize, hhMin, lambda, offset = 0,
  hazV = 0, hazW = 0, hazI = 0, haz = 0, enterP = 1, xLim = c(0, 1),
  yLim = c(0, 1), aquaSD = 0.025, ...)
}
\arguments{
\item{nFeed}{number of feeding sites (generated via pointGen(...))}

\item{nAqua}{number of aquatic habitats}

\item{pointGen}{spatial point pattern generation function}

\item{hhSize}{average number of humans at feeding sites}

\item{hhMin}{minimum number of humans at feeding sites}

\item{lambda}{mean emerging females per human per day for entire landscape summing out seasonality}

\item{offset}{seasonal offset}

\item{hazV}{mean value for feeding site vegetation landing hazard (if 0 it is set to 0 for all sites)}

\item{hazW}{mean value for feeding site outside wall landing hazard (if 0 it is set to 0 for all sites)}

\item{hazI}{mean value for feeding site indoor wall landing hazard (if 0 it is set to 0 for all sites)}

\item{haz}{mean value for aquatic habitat landing hazard (if 0 it is set to 0 for all sites)}

\item{xLim}{x-axis bounds for simulated points}

\item{yLim}{y-axis bounds for simulated points}

\item{aquaSD}{standard deviation of aquatic habitat scatter around feeding sites}

\item{...}{additional named arguments for pointGen()}
}
\value{
a named list of parameters
\itemize{
\item nFeed: number of feeding sites
\item nAqua: number of aquatic habitats
\item feedXY: 2 element list (x and y coords, respectively)
\item feedWt: search weight for feeding sites
\item sugar: opportunistic sugar weight at feeding sites
\item enterP: probability to enter house
\item hazV: feeding site vegetation landing hazards
\item hazW: feeding site outside wall landing hazards
\item hazI: feeding site indoor wall landing hazards
\item humanPop: size of human population at each site
\item humanPopIx: id of humans at each site
\item humanPopN: total population size
\item aquaIx: 'parent' feeding sites of aquatic habitats
\item aquaXY: 2 element list (x and y coords, respectively)
\item aquaWt: search weight for aquatic habitats
\item haz: landing hazard for aquatic habitats
}
}
\description{
This function is a specific instantiation of a generic system to generate parameters for a
chosen landscape. Any user-specified function can be written to generate parameters, as long as the
return list is in the same format.
}
\examples{

}
