% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Landscape-Class.R
\docType{class}
\name{Landscape}
\alias{Landscape}
\title{MICRO Landscape Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
Landscape
}
\description{
This is a Landscape object that is part of a MICRO microsimulation Tile object, defined in \code{\link{MicroTile}}.
}
\section{Methods}{

\itemize{
\item \strong{Constructor}
\itemize{
\item new: initialize a new \code{Landscape} object
\itemize{
\item Arguments:
\itemize{
\item \code{Landscape_PAR}: see \code{\link{Landscape.Parameters}} for parameter generation function.
}
}
}
\item \strong{Getters & Setters}
\itemize{
\item get_FeedingSites: get \code{private$FeedingSites}
\itemize{
\item Arguments:
\itemize{
\item \code{ixS}: if \code{NULL} return all sites, otherwise return site indexed by \code{ixS}
}
}
\item set_FeedingSites: set \code{private$FeedingSites}
\itemize{
\item Arguments:
\itemize{
\item \code{FeedingSites}: replacement object (see below)
\item \code{ixS}: if \code{NULL} replace all sites, otherwise replace site indexed by \code{ixS}
}
}
\item get_AquaSites: get \code{private$AquaSites}
\itemize{
\item Arguments:
\itemize{
\item \code{ixS}: if \code{NULL} return all sites, otherwise return site indexed by \code{ixS}
}
}
\item set_AquaSites: set \code{private$AquaSites}
\itemize{
\item Arguments:
\itemize{
\item \code{AquaSites}: replacement object (see below)
\item \code{ixS}: if \code{NULL} replace all sites, otherwise replace site indexed by \code{ixS}
}
}
\item get_SugarSites: get \code{private$SugarSites}
\item set_SugarSites: set \code{private$SugarSites}
\item get_MatingSites: get \code{private$MatingSites}
\item set_MatingSites: set \code{private$MatingSites}
}
\item \strong{Pointers}
\itemize{
\item get_TilePointer: get \code{\link{MicroTile}} pointer
\item set_TilePointer: set \code{\link{MicroTile}} pointer
\item set_FemalePopPointer:
\item set_FemalePopPointer:
\item get_MosquitoPopMalePointer:
\item set_MosquitoPopMalePointer:
\item get_HumansPointer: get \code{\link{HumanPop}} pointer
\item set_HumansPointer: set \code{\link{HumanPop}} pointer
}
}
}

\keyword{R6}
\keyword{class}
