% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-MosquitoPop-Methods.R
\name{layEggs_MacroMosquitoPop}
\alias{layEggs_MacroMosquitoPop}
\title{\code{MacroMosquitoPop} layEggs}
\usage{
layEggs_MacroMosquitoPop()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
oviposition
}
\examples{
some_function()
}
