% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{lmTest_PfMOI}
\alias{lmTest_PfMOI}
\title{PfMOI \code{Human} Method: Light Microscopy Test}
\usage{
lmTest_PfMOI(tEvent, PAR)
}
\description{
Administer light microscopy to this human.
\itemize{
\item if infected: true positive is detected with probability \code{lmSensPf}, see \code{\link{PfMOI.Parameters}}
\item if susceptible: false positive is detected with probability \code{lmSpecPf}, see \code{\link{PfMOI.Parameters}}
}
}
