% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Setup.R, R/MICRO-Human-Setup.R
\name{MACRO.Humans.Setup}
\alias{MACRO.Humans.Setup}
\alias{MACRO.Humans.Setup}
\title{MACRO: Initialize Additional Methods & Fields in \code{Human}}
\usage{
MACRO.Humans.Setup(overwrite = TRUE)

MACRO.Humans.Setup(overwrite = TRUE)
}
\arguments{
\item{a}{parameter}

\item{a}{parameter}
}
\value{
does stuff

does stuff
}
\description{
Write me! \code{\link{Human}}

Write me! \code{\link{Human}}
}
\examples{
MACRO.Humans.Setup()
MACRO.Humans.Setup()
}
