% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Macro-Emerge.R
\name{MACRO.Patch.Emerge.Setup}
\alias{MACRO.Patch.Emerge.Setup}
\title{Initialize Additional Methods & Fields in \code{MacroPatch} for Emerge Module of Aquatic Ecology}
\usage{
MACRO.Patch.Emerge.Setup(overwrite = TRUE)
}
\arguments{
\item{a}{parameter}
}
\value{
do stuff
}
\description{
Write me! See for parameters that are initialized after this; ie: anything that \code{\link{MACRO.Patch.Parameters}} calls for 'Emerge' should be defined here.
}
\examples{
MACRO.Patch.Emerge.Setup()
}
