% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Patch-Parameters.R
\name{MACRO.Patch.Parameters}
\alias{MACRO.Patch.Parameters}
\title{Initialize MACRO Patch Parameters for \code{MacroPatch}}
\usage{
MACRO.Patch.Parameters(N, hhID, humanIDs, bWeightHuman = NULL,
  bWeightZoo1 = 1, bWeightZoo2 = 1, aquaModule = "emerge", aquaPars)
}
\arguments{
\item{N}{number of patches}

\item{bWeightZoo1}{shape parameter of gamma zoophilic biting weights}

\item{bWeightZoo2}{rate parameter of gamma zoophilic biting weights}
}
\value{
return a list
}
\description{
This is used to generate a list of parameters for \code{\link{MacroPatch}} and should be used during its initialization.
}
\examples{
MACRO.Patch.Parameters()
}
