% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-Macro-PfMethods.R
\name{MACRO.PatchPf.Setup}
\alias{MACRO.PatchPf.Setup}
\title{MACRO: Initialize Additional PATHOGEN PatchPf Methods & Fields in \code{MacroPatch}}
\usage{
MACRO.PatchPf.Setup()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me! \code{\link{Human}}
}
\examples{
MACRO.Humans.Setup()
}
