% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Tile-Parameters.R
\name{MACRO.Tile.Parameters}
\alias{MACRO.Tile.Parameters}
\title{Initialize MACRO Tile Parameters for \code{MacroTile}}
\usage{
MACRO.Tile.Parameters(N = 10, patchSize = 20, patchMin = 10, EIP = 10,
  aquaModule = "emerge", aquaPars = list(lambda = rep(10, 10)),
  pathogenModule = "PfSI", humanPars = "auto")
}
\arguments{
\item{N}{number of patches}

\item{patchSize}{passed to \code{\link{sitePops}}}

\item{patchMin}{passed to \code{\link{sitePops}}}

\item{EIP}{length of entomological incubation period}

\item{aquaModule}{character in 'emerge', 'EL4P'}

\item{aquaPars}{Aquatic Ecology module specific parameters (see \code{\link{MACRO.Patch.Parameters}} for details)
\itemize{
\item Emerge: list with elements 'lambda'
\item EL4P: list with elements
}}

\item{pathogenModule}{character in 'PfSI', 'PfMOI'}

\item{humanPars}{character in 'auto', 'manual' ; if \code{manual} the user should pass}
}
\value{
return a list
}
\description{
This is used to generate a list of parameters for \code{\link{MacroTile}} and should be used during its initialization.
\code{MACRO.Tile.Parameters} will generate parameters that are passed to the \code{\link[R6]{initialize}} method of
the component objects in a tile:
\itemize{
\item \code{\link{HumanPop}}: creates parameters by calling \code{\link{HumanPop.Parameters}}
\item \code{\link{MacroPatch}}: creates parameters by calling \code{\link{MACRO.Patch.Parameters}}
\item \code{\link{MacroMosquitoPop}}: directly initializes parameters
}
}
\examples{
MACRO.Tile.Parameters()
}
