% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Biting.R
\name{MacroHuman_add2Q_Bites}
\alias{MacroHuman_add2Q_Bites}
\title{MACRO \code{Human} Method: add2Q_Bites}
\usage{
MacroHuman_add2Q_Bites(tEvent, PAR)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me! a method for \code{\link{Human}}
This method is bound to \code{Human$add2Q_Bites()}
}
\examples{
some_function()
}
