% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Travel.R
\name{MacroHuman_add2Q_takeTrip}
\alias{MacroHuman_add2Q_takeTrip}
\title{add2Q_takeTrip}
\usage{
MacroHuman_add2Q_takeTrip(tEvent, PAR = NULL)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
some_function()
}
