% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Travel.R
\name{MacroHuman_event_takeTrip}
\alias{MacroHuman_event_takeTrip}
\title{event_takeTrip}
\usage{
MacroHuman_event_takeTrip(tEvent, PAR)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
some_function()
}
