% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Biting.R
\name{MacroHuman_expectedBites}
\alias{MacroHuman_expectedBites}
\title{MACRO \code{Human} Method: expectedBites}
\usage{
MacroHuman_expectedBites()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me! a method for \code{\link{Human}}
This method is bound to \code{Human$expectedBites()}
}
\examples{
some_function()
}
