% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Setup.R
\name{MacroHuman_get_location}
\alias{MacroHuman_get_location}
\title{MACRO: Get \code{Human} location}
\usage{
MacroHuman_get_location()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
some_function()
}
