% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Setup.R
\name{MacroHuman_get_MosquitoPointer}
\alias{MacroHuman_get_MosquitoPointer}
\title{Get \code{Human} \code{MacroMosquitoPop} Pointer}
\usage{
MacroHuman_get_MosquitoPointer()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
some_function()
}
