% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Biting.R
\name{MacroHuman_get_myEIR}
\alias{MacroHuman_get_myEIR}
\title{Get \code{Human} myEIR}
\usage{
MacroHuman_get_myEIR()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
This method is bound to \code{Human$get_myEIR()}
}
\examples{
some_function()
}
