% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Setup.R
\name{MacroHuman_get_patchID}
\alias{MacroHuman_get_patchID}
\title{Get \code{Human} patchID}
\usage{
MacroHuman_get_patchID()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
some_function()
}
