% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Travel.R
\name{MacroHuman_get_travelHistoryHuman}
\alias{MacroHuman_get_travelHistoryHuman}
\title{Return \code{Human} Travel History}
\usage{
MacroHuman_get_travelHistoryHuman()
}
\value{
a list
\itemize{
\item location: vector of integer locations
\item tTravel: vector of numeric trip times
}
}
\description{
Write me! Called as \code{self$get_travelHistory()}  Defined in MACRO-Human-Travel.R
}
\examples{
some_function()
}
