% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Travel.R
\name{MacroHuman_get_travelHistoryHumanPop}
\alias{MacroHuman_get_travelHistoryHumanPop}
\title{Return \code{HumanPop} Travel History}
\usage{
MacroHuman_get_travelHistoryHumanPop()
}
\value{
a list
}
\description{
Write me! Called as \code{self$get_travelHistory()} defined in MACRO-Human-Travel.R
}
\examples{
some_function()
}
