% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Travel.R
\name{MacroHuman_init_travel}
\alias{MacroHuman_init_travel}
\title{MACRO \code{\link{Human}} Method: Initialize Travel}
\usage{
MacroHuman_init_travel(n, freqMean = 7, freqSd = 2, lengthMean = 2,
  lengthSd = 1, tNow = 0)
}
\arguments{
\item{n}{number of other patches to visit}

\item{freqMean}{meanlog}

\item{freqSd}{sdlog}
}
\value{
does stuff
}
\description{
Write me! a method for \code{\link{Human}}
}
\examples{
some_function()
}
