% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Travel.R
\name{MacroHuman_json_travelHistory}
\alias{MacroHuman_json_travelHistory}
\title{Write \code{HumanPop} Travel History to JSON}
\usage{
MacroHuman_json_travelHistory(con)
}
\arguments{
\item{con}{a connection open for writing}
}
\value{
none
}
\description{
Write each \code{\link{Human}} travel history as a .json object to a connection. Depends on \code{\link{get_travelHistoryHumanPop}} defined in MACRO-Human-Travel.R
}
\examples{
HumanPop$MacroHuman_json_travelHistory(con = file(description = paste0(directory, "OUTPUT/humanTravel.json"),open = "wt"))
}
