% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Biting.R
\name{MacroHuman_queueInfectiousBites}
\alias{MacroHuman_queueInfectiousBites}
\title{MACRO \code{HumanPop} Method: queueInfectiousBites}
\usage{
MacroHuman_queueInfectiousBites()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me! a method for \code{\link{HumanPop}}
This method is bound to \code{HumanPop$queueInfectiousBites()}
}
\examples{
some_function()
}
