% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Travel.R
\name{MacroHuman_returnHome}
\alias{MacroHuman_returnHome}
\title{returnHome}
\usage{
MacroHuman_returnHome(tEvent, PAR)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
some_function()
}
