% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Setup.R
\name{MacroHuman_set_location}
\alias{MacroHuman_set_location}
\title{MACRO: Set \code{Human} location}
\usage{
MacroHuman_set_location(location)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
some_function()
}
