% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Biting.R
\name{MacroHuman_set_myEIR}
\alias{MacroHuman_set_myEIR}
\title{Set \code{Human} myEIR}
\usage{
MacroHuman_set_myEIR(myEIR)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
This method is bound to \code{Human$set_myEIR()}
}
\examples{
some_function()
}
