% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Setup.R
\name{MacroHuman_set_PatchesPointer}
\alias{MacroHuman_set_PatchesPointer}
\title{Set \code{Human} \code{MacroPatch} Pointer}
\usage{
MacroHuman_set_PatchesPointer(PatchesPointer)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
some_function()
}
