% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Setup.R
\name{MacroHuman_set_patchID}
\alias{MacroHuman_set_patchID}
\title{Set \code{Human} patchID}
\usage{
MacroHuman_set_patchID(patchID)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
some_function()
}
