% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Setup.R
\name{MacroHuman_set_TilePointer}
\alias{MacroHuman_set_TilePointer}
\title{Set \code{Human} \code{MacroTile} Pointer}
\usage{
MacroHuman_set_TilePointer(TilePointer)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
some_function()
}
