% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Biting.R
\name{MacroHuman_sumKappa}
\alias{MacroHuman_sumKappa}
\title{MACRO: Update \code{Human} kappa For a Patch}
\usage{
MacroHuman_sumKappa()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Add my contribution to kappa to current patch (current value of \code{location}).
This method is bound to \code{Human$sumKappa()}
}
\examples{
some_function()
}
