% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Biting.R
\name{MacroHuman_sumKappa_PfSI}
\alias{MacroHuman_sumKappa_PfSI}
\title{MACRO: Update \code{\link{Human}} PfSI kappa For a Patch}
\usage{
MacroHuman_sumKappa_PfSI()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Add my contribution to kappa to current patch (current value of \code{location}).
This method is bound to \code{Human$sumKappa()}
}
\examples{
some_function()
}
