% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Travel.R
\name{MacroHuman_takeTrip}
\alias{MacroHuman_takeTrip}
\title{takeTrip}
\usage{
MacroHuman_takeTrip(tEvent, PAR)
}
\arguments{
\item{PAR}{a list of length two
\itemize{
\item there ID of patch this human will visit
\item ixTravel index of the patch this human will visit in their travel destinations
}}

\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
some_function()
}
