% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Biting.R
\name{MacroHuman_updateKappa}
\alias{MacroHuman_updateKappa}
\title{MACRO: Update \code{HumanPop} kappa For all Patches}
\usage{
MacroHuman_updateKappa()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Update normalized biting propensities (kappa) for all patches.
This method is bound to \code{HumanPop$updateKappa()}
}
\examples{
some_function()
}
