% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-MosquitoPop-Class.R
\docType{class}
\name{MacroMosquitoPop}
\alias{MacroMosquitoPop}
\title{MACRO Mosquito Population Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
MacroMosquitoPop
}
\description{
This is a generic MACRO mosquito population blah blah ...
below i describe the basic structure of the patch. methods and fields for specific COMPONENTS can be found in:
\itemize{
\item somewhere 1
\item somewhere 2
}
}
\details{
talk about me in detail!
}
\section{Methods}{

\itemize{
\item Constructor:
\item another method
}
}

\section{Fields}{

\itemize{
\item p:
\item f:
\item Q:
\item v:
\item EIP:
\item maxEIP:
\item M:
\item Y:
\item Z:
\item ZZ:
\item psi:
\item P:
\item TilePointer:
\item PatchesPointer:
\item HumansPointer:
}
}

\keyword{R6}
\keyword{class}
