% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Pointers.R
\docType{class}
\name{macroPointer}
\alias{macroPointer}
\title{MACRO Pointer Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
macroPointer
}
\description{
This is a generic MACRO pointer blah blah ...
this class does some stuff:
\itemize{
\item stuff 1
\item stuff 2
}
}
\details{
talk about me in detail!
}
\section{public}{

\itemize{
\item{\code{initialize(N)}}{
talk about me!
}
\item{\code{function()}}{
talk about me!
}
}
}

\section{private}{

\itemize{
\item{\code{function()}}{
talk about me!
}
\item{\code{function()}}{
talk about me!
}
}
}

\section{Active Bindings}{

\itemize{
\item{\code{coolActiveBinding}}{...}
}
}

\keyword{R6}
\keyword{class}
