% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-LANDSCAPE.R
\name{makeAquaSite}
\alias{makeAquaSite}
\title{Generate a Aquatic Habitat Object}
\usage{
makeAquaSite(siteIx, siteXY, searchWt, landHaz, aquaMod, nBatches = 5,
  nAdults = 5)
}
\arguments{
\item{siteIx}{index of this site}

\item{siteXY}{length 2 numeric vector of xy-coordinates}

\item{searchWt}{search weight for MvOb kernel}

\item{landHaz}{resting or landing hazard}

\item{aquaMod}{string; either "el4p" or "emerge" depending on which AQUATIC ECOLOGY module is used}

\item{nBatches}{number of egg batches to pre-allocate for EggQ}

\item{nAdults}{number of adult slots fo pre-allocate for ImagoQ}
}
\value{
a named list
}
\description{
This wil create a single aquatic habitat object for MICRO-LANDSCAPE
}
\examples{
makeAquaSite(siteIx, siteXY, searchWt, landHaz, aquaMod, nBatches = 5, nAdults = 5)
}
