% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-LANDSCAPE.R
\name{makeFeedingSite}
\alias{makeFeedingSite}
\title{Generate a Feeding Site Object}
\usage{
makeFeedingSite(siteIx, siteXY, humanN, humanIx, searchWt, landHaz, enterHouseP)
}
\arguments{
\item{siteIx}{index of this site}

\item{siteXY}{length 2 numeric vector of xy-coordinates}

\item{humanN}{number of humans who live here}

\item{humanIx}{indices of humans who live here}

\item{searchWt}{search weight for MvOb kernel}

\item{landHaz}{resting or landing hazard}

\item{enterHouseP}{probability for mosquito to enter the house}
}
\value{
a named list
}
\description{
This wil create a single feeding site object for MICRO-LANDSCAPE
}
\examples{
makeFeedingSite(siteIx , siteXY, humanN, humanIx, searchWt, landHaz, enterHouseP)
}
