% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-Emerge-Methods.R
\name{makeLambda_MicroEmerge}
\alias{makeLambda_MicroEmerge}
\title{MICRO: Generate Seasonal Emergence (Lambda) for \code{Emerge} model of Aquatic Ecology}
\usage{
makeLambda_MicroEmerge(aquaPars)
}
\arguments{
\item{aquaPars}{a list of the following structure
\itemize{
\item N: number of aquatic habitats (required)
\item lambda: number of emerging adult females per human per day averaged over one year for the entire \code{\link{Landscape}} (required)
\item lambdaWeight: vector of weights applied to each site (if not specified or set to \code{NULL} initialize to Gamma(1,1) distribution)
\item offset: vector of seasonal offsets in peak emergence applied to each site (if not specified or set to \code{NULL} initialize to 0 for all sites)
}}
}
\value{
list \code{lambda} where each element is the daily emergence for that \code{\link{FeedingSite}}
}
\description{
Generate lambda for all sites.
}
\examples{
makeLambda_MicroEmerge(aquaPars= list(lambda = c(2,3,4)))
}
