% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-LANDSCAPE.R
\name{makeLandscape}
\alias{makeLandscape}
\title{Generate the LANDSCAPE Object}
\usage{
makeLandscape(nF, nA, nS, nM, aquaMod, pointGen, hhSize, hhMin = 2,
  xLim = c(0, 1), yLim = c(0, 1), aquaSD = 0.01, swarmSD = 0.05, ...)
}
\arguments{
\item{nF}{number of feeding sites (generated via pointGen(...))}

\item{nA}{number of aquatic habitats}

\item{nS}{number of sugar sites (generated via pointGen(...))}

\item{nM}{number of swarming sites}

\item{aquaMod}{string; "emerge" or "el4p"}

\item{pointGen}{point generation function}

\item{hhSize}{average number of hosts at feeding sites}

\item{hhMin}{minimum number of hosts at feeding sites}

\item{xLim}{x-axis bounds for simulated points}

\item{yLim}{y-axis bounds for simulated points}

\item{aquaSD}{standard deviation of aquatic habitat scatter around feeding sites}

\item{swarmSD}{standard deviation of swarm site scatter around feeding sites}

\item{...}{additional named arguments for pointGen()}
}
\value{
LANDSCAPE object
}
\description{
Generate the global LANDSCAPE object for MICRO-LANDSCAPE
}
\examples{
makeLandscape(nF, nA, nS, nM, aquaMod, pointGen, hhSize, hhMin = 2, xLim=c(0,1), yLim=c(0,1), aquaSD = 0.01, swarmSD = 0.05, ...)
}
