% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Calibration.R
\name{makePAR_EL4P}
\alias{makePAR_EL4P}
\title{Calculate Parameters for EL4P Aquatic Ecology Module}
\usage{
makePAR_EL4P(nA, nH, R0, M, aquaEq, par, summary, p = 0.9, P = 0.8, a = 1,
  b = 1)
}
\arguments{
\item{nA}{number of aquatic habitats}

\item{nH}{number of humans}

\item{R0}{desired pathogen transmission level at equilibrium}

\item{M}{vector of densities of mosquitoes at each aquatic habitat}

\item{aquaEq}{vector of probability of oviposition at each aquatic habitat (see \code{\link{aquaIx_equilibrium}})}

\item{par}{parameter list from \code{\link{MBITES.PAR}}}

\item{summary}{output from \code{\link{cohortBionomics}} which can be calculated from a run of MASH, or, less laboriously, from \code{\link{MBITES.basic}}.}

\item{p}{expected fraction of cohort that advances to next life stage (1/p is expected time spent in each stage E,L1,L2,L3,L4,P)}

\item{P}{density independent survival from egg to adult}

\item{a}{shape parameter for gamma distributed weights on K}

\item{b}{scale parameter for gamma distributed weights on K}
}
\value{
named list of fitted parameters
\itemize{
\item M: M
\item aquaEq: aquaEq
\item nA: nA
\item nH: nH
\item lifespan: mean length of mosquito lifespan
\item G: mean total lifetime egg production of an adult female
\item nu: mean batch size for one oviposition
\item p: p
\item P: P
\item a: a
\item b: b
\item lambda: daily female emergence required to sustain transmission at input R0 (see \code{\link{getLambda}})
}
}
\description{
Calculate parameters needed for EL4P fitting.
}
