% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfPedigree.R
\name{makePf0}
\alias{makePf0}
\title{Generate the Pf Object for Simulated Biting}
\usage{
makePf0(ixH, tBite, ixS = NULL)
}
\arguments{
\item{ixH}{index of human}

\item{tBite}{time of bite}

\item{ixS}{NULL; index of site}
}
\value{
a list of two elements
\itemize{
\item spz: TRUE
\item PfM: The PfM object
\itemize{
\item tm: time of human to vector transmission
\item ixS: NULL
\item ixH: site of human to vector transmission
\item damID: NULL
\item sireID: NULL
\item pfid: NULL
}
}
}
\description{
This generates the Pf object for \code{\link{simbite_PfSI}}. It does not have the element \code{infected} in the returned list
because this is for vector to human transmission; therefore \code{spz} must be set to \code{TRUE}.
}
\examples{
makePfM(ixH, tBite, ixS)
}
