% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfPedigree.R
\name{makePfM}
\alias{makePfM}
\title{Generate the Pf Object for Mosquito Stage of Parasite Life Cycle}
\usage{
makePfM(ixH, tBite, ixS)
}
\arguments{
\item{ixH}{index of human}

\item{tBite}{time of bite}

\item{ixS}{index of site}
}
\value{
a list of two elements
\itemize{
\item infected: TRUE
\item PfM: The PfM object
\itemize{
\item tm: time of human to vector transmission
\item ixS: site of human to vector transmission
\item ixH: site of human to vector transmission
\item damID: female gametocyte IDs
\item sireID: male gametocyte IDs
\item pfid: Pf ID from the infected human
}
}
}
\description{
During successful human to mosquito transmission makePfM() generates the PfM object that is passed from human to mosquito.
This object contains all of the Pf information necessary for the module.
}
\examples{
makePfM(ixH, tBite, ixS)
}
