% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-LANDSCAPE.R
\name{makeSugarSite}
\alias{makeSugarSite}
\title{Generate a Sugar Feeding Site Object}
\usage{
makeSugarSite(siteIx, siteXY, searchWt, landHaz)
}
\arguments{
\item{siteIx}{index of this site}

\item{siteXY}{length 2 numeric vector of xy-coordinates}

\item{searchWt}{search weight for MvOb kernel}

\item{landHaz}{resting or landing hazard}
}
\value{
a named list
}
\description{
This wil create a single sugar feeding site object for MICRO-LANDSCAPE
}
\examples{
makeSugarSite(siteIx, siteXY, searchWt, landHaz)
}
