% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Macro-EL4P.R
\name{manage_PatchesImagoQ}
\alias{manage_PatchesImagoQ}
\title{Manage \code{MacroPatch} ImagoQ (Emerging Adults Queue)}
\usage{
manage_PatchesImagoQ(ixP, N = 10L)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
If all the ImagoQ slots for patch \code{ixP} are full then call \code{\link{extend_PatchesImagoQ}}
}
\examples{
some_function()
}
