% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{MASH-package}
\alias{MASH}
\alias{MASH-package}
\title{MASH (Modular Analysis and Simulation for Human Health)}
\description{
MASH is a generic, modular, and extensible modeling framework for the modeling and analysis of
mosquito-borne pathogens.
}
\details{
Please refer to \url{http://smitdave.github.io/MASH-Development/index.html} for documentation, citations, publications,
acknowledgements, and examples. Please refer to https://smitdave.github.io/MASH/index.html for reference and vignettes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/smitdave/MASH}
  \item Report bugs at \url{https://github.com/smitdave/MASH/issues}
}

}
\author{
\strong{Maintainer}: David L. Smith \email{smitdave@gmail.com}

Authors:
\itemize{
  \item Héctor Manuel Sánchez Castellanos \email{sanchez.hmsc@gmail.com}
  \item Sean Wu \email{slwu89@berkeley.edu}
  \item Robert C. Reiner, Jr. \email{bcreiner@uw.edu}
  \item Amit Verma \email{amit.verma13@gmail.com}
}

Other contributors:
\itemize{
  \item Arnaud Le Menach [contributor]
  \item Nick Ruktanonchai [contributor]
  \item Samson Kiware [contributor]
  \item Gerry Killeen [contributor]
  \item Tom Scott [contributor]
  \item Ellis McKenzie [contributor]
  \item Steven W. Lindsay [contributor]
  \item Willem Takken [contributor]
  \item Philip Eckhoff [contributor]
  \item Nicole Achee [contributor]
  \item Chris Barker [contributor]
  \item Nakul Chitnis [contributor]
  \item Justin Cohen [contributor]
  \item Su Yun Kang [contributor]
  \item Audrey Lenhart [contributor]
  \item John Marshall [contributor]
  \item Phil McCall [contributor]
  \item Catherine Moyes [contributor]
  \item Doug Norris [contributor]
  \item Alex Perkins [contributor]
  \item Chris Stone [contributor]
  \item Edward Wenger [contributor]
  \item Anne Wilson [contributor]
  \item Andrew J. Dolgert [contributor]
}

}
