% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO-Cohort-Setup.R
\name{MBITES.BRO.Cohort.Setup}
\alias{MBITES.BRO.Cohort.Setup}
\title{MBITES-BRO-Cohort: Initialize Additional Methods & Fields in \code{\link{MicroMosquitoPopFemale}} and \code{\link{MicroMosquitoFemale}}}
\usage{
MBITES.BRO.Cohort.Setup(overwrite = TRUE)
}
\arguments{
\item{overwrite}{overwrite methods}
}
\description{
WRITE MEEEEE Initialize M-BITES BRO (Blood Feeding, Resting, Oviposition) lifecycle model.
}
\examples{
MBITES.BRO.Setup()
}
