% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO-Parameters.R
\name{MBITES.BRO.Parameters}
\alias{MBITES.BRO.Parameters}
\title{Generate Parameters for M-BITES 'BRO' Module}
\usage{
MBITES.BRO.Parameters(B_wts = rep(1, 6), R_wts = rep(1, 6), O_wts = rep(1,
  6), InAndOut = matrix(c(c(4, 2, 1, 0, 1), c(2, 1, 1, 0, 1), c(2, 1, 1, 0,
  2), c(1, 1, 1, 0, 1), c(6, 4, 2, 1, 0)), byrow = FALSE, ncol = 5),
  B_surv = (0.98 * 0.9), bfa.s = 0.3, B_time = 2/3, Q = 0.9,
  bfs.L = 0.1, bfs.E = 0.95, F_time = 2/24, bms.a = 7.5, bms.b = 2.5,
  bms.z = 0.01, rf.a = 18, rf.b = 10000, of.a = 5, of.b = 500,
  R_surv = 0.9, R_time = 0.8, reFeed = 0.2, O_surv = (0.9 * 0.9),
  ela.s = 0.7, O_time = 1, L_time = 0.3, els.a = 0.95, sns.a = 0.1,
  sns.b = 100, sgr.d = 1/10, sgr.p = 0.95, sgr.t = 1/24, sgr.a = 50,
  sgr.b = 10, sgr.sa = 50, sgr.sb = 200, sgr.sz = 1, sgr.cvr = 0.2,
  ttsz.p = 0.5, ttsz.a = 5, ttsz.b = 95, ttr.a = 10, ttr.b = 500,
  rpl.cvr = 0, rpl.rpl = 0, rpl.die = 0, eggT = 0, eggP = 0,
  energyPreG = 0, Pathogen = "PfSI")
}
\arguments{
\item{B_wts}{landing spot weights}

\item{R_wts}{landing spot weights}

\item{O_wts}{landing spot weights}

\item{InAndOut}{matrix of landing probabilities}

\item{bfa.s}{Blood Feed Attempt: Prob Succeeds}

\item{B_time}{Blood Feed Attempt: Mean Time Elapsed (in Days)}

\item{Q}{Proportion of Bites on Humans}

\item{bfs.L}{Blood Feed Search: Prob Leaves Location}

\item{bfs.E}{Blood Feed Search: Prob Enters New Location}

\item{F_time}{Blood Feed Search: Mean Time Elapsed (in Days)}

\item{bms.a}{Blood Meal Size, a}

\item{bms.b}{Blood Meal Size, b}

\item{bms.z}{Blood Feeding Hazard}

\item{rf.a}{ReFeed WRITE ME}

\item{rf.b}{ReFeed WRITE ME}

\item{of.a}{Overfeed WRITE ME}

\item{of.b}{Overfeed WRITE ME}

\item{R_time}{Resting: Mean Time Elapsed (in Days)}

\item{reFeed}{WRITE ME}

\item{ela.s}{Egg Laying Attempt: Prob Success}

\item{O_time}{Egg Laying Attempt: Mean Time Elapsed}

\item{L_time}{Egg Laying Search: Mean Time Elapsed}

\item{els.a}{Egg Laying Search: Prob Accepts}

\item{sns.a}{WRITE ME}

\item{sns.b}{WRITE ME}

\item{sgr.d}{WRITE ME}

\item{sgr.p}{WRITE ME}

\item{sgr.t}{WRITE ME}

\item{sgr.a}{WRITE ME}

\item{sgr.b}{WRITE ME}

\item{sgr.sa}{WRITE ME}

\item{sgr.sb}{WRITE ME}

\item{sgr.sz}{WRITE ME}

\item{sgr.cvr}{WRITE ME}

\item{ttsz.p}{WRITE ME}

\item{ttsz.a}{WRITE ME}

\item{ttsz.b}{WRITE ME}

\item{ttr.a}{WRITE ME}

\item{ttr.b}{WRITE ME}

\item{rpl.cvr}{WRITE ME}

\item{rpl.rpl}{WRITE ME}

\item{rpl.die}{WRITE ME}

\item{eggT}{the minimum time before eggs are mature}

\item{eggP}{the minimum provision before eggs are mature}

\item{energyPreG}{pre-gonotrophic energy requirement}

\item{Pathogen}{name of the PATHOGEN module to use, can be in "PfSI", "PfMOI"}

\item{bfa.p}{Blood Feed Attempt: Prob Survives (Base)}

\item{bfs.p}{Blood Feed Search: Prob Survives}

\item{ppr.p}{Resting: Prob Survives (Base)}

\item{ela.p}{Egg Laying Attempt: Prob Survives (Base)}

\item{els.p}{Egg Laying Search: Prob Survives (Base)}
}
\value{
a named list of parameters
}
\description{
Generate named list of parameters used throughout MICRO/M-BITES. All arguments have default values which are listed below before the definition.
}
\examples{
P = MBITES.BRO.Parameters()
}
