% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO-Setup.R
\name{MBITES.BRO.Setup}
\alias{MBITES.BRO.Setup}
\title{MBITES-BRO: Initialize Additional Methods & Fields in \code{\link{MicroMosquitoPopFemale}} and \code{\link{MicroMosquitoFemale}}}
\usage{
MBITES.BRO.Setup(overwrite = TRUE, aquaModule)
}
\arguments{
\item{overwrite}{overwrite methods}

\item{aquaModule}{character in "emerge" or "EL4P"}
}
\description{
Initialize M-BITES BRO (Blood Feeding, Resting, Oviposition) lifecycle model.
}
\examples{
MBITES.BRO.Setup()
}
