% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Parameters.R
\name{MBITES.FULL.Parameters}
\alias{MBITES.FULL.Parameters}
\title{Generate Parameters for M-BITES 'Full' Module}
\usage{
MBITES.FULL.Parameters(enterhouse.a = 9, enterhouse.b = 1, feedHaz.a = 99,
  feedHaz.b = 1, aquaHaz.a = 99, aquaHaz.b = 1, sugarHaz.a = 99,
  sugarHaz.b = 1, swarmHaz.a = 99, swarmHaz.b = 1, F.t = 1,
  B.t = 0.75, R.t = 1.5, L.t = 0.75, O.t = 1, M.t = 1.5, S.t = 0.5,
  F.s = 0.99, B.s = 0.99, L.s = 0.99, O.s = 0.99, M.s = 0.95,
  S.s = 0.95, F.p = 0.95, B.p = 0.98, R.p = 0.98, L.p = 0.8,
  O.p = 0.98, M.p = 0.98, S.p = 0.98, maleM.s = 0.99, maleM.p = 0.95,
  Q = 0.9, reFeed = 0.01, bm.a = 7.5, bm.b = 2.5, rf.a = 60,
  rf.b = 5000, of.a = 5, of.b = 5000, S.a = 20, S.b = 10, S.u = 1/7,
  S.sa = 15, S.sb = 5, sf.a = 7.5, sf.b = 4.5, B.energy = 1/10,
  energyPreG = 0, preGsugar = 0, preGblood = 0, eggP.min = 0.5,
  sns.a = 0.085, sns.b = 100, ttsz.p = 0.5, ttsz.a = 5, ttsz.b = 95,
  ttr.a = 15, ttr.b = 500, maxBatch = 30, E.p = 0.5, Emax = 90,
  Eb = 0.9, eEndm = 180, eEndV = 30, surviveH = 1, probeH = 1,
  surviveprobeH = 1, feedH = 1, surviveL = 1, feedL = 1, SUGAR = TRUE,
  ESTIVATION = FALSE, MATE = TRUE, SENESCE = TRUE, TATTER = TRUE,
  REFEED = TRUE, OVERFEED = TRUE, HISTORY = TRUE, batchSize = "bms",
  eggMatT = "off", InAndOut = matrix(c(c(4, 2, 1, 0, 1), c(2, 1, 1, 0, 1),
  c(2, 1, 1, 0, 2), c(1, 1, 1, 0, 1), c(6, 4, 2, 1, 0)), byrow = FALSE, ncol =
  5), Fwts = rep(1, 5), Rwts = rep(1, 5), Lwts = rep(1, 5),
  Owts = rep(1, 5), Mwts = rep(1, 5), Swts = rep(1, 5), eggT = 0,
  eggP = 0, Pathogen = "PfSI")
}
\arguments{
\item{enterhouse.a}{= 9, # beta house entry parameters}

\item{enterhouse.b}{= 1, # beta house entry parameters}

\item{feedHaz.a}{= 99, # beta feeding site hazard weights}

\item{feedHaz.b}{= 1, # beta feeding site hazard weights}

\item{aquaHaz.a}{= 99, # beta aqua site hazard weights}

\item{aquaHaz.b}{= 1, # beta aqua site hazard weights}

\item{sugarHaz.a}{= 99, # beta sugar site hazard weights}

\item{sugarHaz.b}{= 1, # beta sugar site hazard weights}

\item{swarmHaz.a}{= 99, # beta swarm site hazard weights}

\item{swarmHaz.b}{= 1, # beta swarm site hazard weights}

\item{F.t}{= 1, #blood feed search bout timing}

\item{B.t}{= 0.75, #blood feed attempt bout timing}

\item{R.t}{= 1.5, #resting bout timing}

\item{L.t}{= 0.75, #egg laying search bout timing}

\item{O.t}{= 1, #egg laying attempt bout timing}

\item{M.t}{= 1.5, #mating bout timing}

\item{S.t}{= 0.5, #sugar feeding bout timing}

\item{F.s}{= 0.99,}

\item{B.s}{= 0.99,}

\item{L.s}{= 0.99,}

\item{O.s}{= 0.99,}

\item{M.s}{= 0.95,}

\item{S.s}{= 0.95,}

\item{F.p}{= 0.95,}

\item{B.p}{= 0.98,}

\item{R.p}{= 0.98,}

\item{L.p}{= 0.80,}

\item{O.p}{= 0.98,}

\item{M.p}{= 0.98,}

\item{S.p}{= 0.98,}

\item{maleM.s}{= .99,}

\item{maleM.p}{= .95,}

\item{Q}{= 0.9, #human blood index}

\item{reFeed}{= 0.01, #probability to refeed post resting bout}

\item{bm.a}{= 7.5, #shape param for bloodmeal size}

\item{bm.b}{= 2.5, #shape param for bloodmeal size}

\item{rf.a}{= 60, #exp param for refeeding as function of bmSize}

\item{rf.b}{= 5e3, #exp param for refeeding as function of bmSize}

\item{of.a}{= 5, #exp param for overfeeding as function of bmSize}

\item{of.b}{= 5e3, #exp param for overfeeding as function of bmSize}

\item{S.a}{= 20, #exp param for death as function of energy}

\item{S.b}{= 10, #exp param for death as function of energy}

\item{S.u}{= 1/7, #energy expenditure for each bout}

\item{S.sa}{= 15, #exp param for sugar bout as function of energy}

\item{S.sb}{= 5, #exp param for sugar bout as function of energy}

\item{sf.a}{= 7.5, #shape param for sugar feed size}

\item{sf.b}{= 4.5, #shape param for sugar feed size}

\item{B.energy}{= 1/10, #scaling factor energy from a bloodmeal}

\item{energyPreG}{= 0, #pre-gonotrophic energy requirement}

\item{preGsugar}{= 0,}

\item{preGblood}{= 0,}

\item{sns.a}{= 0.085, #exp param for senescence}

\item{sns.b}{= 100, #exp param for senescence}

\item{ttsz.p}{= 0.5, #zero-inflation for tattering damage}

\item{ttsz.a}{= 5, #shape param for tattering damage}

\item{ttsz.b}{= 95, #shape param for tattering damage}

\item{ttr.a}{= 15, #exp param for tattering survival}

\item{ttr.b}{= 500, #exp param for tattering survival}

\item{maxBatch}{= 30, #maximum batch size}

\item{E.p}{= 0.5, # probability of surviving estivation}

\item{Emax}{= 90, # onset of the dry season; a day of the calendar year}

\item{Eb}{= 0.9,}

\item{eEndm}{= 180, # end of estivation; a day of the calendar year}

\item{eEndV}{= 30,}

\item{surviveH}{= 1,}

\item{probeH}{= 1,}

\item{surviveprobeH}{= 1,}

\item{feedH}{= 1,}

\item{surviveL}{= 1,}

\item{feedL}{= 1,}

\item{SUGAR}{= TRUE,}

\item{ESTIVATION}{= FALSE,}

\item{MATE}{= TRUE,}

\item{SENESCE}{= TRUE, #senesce}

\item{TATTER}{= TRUE, #tattering}

\item{REFEED}{= TRUE,}

\item{OVERFEED}{TRUE; simulate overfeeding? See \code{\link{overFeed}}}

\item{HISTORY}{TRUE; record history? Needed for \code{\link{cohortBionomics}}, see \code{\link{historyTrack}}}

\item{batchSize}{"bms"; switches for how egg batch size is calculated, should be either "bms" or "norm", see \code{\link{BatchSize.bms}} or \code{\link{BatchSize.norm}}}

\item{eggMatT}{"off"; switches for how egg maturation is calculated, should be either "off" or "norm", see \code{\link{eggMaturationTime.off}} or \code{\link{eggMaturationTime.norm}}}

\item{InAndOut}{matrix of weights for landing spot, see \code{\link{newSpot}}}

\item{Fwts}{rep(1,5); blood feeding search bout landing spot weights, see \code{\link{newSpot}}}

\item{Rwts}{rep(1,5); resting bout landing spot weights, see \code{\link{newSpot}}}

\item{Lwts}{rep(1,5); egg laying search bout landing spot weights, see \code{\link{newSpot}}}

\item{Owts}{rep(1,5); egg laying attempt bout landing spot weights, see \code{\link{newSpot}}}

\item{Mwts}{rep(1,5); mating bout landing spot weights, see \code{\link{newSpot}}}

\item{Swts}{rep(1,5); sugar bout landing spot weights, see \code{\link{newSpot}}}

\item{eggT}{the minimum time before eggs are mature}

\item{eggP}{the minimum provision before eggs are mature}

\item{Pathogen}{name of the PATHOGEN module to use, can be in "PfSI", "PfMOI"}
}
\value{
a named list of parameters
}
\description{
Generate named list of parameters used throughout M-BITES 'Full' Module. All arguments have default values which are listed below before the definition.
}
\examples{
P = MBITES.FULL.Parameters()
}
