% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Setup.R
\name{MBITES.Generic.Setup}
\alias{MBITES.Generic.Setup}
\title{Initialize Generic Methods for M-BITES}
\usage{
MBITES.Generic.Setup(overwrite = TRUE, batchSize = "bms", eggMatT = "off")
}
\arguments{
\item{batchSize}{character switch that should be one of \code{"bms","norm"} for egg batch sizes dependent on bloodmeal size or normally distributed}

\item{eggMatT}{character switch that should be one of \code{"off","norm"} for egg batch maturation time turned off or normally distributed}
}
\value{
modifies the \code{\link{MicroMosquitoFemale}} and \code{\link{MicroMosquitoMale}} classes.
}
\description{
This function initializes generic methods for M-BITES models; please note that the
switches for this function modify only the methods that are added to the \code{\link{MicroMosquitoFemale}}
and \code{\link{MicroMosquitoMale}} classes.
}
