% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO-Setup.R
\name{mbitesBRO.Setup}
\alias{mbitesBRO.Setup}
\title{MBITES-BRO: Initialize Additional Methods & Fields in \code{MicroMosquitoPop} and \code{MicroMosquito}}
\usage{
mbitesBRO.Setup(overwrite = TRUE)
}
\arguments{
\item{overwrite}{overwrite methods}
}
\description{
Initialize M-BITES BRO (Blood Feeding, Resting, Oviposition) lifecycle model.
}
\examples{
mbitesBRO.Setup()
}
