% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO-Energetics.R
\name{mbitesBRO_BloodMeal}
\alias{mbitesBRO_BloodMeal}
\title{MBITES-BRO: Bloodmeal Energetics for \code{\link{MicroMosquitoFemale}}}
\usage{
mbitesBRO_BloodMeal()
}
\description{
Handle bloodmeal size (see \code{\link{mbitesGeneric_rBloodMealSize}}), overfeeding (see \code{\link{mbitesGeneric_pOverFeed}}),
and egg batch size and maturation time.
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$BloodMeal()}.
}
}
