% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO-Bouts.R
\name{mbitesBRO_boutB}
\alias{mbitesBRO_boutB}
\title{MBITES-BRO: Blood Feeding Bout \code{MicroMosquitoFemale}}
\usage{
mbitesBRO_boutB()
}
\description{
A mosquito performs a blood feeding bout (all actions taken launch to launch when blood feeding required).
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$boutB()}.
}
}
