% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO-Bouts.R
\name{mbitesBRO_boutO}
\alias{mbitesBRO_boutO}
\title{MBITES-BRO: Egg Laying Bout \code{\link{MicroMosquitoFemale}}}
\usage{
mbitesBRO_boutO()
}
\description{
A mosquito performs a oviposition bout (all actions taken launch to launch when oviposition required).
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$boutO()}.
}
}
