% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO-Bouts.R
\name{mbitesBRO_boutR}
\alias{mbitesBRO_boutR}
\title{MBITES-BRO: Post-Prandial Resting Bout \code{MicroMosquitoFemale}}
\usage{
mbitesBRO_boutR()
}
\description{
A mosquito performs a post-prandial resting bout (all actions taken launch to launch when resting required).
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$boutR()}.
}
}
