% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO-Cohort.R
\name{mbitesBRO_cohort_humanEncounter}
\alias{mbitesBRO_cohort_humanEncounter}
\title{MBITES-BRO-Cohort: Human Host Encounter for \code{\link{MicroMosquitoFemale}}}
\usage{
mbitesBRO_cohort_humanEncounter()
}
\description{
After calling \code{\link{mbitesGeneric_chooseHost}}, the mosquito encounters a human host and attempts to feed.
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$humanEncounter()}.
}
}
