% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO-Cohort.R
\name{mbitesBRO_cohort_oneMosquito_MBITES}
\alias{mbitesBRO_cohort_oneMosquito_MBITES}
\title{MBITES-BRO-Cohort: Run Simulation for \code{\link{MicroMosquitoFemale}}}
\usage{
mbitesBRO_cohort_oneMosquito_MBITES()
}
\description{
Run the M-BITES life cycle simulation algorithm until mosquito is dead.
This method calls \code{\link{mbitesBRO_oneBout}} to simulate each life stage.
For differences between cohort and microsimulation versions see \code{\link{MBITES.BRO.Cohort.Setup}}
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$MBITES_Cohort()}.
}
}
