% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO-Cohort.R
\name{mbitesBRO_cohort_simCohort}
\alias{mbitesBRO_cohort_simCohort}
\title{MBITES-BRO-Cohort: Human Host Encounter for \code{\link{MicroMosquitoFemale}}}
\usage{
mbitesBRO_cohort_simCohort(N, writeJSON)
}
\arguments{
\item{N}{number of mosquitoes in cohort}

\item{writeJSON}{if \code{TRUE} write output to JSON in the directory initialized in the enclosing \code{\link{MicroTile}}, else return a list}
}
\description{
After calling \code{\link{mbitesGeneric_chooseHost}}, the mosquito encounters a human host and attempts to feed.
\itemize{
\item This method is bound to \code{MicroMosquitoPopFemale$simCohort()}.
}
}
