% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO-Bouts.R
\name{mbitesBro_enterHouse}
\alias{mbitesBro_enterHouse}
\title{MBITES-BRO: Attempt to Enter a House for \code{MicroMosquitoFemale}}
\usage{
mbitesBro_enterHouse()
}
\description{
Method to simulate attempted house entry for mosquito, and call appropriate events if the mosquito enters.
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$enterHouse()}.
}
}
