% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO-Survival.R
\name{mbitesBRO_get_restHaz}
\alias{mbitesBRO_get_restHaz}
\title{MBITES-BRO: Get Resting Hazards for \code{\link{MicroMosquitoFemale}}}
\usage{
mbitesBRO_get_restHaz()
}
\description{
Get resting hazards for \code{\link{mbitesBRO_surviveFlight}}.
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$get_restHaz()}.
}
}
