% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO-Survival.R
\name{mbitesBRO_get_surviveFlightProb}
\alias{mbitesBRO_get_surviveFlightProb}
\title{MBITES-BRO: Get Baseline Survival Probability for \code{\link{MicroMosquitoFemale}}}
\usage{
mbitesBRO_get_surviveFlightProb()
}
\description{
Get baseline survival probability for \code{\link{mbitesGeneric_surviveFlight}}.
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$get_surviveFlightProb()}.
}
}
