% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO-Bouts.R
\name{mbitesBRO_getMySiteType}
\alias{mbitesBRO_getMySiteType}
\title{MBITES-BRO: Return Site Type \code{MicroMosquitoFemale}}
\usage{
mbitesBRO_getMySiteType()
}
\value{
vector of landing spot weights
}
\description{
Method to return integer corresponding to site type of \code{\link{MicroSite}} this mosquito is currently at.
Site Types:
\itemize{
\item 1: peri-domestic site
\item 0: not peri-domestic site
}
}
\details{
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$getMySiteType()}.
}
}
