% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO-HostEncounter.R
\name{mbitesBRO_humanEncounter}
\alias{mbitesBRO_humanEncounter}
\title{MBITES-BRO: Human Host Encounter for \code{\link{MicroMosquitoFemale}}}
\usage{
mbitesBRO_humanEncounter()
}
\description{
After calling \code{\link{mbitesGeneric_chooseHost}}, the mosquito encounters the human host and attempts to feed.
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$humanEncounter()}.
}
}
