% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO-Bouts.R
\name{mbitesBRO_layEggs_EL4P}
\alias{mbitesBRO_layEggs_EL4P}
\title{MBITES-BRO: Lay Eggs for 'EL4P' \code{\link{MicroMosquitoFemale}}}
\usage{
mbitesBRO_layEggs_EL4P()
}
\description{
During an oviposition bout \code{\link{mbitesBRO_boutO}}, lay eggs for 'EL4P' module of Aquatic Ecology.
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$layEggs()}.
}
}
