% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO-Bouts.R
\name{mbitesBRO_layEggs_Emerge}
\alias{mbitesBRO_layEggs_Emerge}
\title{MBITES-BRO: Lay Eggs for 'Emerge' \code{\link{MicroMosquitoFemale}}}
\usage{
mbitesBRO_layEggs_Emerge()
}
\description{
During an oviposition bout \code{\link{mbitesBRO_boutO}}, lay eggs (this is just a filler to clear out the \code{batch} field of the mosquito; egg laying is not implemented in any modules relying on "Emerge" Aquatic Ecology module)
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$layEggs()}.
}
}
