% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO-Bouts.R
\name{mbitesBro_MBITES}
\alias{mbitesBro_MBITES}
\title{MBITES-BRO: Run Simulation for \code{MicroMosquitoFemale}}
\usage{
mbitesBro_MBITES()
}
\description{
Run the M-BITES life cycle simulation algorithm while alive and has not overrun time in enclosing \code{\link{MicroTile}}.
This method calls \code{\link{mbitesBro_oneBout}} to simulate each life stage.
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$MBITES()}.
}
}
