% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO-Bouts.R
\name{mbitesBRO_newSpot}
\alias{mbitesBRO_newSpot}
\title{MBITES-BRO: Generate New Landing Spot for \code{MicroMosquitoFemale}}
\usage{
mbitesBRO_newSpot()
}
\value{
integer value corresponding to new landing spot
}
\description{
Method for return a new landing spot based on behavioral state of mosquito and weights from \code{\link{mbitesBRO_get_WTS}}.
New landing spots generated at the end of the search bout, attempt bout, or after oviposition a mosquito has entered
the area around a feeding site and either rested or attempted to rest.
\itemize{
\item i: 1 rest on the inside wall of a structure
\item w: 2 rest on the outside wall of a structure
\item v: 3 rest on vegetation
\item r: 4 reattempt without resting
\item l: 5 leave the area
}
}
\details{
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$newSpot()}.
}
}
