% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO-Bouts.R
\name{mbitesBRO_oneMosquito_MBITES}
\alias{mbitesBRO_oneMosquito_MBITES}
\title{MBITES-BRO: Run Simulation for \code{\link{MicroMosquitoFemale}}}
\usage{
mbitesBRO_oneMosquito_MBITES()
}
\description{
Run the M-BITES life cycle simulation algorithm while alive and has not overrun time in enclosing \code{\link{MicroTile}}.
This method calls \code{\link{mbitesBRO_oneBout}} to simulate each life stage.
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$MBITES()}.
}
}
