% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO.R
\name{MbitesBRO_timingExponential}
\alias{MbitesBRO_timingExponential}
\title{MBITES-BRO: Exponential Timing for \code{MicroMosquitoFemale}}
\usage{
MbitesBRO_timingExponential()
}
\description{
Method for exponentially-distributed bout lengths (model mosquito walk through state space as a strictly Markovian process).
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$timingExponential()}.
}
}
