% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO-HostEncounter.R
\name{mbitesBRO_zooEncounter}
\alias{mbitesBRO_zooEncounter}
\title{MBITES-BRO: Non-Human Host Encounter for \code{\link{MicroMosquitoFemale}}}
\usage{
mbitesBRO_zooEncounter()
}
\description{
After calling \code{\link{mbitesGeneric_chooseHost}}, the mosquito encounters the non-human host and attempts to feed.
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$zooEncounter()}.
}
}
