% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-ChooseHost.R
\name{mbitesCohort_chooseHost}
\alias{mbitesCohort_chooseHost}
\title{MBITES-Cohort: Host Choosing for \code{\link{MicroMosquitoFemale}}}
\usage{
mbitesCohort_chooseHost()
}
\description{
Choose a human or animal host for MBITES-Cohort algorithms (moquito life-cycle simulations).
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$chooseHost()}.
}
}
