% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Setup.R
\name{mbitesGeneric.Setup}
\alias{mbitesGeneric.Setup}
\title{Initialize Generic Methods for M-BITES}
\usage{
mbitesGeneric.Setup(overwrite = TRUE, batchSize = "bms", eggMatT = "off")
}
\arguments{
\item{batchSize}{character switch that should be one of \code{"bms","norm"} for egg batch sizes dependent on bloodmeal size or normally distributed}

\item{eggMatT}{character switch that should be one of \code{"off","norm"} for egg batch maturation time turned off or normally distributed}
}
\value{
modifies the \code{MicroMosquitoFemale} and \code{MicroMosquitoMale} classes.
}
\description{
This function initializes generic methods for M-BITES models; please note that the
switches for this function modify only the methods that are added to the MicroMosquitoFemale
and MicroMosquitoMale classes. Different genotypes still depend on the internal list of parameters
to parameterize these functions and functional forms for equations.
}
