% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-ChooseHost.R
\name{mbitesGeneric_chooseHost}
\alias{mbitesGeneric_chooseHost}
\title{MBITES-Generic: Host Choosing for \code{\link{MicroMosquitoFemale}}}
\usage{
mbitesGeneric_chooseHost()
}
\description{
Choose a human or animal host at a site.
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$chooseHost()}.
}
}
