% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Generic-Bouts.R
\name{mbitesGeneric_enterHouse}
\alias{mbitesGeneric_enterHouse}
\title{M-BITES (All Modules): Attempt to Enter a House for \code{MicroMosquitoFemale}}
\usage{
mbitesGeneric_enterHouse()
}
\description{
Method to simulate attempted house entry for mosquito, and call appropriate events if the mosquito enters.
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$enterHouse()}.
}
}
