% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Generic.R
\name{MbitesGeneric_FemaleBout}
\alias{MbitesGeneric_FemaleBout}
\title{MBITES: Generic Female Bout for \code{MicroMosquitoFemale}}
\usage{
MbitesGeneric_FemaleBout()
}
\description{
Manage lifecycle methods across modules.
}
\section{Generic Bout}{

\itemize{
\item hello.
}
}

