% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Generic.R
\name{MbitesGeneric_isActive}
\alias{MbitesGeneric_isActive}
\title{MBITES: Generic Active Check for \code{MicroMosquito}}
\usage{
MbitesGeneric_isActive()
}
\description{
Check if this mosquito is active and return a logical value.
\itemize{
\item This method is bound to \code{MicroMosquito$isActive()}.
}
}
