% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Generic.R
\name{MbitesGeneric_isAlive}
\alias{MbitesGeneric_isAlive}
\title{MBITES: Generic Alive Check for \code{MicroMosquito}}
\usage{
MbitesGeneric_isAlive()
}
\description{
Check if this mosquito is alive and return a logical value.
\itemize{
\item This method is bound to \code{MicroMosquito$isAlive()}.
}
}
