% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Generic-Bouts.R
\name{mbitesGeneric_newSpot}
\alias{mbitesGeneric_newSpot}
\title{M-BITES (All Modules): Generate New Landing Spot for \code{MicroMosquitoFemale}}
\usage{
mbitesGeneric_newSpot()
}
\value{
integer value corresponding to new landing spot
}
\description{
Method for return a new landing spot based on behavioral state of mosquito and weights from \code{\link{mbitesBRO_getWTS}}.
New landing spots generated at the end of the search bout, attempt bout, or after oviposition a mosquito has entered
the area around a feeding site and either rested or attempted to rest.
\itemize{
\item l: 1 leave the area
\item r: 2 reattempt without resting
\item v: 3 rest on vegetation
\item w: 4 rest on the outside wall of a structure
\item i: 5 rest on the inside wall of a structure
}
}
\details{
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$newSpot()}.
}
}
