% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Energetics.R
\name{mbitesGeneric_pReFeed}
\alias{mbitesGeneric_pReFeed}
\title{MBITES-Generic: Probaility to re-enter Blood Feeding Cycle from Incomplete Feeding for \code{\link{MicroMosquitoFemale}}}
\usage{
mbitesGeneric_pReFeed()
}
\description{
Probability to re-enter blood feeding cycle after incomplete blood feeding given by \deqn{ \frac{2+rf.b}{1+rf.b}-\frac{e^{rf.a\times bmSize}}{rf.b+e^{rf.a\times bmSize}} }
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$pReFeed()}.
}
}
