% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Survival.R
\name{mbitesGeneric_pSenesce}
\alias{mbitesGeneric_pSenesce}
\title{MBITES-Generic: Probability of Death due to Senescence for \code{\link{MicroMosquitoFemale}}}
\usage{
mbitesGeneric_pSenesce()
}
\description{
probability of death due to senescence given by \deqn{ \frac{2+sns.b}{1+sns.b} - \frac{e^{sns.a\times age}}{sns.b+e^{sns.a\times age}} }
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$pSenesce()}.
}
}
