% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Survival.R
\name{mbitesGeneric_pTatter}
\alias{mbitesGeneric_pTatter}
\title{MBITES-Generic: Probability of Death due to Wing Tattering for \code{\link{MicroMosquitoFemale}}}
\usage{
mbitesGeneric_pTatter()
}
\description{
probability of death due to tattering given by \deqn{ \frac{2+ttr.b}{1+ttr.b} - \frac{e^{damage\times ttr.a}}{ttr.b+e^{damage\times ttr.a}} }
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$pTatter()}.
}
}
