% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Energetics.R
\name{mbitesGeneric_rBatchSizeBms}
\alias{mbitesGeneric_rBatchSizeBms}
\title{MBITES-Generic: Egg Batch Size due to Bloodmeal Size for \code{\link{MicroMosquitoFemale}}}
\usage{
mbitesGeneric_rBatchSizeBms()
}
\description{
Give an egg batch size given by \deqn{ bmSize\times maxBatch }
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$rBatchSize()}.
}
}
