% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Energetics.R
\name{mbitesGeneric_rBloodMealSize}
\alias{mbitesGeneric_rBloodMealSize}
\title{MBITES-Generic: Draw Bloodmeal Size for \code{\link{MicroMosquitoFemale}}}
\usage{
mbitesGeneric_rBloodMealSize()
}
\description{
Draw a bloodmeal size from Beta(bm.a,bm.b)
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$rBloodMealSize()}.
}
}
