% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Energetics.R
\name{mbitesGeneric_rEggMaturationTimeNorm}
\alias{mbitesGeneric_rEggMaturationTimeNorm}
\title{MBITES-Generic: Normally-distributed Egg Maturation Time for \code{\link{MicroMosquitoFemale}}}
\usage{
mbitesGeneric_rEggMaturationTimeNorm()
}
\description{
Draw an egg maturation time from Normal(emt.m,emt.v)
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$rEggMaturationTime()}.
}
}
