% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Energetics.R
\name{mbitesGeneric_rEggMaturationTimeOff}
\alias{mbitesGeneric_rEggMaturationTimeOff}
\title{MBITES-Generic: No Egg Maturation Time for \code{\link{MicroMosquitoFemale}}}
\usage{
mbitesGeneric_rEggMaturationTimeOff()
}
\description{
Instant egg maturation.
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$rEggMaturationTime()}.
}
}
