% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Survival.R
\name{mbitesGeneric_rTatterSize}
\alias{mbitesGeneric_rTatterSize}
\title{MBITES-Generic: Wing Tattering for \code{\link{MicroMosquitoFemale}}}
\usage{
mbitesGeneric_rTatterSize()
}
\description{
Draw from a zero-inflated Beta distribution for additive wing damager from tattering.
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$rTatterSize()}.
}
}
