% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Survival.R
\name{mbitesGeneric_surviveFlight}
\alias{mbitesGeneric_surviveFlight}
\title{MBITES-Generic: Flight Survival for \code{\link{MicroMosquitoFemale}}}
\usage{
mbitesGeneric_surviveFlight()
}
\description{
Run generic flight survival probailities for bouts (launch to launch).
Depending on settings from M-BITES parameters, senescence and/or tattering may also be simulated.
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$surviveFlight()}.
}
}
