% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Survival.R
\name{mbitesGeneric_surviveResting}
\alias{mbitesGeneric_surviveResting}
\title{MBITES-Generic: Resting Survival for \code{\link{MicroMosquitoFemale}}}
\usage{
mbitesGeneric_surviveResting()
}
\description{
Run generic resting survival probailities for bouts (launch to launch).
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$surviveResting()}.
}
}
