% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Calibration.R
\name{meshK_EL4P}
\alias{meshK_EL4P}
\title{Fit Psi on Mesh of K Values}
\usage{
meshK_EL4P(lK, uK, EL4P_pops, PAR, tMax = 500, tol = 0.1)
}
\arguments{
\item{lK}{lower bound of K mesh}

\item{uK}{upper bound of K mesh}

\item{EL4P_pops}{list of aquatic populations}

\item{PAR}{list of parameters from \code{\link{makePAR_EL4P}}}

\item{tMax}{maximum time to to run aquatic populations to equilibrium}

\item{tol}{target minimum variance in lambda for aquatic populations equilibrium}

\item{alpha}{vector of density independent mortality}
}
\value{
named list of output
\itemize{
\item EL4P_pops: aquatic populations run to equilibrium values
\item psiHat: fitted values of psi
\item meshK: sampled values of K
}
}
\description{
Generate a mesh of K values in log space and fit psi to mesh via numerical optization, then run input EL4P populations to equilibrium.
This sets values of psi so that lambda \eqn{\lambda=K} at given parameter values.
}
\examples{
meshK_EL4P(lK, uK, EL4P_pops, PAR, tMax = 500, tol = 0.1)
}
