% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-Main.R
\name{MICRO.Aqua.Setup}
\alias{MICRO.Aqua.Setup}
\title{Initialize Aquatic Ecology COMPONENT}
\usage{
MICRO.Aqua.Setup()
}
\arguments{
\item{hi}{spacecats}
}
\value{
stuff
}
\description{
This function initializes generic methods and fields for the Aquatic Ecology COMPONENT.
It modifies \code{\link{AquaticSite}} and \code{\link{Landscape}} classes.
}
\examples{
MICRO.Aqua.Setup()
}
