% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-EL4P-Setup.R
\name{MICRO.EL4P.Setup}
\alias{MICRO.EL4P.Setup}
\title{Initialize EL4P Aquatic Ecology Module}
\usage{
MICRO.EL4P.Setup(overwrite = TRUE)
}
\arguments{
\item{dunno}{sdf}
}
\value{
stuff
}
\description{
This function initializes methods and fields for the 'EL4P' MODULE of Aquatic Ecology COMPONENT.
It modifies \code{\link{AquaticSite}} and \code{\link{Landscape}} classes.
}
\examples{
MICRO.EL4P.Setup()
}
