% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-Emerge-Setup.R
\name{MICRO.Emerge.Setup}
\alias{MICRO.Emerge.Setup}
\title{Initialize Emerge Aquatic Ecology Module}
\usage{
MICRO.Emerge.Setup(overwrite = TRUE)
}
\arguments{
\item{dunno}{sdf}
}
\value{
stuff
}
\description{
This function initializes methods and fields for the 'Emerge' MODULE of Aquatic Ecology COMPONENT.
It modifies \code{\link{AquaticSite}} and \code{\link{Landscape}} classes.
}
\examples{
Aqua.Emerge.Setup()
}
