% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Human-Setup.R
\name{MICRO.Humans.Setup}
\alias{MICRO.Humans.Setup}
\title{MICRO: Initialize Additional Methods & Fields in \code{\link{Human}} and \code{\link{HumanPop}}}
\usage{
MICRO.Humans.Setup(overwrite = TRUE)
}
\arguments{
\item{overwrite}{overwrite existing methods}
}
\description{
Write me! \code{\link{Human}}
}
\examples{
MACRO.Humans.Setup()
}
