% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEARCH-MicroKernel-Generate.R
\name{MicroKernel_exactAll}
\alias{MicroKernel_exactAll}
\title{MICRO Search Kernels: Exact Movement Object Between All Point Sets}
\usage{
MicroKernel_exactAll(Landscape, sigma = 3, eps = 0.1, beta = 0)
}
\arguments{
\item{Landscape}{a microsimulation \code{\link{Landscape}} object}

\item{sigma}{a param}

\item{eps}{a param}

\item{beta}{a param}
}
\value{
list of movement objects between all site classes (see above and \code{\link{MicroKernel_exactMvOb}}) for details.
}
\description{
Generate an exact movement object (one-step Markov transition matrix between all sets of sites).
This function returns nested list of the following form for all site classes in landscape:
\itemize{
\item S2D: output of \code{\link{MicroKernel_exactMvOb}} applied to starting site point set S and destination point set D
The output, MvAll, will be assigned to \code{private$movement} field of \code{\link{MicroMosquitoPopFemale}} or \code{\link{MicroMosquitoPopMale}}. The specific accessor can be found at \code{\link{get_MicroKernel_movement}}.
}
}
