% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEARCH-MicroKernel-Generate.R
\name{MicroKernel_exactMvOb}
\alias{MicroKernel_exactMvOb}
\title{MICRO Search Kernels: Exact Movement Object Between Two Point Sets}
\usage{
MicroKernel_exactMvOb(S, D, sigma = 3, eps = 0.1, beta = 0)
}
\arguments{
\item{S}{pointer to starting sites (for example, Landscape$get_AquaSites())}

\item{D}{pointer to destination sites}

\item{sigma}{a param}

\item{eps}{a param}

\item{beta}{a param}
}
\value{
movement object; each element has the following structure:
\itemize{
\item ix: ID of starting site
\item PR: probabilities of near, around, and moveFar movement
\item near: the near movement object
\itemize{
\item id: sorted ID of destination sites
\item pr: sorted probabilities of destination sites
}
\item around:
\item moveFar:
}
}
\description{
Generate an exact movement object (one-step Markov transition matrix between two sets of sites).
The movement object is a list of lists where each element \code{ix} is the \code{\link{MicroKernel_PowerKernel}}
evaluated at starting location \code{ix}. If the destinations are the same sites as the starting locations
the mosquito is forced to change sites (ie; there are no one-step movements to the same site).
}
