% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Search-MicroKernel.R
\name{MicroKernel_moveMe}
\alias{MicroKernel_moveMe}
\title{MICRO Search Kernels: \code{\link{MicroMosquito}} Movement Function}
\usage{
MicroKernel_moveMe()
}
\description{
Move one mosquito based site and next behavioral state. This method is bound to \code{MicroMosquito$moveMe()}
}
