% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEARCH-MicroKernel-Methods.R
\name{MicroKernel_moveMe_BRO}
\alias{MicroKernel_moveMe_BRO}
\title{MICRO Search Kernels: \code{\link{MicroMosquitoFemale}} Movement Function for Full M-BITES Lifecycle Model}
\usage{
MicroKernel_moveMe_BRO()
}
\description{
Move one mosquito based on site and next behavioral state.
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$moveMe()}
}
}
