% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEARCH-MicroKernel-Methods.R
\name{MicroKernel_moveMe_Male}
\alias{MicroKernel_moveMe_Male}
\title{MICRO Search Kernels: \code{\link{MicroMosquitoMale}} Movement Function for Full M-BITES Lifecycle Model}
\usage{
MicroKernel_moveMe_Male()
}
\description{
Move one mosquito based on site and next behavioral state. This method is bound to \code{MicroMosquitoMale$moveMe()}
}
