% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEARCH-MicroKernel-Generate.R
\name{MicroKernel_PowerKernel}
\alias{MicroKernel_PowerKernel}
\title{MICRO Search Kernels: Calculate Power Kernel}
\usage{
MicroKernel_PowerKernel(S, D, sigma = 3, eps = 0.1, beta = 0)
}
\arguments{
\item{S}{pointer to starting sites (for example, Landscape$get_AquaSites())}

\item{D}{pointer to destination sites}

\item{sigma}{a param}

\item{eps}{a param}

\item{beta}{a param}
}
\value{
matrix
}
\description{
Given pointers \code{S} and \code{D} to two lists of \code{MicroSite} objects,
Calculate one-step Markov transition matrix between sites.
}
