% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEARCH-MicroKernel-Generate.R
\name{MicroKernel_prSort}
\alias{MicroKernel_prSort}
\title{MICRO Search Kernels: Sort ID and Probabilities}
\usage{
MicroKernel_prSort(id, pr)
}
\arguments{
\item{id}{vector of IDs}

\item{pr}{vector of probabilities}
}
\value{
list
\itemize{
\item id: sorted IDs
\item pr: sorted and normalized probabilities
}
}
\description{
Return sorted list of IDs and probabilities
}
