% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Search-MicroKernel.R
\name{MicroKernel_SampleMvOb}
\alias{MicroKernel_SampleMvOb}
\title{MICRO Search Kernels: \code{\link{MicroMosquito}} Sample MvOb}
\usage{
MicroKernel_SampleMvOb(MvOb)
}
\arguments{
\item{MvOb}{output of \code{\link{MicroKernel_exactMvOb}}}
}
\description{
This method is a helper for \code{\link{MicroKernel_moveMe}} and samples the appropriate MvOb in MvAll in the enclosing \code{\link{MicroMosquitoPop}} object.
}
