% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEARCH-MicroKernel-Methods.R
\name{MicroKernel_SampleMvOb_MicroMosquitoPopFemale}
\alias{MicroKernel_SampleMvOb_MicroMosquitoPopFemale}
\title{MICRO Search Kernels: \code{\link{MicroMosquito}} Sample MvOb}
\usage{
MicroKernel_SampleMvOb_MicroMosquitoPopFemale(ixS, state, inPointSet)
}
\arguments{
\item{ixS:}{current site index \code{ix} of this mosquito}

\item{state:}{current behavioral state of this mosquito}

\item{inPointSet:}{current point set this mosquito is in}
}
\value{
new index
}
\description{
This method is a helper for \code{\link{MicroKernel_moveMe}} and samples the appropriate MvOb in MvAll in the enclosing \code{\link{MicroMosquitoPopFemale}} object.
\itemize{
\item This function is bound to \code{MicroMosquitoPopFemale$SampleMove}
}
}
