% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEARCH-MicroKernel-Utility.R
\name{MicroKernelPlot_utility}
\alias{MicroKernelPlot_utility}
\title{MICRO Search Kernels: Plot Kernel Functions}
\usage{
MicroKernelPlot_utility(S, D, N = 50, dMesh = 0.01, dMax = NULL,
  sigma = 3, eps = 0.1, beta = 0)
}
\arguments{
\item{S}{pointer to list of starting sites (for example, Landscape$get_AquaSites())}

\item{D}{pointer to list of destination sites (for example, Landscape$get_FeedingSites())}

\item{N}{number of random starting sites}

\item{dMesh}{granularity of distance bins}

\item{dMax}{maximum distance for distance bins}

\item{sigma}{a param}

\item{eps}{a param}

\item{beta}{a param}
}
\description{
MICRO Search Kernels: Plot Kernel Functions
}
