% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Landscape-Utilities.R
\name{MicroLandscapePlot_utility}
\alias{MicroLandscapePlot_utility}
\title{Plot a MICRO Landscape}
\usage{
MicroLandscapePlot_utility(Landscape, xLim = c(0, 1), yLim = c(0, 1),
  offset = 0.05, bgCol = "grey20", cex = 0.75)
}
\arguments{
\item{Landscape}{\code{\link{Landscape}} object}

\item{xLim}{x-limits of landscape}

\item{yLim}{y-limits of landscape}

\item{offset}{additional space on edges of plot}

\item{bgCol}{background color of plotting surface}

\item{cex}{size of points}
}
\description{
Given object of class \code{\link{Landscape}}, produce a plot.
}
