% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Class.R
\docType{class}
\name{MicroMosquito}
\alias{MicroMosquito}
\title{MICRO Generic Mosquito Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
MicroMosquito
}
\description{
This is a generic Mosquito class definition, it is a superclass for \code{\link{MicroMosquitoFemale}} and \code{\link{MicroMosquitoMale}} and cannot be independently instantiated.
}
\section{Methods}{

\itemize{
\item \strong{Constructor}
\itemize{
\item none: see \code{\link{MicroMosquitoFemale}} and \code{\link{MicroMosquitoMale}} for inheriting classes
}
\item \strong{Getters & Setters}
\itemize{
\item get_id:
}
\item \strong{Pointers}
}
}

\keyword{R6}
\keyword{class}
