% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Class.R
\docType{class}
\name{MicroMosquitoMale}
\alias{MicroMosquitoMale}
\title{MICRO Male Mosquito Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
MicroMosquitoMale
}
\description{
This is a male mosquito class definition for MICRO; it inherits (superclass) from \code{\link{MicroMosquito}}.
}
\section{Fields}{

\itemize{
\item \strong{ID and Time}
\itemize{
\item id: mosquito ID (integer)
\item bDay: time of emergence (numeric)
\item tNow: time of current behavioral state (numeric)
\item tNext: time to next behavioral state change (numeric)
\item genotype: genotype of mosquito (integer)
}
\item \strong{State and Location}
\itemize{
\item state: current behavioral state of mosquito (character)
\itemize{
\item M: Male Mating Bout
\item S: Sugar Feeding Attempt Bout
\item R: Male Resting Bout
}
}
}
}

\section{Methods}{

\itemize{
\item \strong{Constructor}
\itemize{
\item new: initialize a new \code{MicroMosquitoMale} object
\itemize{
\item Arguments:
\itemize{
\item arg1: something.
}
}
}
\item \strong{Getters & Setters}
\itemize{
\item a getter:
}
\item \strong{Pointers}
\itemize{
\item a pointer
}
}
}

\keyword{R6}
\keyword{class}
