% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoPop-Parameters.R
\name{MicroMosquitoPop.Parameters}
\alias{MicroMosquitoPop.Parameters}
\title{Initialize MICRO Mosquito Population Parameters}
\usage{
MicroMosquitoPop.Parameters(module, aquaModule, N_female, N_male = NULL,
  time = 0, ix_female, ix_male = NULL, genotype_female,
  genotype_male = NULL, ...)
}
\arguments{
\item{module}{which M-BITES module to use (must be a character in "BRO","BROM","BROS","BROMS","FULL")}

\item{aquaModule}{which Aquatic Ecology module to use (must be a character in "emerge", "EL4P")}

\item{N_female}{number of female mosquitoes}

\item{N_male}{number of male mosquitoes}

\item{time}{initial time to start populations}

\item{ix_female}{vector of starting location indices for females (indices of \code{\link{AquaticSite}} they emerge from)}

\item{ix_male}{vector of starting location indices for males (indices of \code{\link{AquaticSite}} they emerge from)}

\item{genotype_female}{vector of genotpes for females}

\item{genotype_male}{vector of genotpes for males}

\item{...}{additional named parameters to be passed to the specific M-BITES parameters function
\itemize{
\item MBITES-BRO: pass to \code{\link{MBITES.BRO.Parameters}}
}}

\item{batchSize}{passed to \code{\link{mbitesGeneric.Setup}}}

\item{eggMatT}{passed to \code{\link{mbitesGeneric.Setup}}}
}
\value{
return a list
}
\description{
Generate a list of parameters to initialize \code{\link{MicroMosquitoPopFemale}} and (optionally) \code{\link{MicroMosquitoPopMale}} for initialization of
mosquito populations in a microsimulation tile \code{\link{MicroTile}}. For MICRO simulations, this function automatically interfaces with the necessary M-BITES initialization functions.
}
\examples{
MicroMosquitoPop.Parameters()
}
