% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoPop-Class.R
\docType{class}
\name{MicroMosquitoPop}
\alias{MicroMosquitoPop}
\title{MICRO Generic Mosquito Population Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
MicroMosquitoPop
}
\description{
This is a generic Mosquito Population class definition, it is a superclass for \code{\link{MicroMosquitoPopFemale}} and \code{\link{MicroMosquitoPopMale}} and cannot be independently instantiated.
It contains methods and fields which are generic to both the male and female populations, described below.
}
\section{Fields}{

\itemize{
\item \strong{Fields}
\itemize{
\item pop: list of \code{\link{MicroMosquitoFemale}} or \code{\link{MicroMosquitoMale}} objects
\item MvOb: Movement object WRITE ME
\item MBITES_PAR: M-BITES parameters; see \code{\link{MBITES.Parameters}}
}
\item \strong{Pointers}
\itemize{
\item LandscapePointer: pointer to \code{\link{Landscape}} object in enclosing microsimulation Tile \code{\link{MicroTile}}
\item HumansPointer: pointer to \code{\link{HumanPop}} object in enclosing microsimulation Tile \code{\link{MicroTile}}
\item TilePointer: pointer to enclosing \code{\link{MicroTile}}
}
}
}

\section{Methods}{

\itemize{
\item \strong{Getters & Setters}
\itemize{
\item get_MosquitoIxM: get mosquito corresponding to position \code{IxM} in list (or return all if argument \code{ixM = NULL})
\item get_MosquitoID: get mosquito with private id corresponding to argument \code{ID}
}
\item \strong{Pointers}
\itemize{
\item get_LandscapePointer: get pointer to \code{\link{Landscape}} in same enclosing microsimulation Tile \code{\link{MicroTile}}
\item set_LandscapePointer: set pointer to \code{\link{Landscape}} in same enclosing microsimulation Tile \code{\link{MicroTile}}
\item get_HumansPointer: get pointer to \code{\link{HumanPop}} in same enclosing microsimulation Tile \code{\link{MicroTile}}
\item set_HumansPointer: set pointer to \code{\link{HumanPop}} in same enclosing microsimulation Tile \code{\link{MicroTile}}
\item get_TilePointer: get pointer to enclosing microsimulation Tile \code{\link{MicroTile}}
\item set_TilePointer: get pointer to enclosing microsimulation Tile \code{\link{MicroTile}}
}
}
}

\keyword{R6}
\keyword{class}
