% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoPop-Setup.R
\name{MicroMosquitoPop.Setup}
\alias{MicroMosquitoPop.Setup}
\title{Setup MICRO Mosquito Population & M-BITES Module for a \code{\link{MicroTile}}}
\usage{
MicroMosquitoPop.Setup(cohort = FALSE, module, aquaModule, N_female,
  N_male = NULL, time = 0, ix_female, ix_male = NULL, genotype_female,
  genotype_male = NULL, batchSize = "bms", eggMatT = "off", ...)
}
\arguments{
\item{cohort}{run the chosen M-BITES module in Cohort mode}

\item{module}{which M-BITES module to use (must be a character in "BRO","BROM","BROS","BROMS","FULL")}

\item{aquaModule}{which Aquatic Ecology module to use (must be a character in "emerge", "EL4P")}

\item{N_female}{number of female mosquitoes}

\item{N_male}{number of male mosquitoes}

\item{time}{initial time to start populations (typically 0)}

\item{ix_female}{vector of starting location indices for females (indices of \code{\link{AquaticSite}} they emerge from)}

\item{ix_male}{vector of starting location indices for males (indices of \code{\link{AquaticSite}} they emerge from)}

\item{genotype_female}{vector of genotpes for females}

\item{genotype_male}{vector of genotpes for males}

\item{batchSize}{passed to \code{\link{MBITES.Generic.Setup}}}

\item{eggMatT}{passed to \code{\link{MBITES.Generic.Setup}}}

\item{...}{additional named parameters to be passed to the specific M-BITES parameters function
\itemize{
\item MBITES-BRO: pass to \code{\link{MBITES.BRO.Parameters}}
}}
}
\value{
return a list
}
\description{
Generate a list of parameters to initialize \code{\link{MicroMosquitoPopFemale}} and (optionally) \code{\link{MicroMosquitoPopMale}} for initialization of
mosquito populations in a microsimulation tile \code{\link{MicroTile}}. For MICRO simulations, this function automatically interfaces with the necessary M-BITES setup functions.
}
\section{MBITES-Cohort}{

For Cohort simulations the details of the population can be set to arbitrary values as they will be overwritten by the following functions.
For more details on using MBITES-Cohort to fit EL4P, see \code{\link{set_FemalePop_MicroTile}}.
\itemize{
\item \code{\link{mbitesBRO_cohort_simCohort}}
\item others.
}
}

