% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoPop-Methods.R
\name{MicroMosquitoPopFemale_clear_pop}
\alias{MicroMosquitoPopFemale_clear_pop}
\title{\code{\link{MicroMosquitoPopFemale}} Method: Push Emerging Mosquitoes to Adult Population}
\usage{
MicroMosquitoPopFemale_clear_pop(historyTrack = FALSE)
}
\description{
Manage the pop vector (find dead mosquitoes; if 'con' is provided, write their histories out to JSON).
\itemize{
\item This method is bound to \code{MicroMosquitoPopFemale$clear_pop}
}
}
