% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoPop-Methods.R
\name{MicroMosquitoPopFemale_extend_pop}
\alias{MicroMosquitoPopFemale_extend_pop}
\title{\code{\link{MicroMosquitoPopFemale}} Method: Extend the Population Vector}
\usage{
MicroMosquitoPopFemale_extend_pop()
}
\description{
Used to handle emergence; typically from ImagoQ to the population.
\itemize{
\item This method is bound to \code{MicroMosquitoPopFemale$extend_pop}
}
}
