% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoPop-Methods.R
\name{MicroMosquitoPopFemale_push_pop}
\alias{MicroMosquitoPopFemale_push_pop}
\title{\code{\link{MicroMosquitoPopFemale}} Method: Push Emerging Mosquitoes to Adult Population}
\usage{
MicroMosquitoPopFemale_push_pop(N, tEmerge, ix, genotype, damID, sireID)
}
\description{
Used to handle emergence; typically from ImagoQ to the population.
\itemize{
\item This method is bound to \code{MicroMosquitoPopFemale$push_pop}
}
}
