% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoPop-Methods.R
\name{MicroMosquitoPopFemale_update_nullPop}
\alias{MicroMosquitoPopFemale_update_nullPop}
\title{\code{\link{MicroMosquitoPopFemale}} Method: Update \code{nullPop} Index of \code{NULL} Entries}
\usage{
MicroMosquitoPopFemale_update_nullPop()
}
\description{
Update indices of \code{NULL} mosquitoes used for memory allocation.
\itemize{
\item This method is bound to \code{MicroMosquitoPopFemale$update_nullPop}
}
}
