% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoPop-Methods.R
\name{MicroMosquitoPopFemale_update_pop}
\alias{MicroMosquitoPopFemale_update_pop}
\title{\code{\link{MicroMosquitoPopFemale}} Method: Update Adult Population}
\usage{
MicroMosquitoPopFemale_update_pop(N, time_init = 0, ix_init, genotype_init,
  module)
}
\description{
Mostly used to update the population after fitting EL4P (see \code{\link{EL4P.Mesh.Fit}}) with equilibrium mosquito density and site locations.
\itemize{
\item This method is bound to \code{MicroMosquitoPopFemale$update_pop}
}
}
