% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoPop-Methods.R
\name{MicroMosquitoPopFemale_which_alive}
\alias{MicroMosquitoPopFemale_which_alive}
\title{\code{\link{MicroMosquitoPopFemale}} Method: Find Indices of Alive Mosquitoes}
\usage{
MicroMosquitoPopFemale_which_alive()
}
\description{
Find indices of non-null, non-dead mosquitoes in population.
\itemize{
\item This method is bound to \code{MicroMosquitoPopFemale$which_alive}
}
}
