% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Tile-Class.R
\docType{class}
\name{MicroTile}
\alias{MicroTile}
\title{MICRO Tile Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
MicroTile
}
\description{
This is a generic MICRO microsimulation tile blah blah ...
I talk about something here
\itemize{
\item somewhere 1
\item somewhere 2
}
}
\details{
talk about me in detail!
}
\section{public}{

\itemize{
\item{\code{initialize(N)}}{
talk about me!
}
\item{\code{function()}}{
talk about me!
}
}
}

\section{private}{

\itemize{
\item{\code{function()}}{
talk about me!
}
\item{\code{function()}}{
talk about me!
}
}
}

\section{Active Bindings}{

\itemize{
\item{\code{coolActiveBinding}}{...}
}
}

\keyword{R6}
\keyword{class}
