% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Class.R
\docType{class}
\name{MosquitoFemale}
\alias{MosquitoFemale}
\title{MICRO Female Mosquito Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
MosquitoFemale
}
\description{
This is a female mosquito class definition for MICRO; it inherits (superclass) from \code{\link{MicroMosquito}}.
}
\section{Methods}{

\itemize{
\item \strong{Constructor}
\itemize{
\item new: initialize a new \code{MosquitoFemale} object
\itemize{
\item Arguments:
\itemize{
\item arg1: something.
}
}
}
\item \strong{Getters & Setters}
\itemize{
\item get_id:
}
\item \strong{Pointers}
\itemize{
\item get_FemalePopPointer: get pointer to enclosing \code{\link{MosquitoPopFemale}}
\item set_FemalePopPointer: set pointer to enclosing \code{\link{MosquitoPopFemale}}
\item get_MalePopPointer: get pointer to \code{\link{MosquitoPopMale}} in same enclosing microsimulation Tile \code{\link{MicroTile}}
\item set_MalePopPointer: set pointer to \code{\link{MosquitoPopMale}} in same enclosing microsimulation Tile \code{\link{MicroTile}}
\item get_LandscapePointer: get pointer to \code{\link{Landscape}} in same enclosing microsimulation Tile \code{\link{MicroTile}}
\item set_LandscapePointer: set pointer to \code{\link{Landscape}} in same enclosing microsimulation Tile \code{\link{MicroTile}}
\item get_HumansPointer: get pointer to \code{\link{HumanPop}} in same enclosing microsimulation Tile \code{\link{MicroTile}}
\item set_HumansPointer: set pointer to \code{\link{HumanPop}} in same enclosing microsimulation Tile \code{\link{MicroTile}}
\item get_TilePointer: get pointer to enclosing microsimulation Tile \code{\link{MicroTile}}
\item set_TilePointer: set pointer to enclosing microsimulation Tile \code{\link{MicroTile}}
}
}
}

\keyword{R6}
\keyword{class}
