% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{MosquitoFemaleHistory}
\alias{MosquitoFemaleHistory}
\title{Mosquito Female History Class}
\usage{
MosquitoFemaleHistory()
}
\description{
Class to store female mosquito histories; \code{\link{MicroMosquitoFemale}} and \code{\link{MicroMosquito}}
}
\section{Methods}{

\itemize{
\item historyInit: initialize the mosquito history
\item historyTrack: track history at the end of a generic bout (MBITES-XX-Bouts.R)
\item historyFeed: track history from a host encounter (MBITES-XX-HostEncounter.R)
\item calcBionomics: calculate bionomic parameters upon mosquito death; not intended to be directly called by user.
\item exportHistory: export the mosquito history object as a named list which consists of the following components:
\itemize{
\item stateH: state trajectory
\item timeH: transition times
\item ixH: sites visited
\item pSetH: point sets visited
\item feedAllH: number of blood meals
\item feedAllT: times of blood meals
\item feedHumanH: number of blood meals on human hosts
\item feedHumanT: times of blood meals on human hosts
\item feedIxH: ids of all blood hosts
\item bmSizeH: size of blood meal
\item batchH: size of egg batch
\item bionomics_mBatch: mean egg batch size
\item bionomics_tBatch: total egg production (sum of all egg batches)
\item bionomics_bmInt: all bloodmeal intervals
\item bionomics_bmIntH: human bloodmeal intervals
\item bionomics_lifespan: lifespan
}
}
}

