% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Class.R
\name{MosquitoMaleHistory}
\alias{MosquitoMaleHistory}
\title{\code{\link{MosquitoMale}} History Object}
\usage{
MosquitoMaleHistory(stateH = NULL, timeH = NULL, ixH, pSetH)
}
\arguments{
\item{stateH}{state trajectory}

\item{timeH}{transition times}

\item{ixH}{sites visited (initialized to site of emergence)}

\item{pSetH}{point sets visited (initialized to site of emergence)}
}
\value{
list
}
\description{
Generate the history object for \code{\link{MosquitoMale}}
}
\examples{
MosquitoMaleHistory(ixH = 0L, pSetH = "l")
}
