% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{mosquitoPfSI}
\alias{mosquitoPfSI}
\title{PfSI Mosquito Stage Pathogen Class}
\usage{
mosquitoPfSI(PfID_init, tInf_init = -1, damID_init = -1, sireID_init = -1,
  infected_init = FALSE)
}
\arguments{
\item{PfID_init}{PfID derived from the infecting human's gametocytes}

\item{tInf_init}{time of current infection (successful human to mosquito transmission event)}

\item{damID_init}{female gametocyte 'mother'}

\item{sireID_init}{male gametocyte 'father'}

\item{infected_init}{infection status}
}
\description{
A class for storing parameters and state variables of mosquito-stage PfSI infections.
}
\section{Methods}{

\itemize{
\item get_PfID:
\item set_PfID:
\item get_tInf:
\item set_tInf:
\item get_damID:
\item set_damID:
\item get_sireID:
\item set_sireID:
\item get_infected:
\item set_infected:
}
}

\section{Fields}{

\itemize{
\item a field: talk about me
}
}

