% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-RM-MosyPop-Class.R
\docType{class}
\name{MosyPop}
\alias{MosyPop}
\title{MACRO Mosquito Population Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
MosyPop
}
\description{
This is a generic MACRO mosquito population blah blah ...
below i describe the basic structure of the patch. methods and fields for specific COMPONENTS can be found in:
\itemize{
\item somewhere 1
\item somewhere 2
}
}
\details{
talk about me in detail!
}
\section{public}{

\itemize{
\item{\code{initializer(N)}}{
talk about me!
}
\item{\code{function()}}{
talk about me!
}
}
}

\section{Private}{

\itemize{
\item{maxEIP: maximum value of extrinsic incubation period}
\item{nPatches: number of patches in \code{\link{Tile}}}
\item{RMparameters}{
\itemize{
\item{p: the probability of surviving one day}
\item{f: the proportion of mosquitoes that blood feed each day}
\item{Q: the proportion of mosquitoes that feed on humans}
\item{v: ?}
}
}
\item{MP}{
\itemize{
\item{M: the number of mosquitoes}
\item{Y: the number of infected mosquitoes (including Z)}
\item{Z: the number of infectious mosquitoes}
\item{ZZ: matrix where each row is the number that will be added to the infectious state on that day}
}
}
\item{P: p^EIP}
\item{Psi: psi^EIP}
}
}

\section{Active Bindings}{

\itemize{
\item{\code{coolActiveBinding}}{...}
}
}

\keyword{R6}
\keyword{class}
