% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUATIC-ECOLOGY.R
\name{newAdults}
\alias{newAdults}
\title{Empty Adult Slot for ImagoQ}
\usage{
newAdults(N = 0, tm = 0, ix = 0, dam = 0, sire = 0)
}
\arguments{
\item{N}{number of emerging adults from this element of queue}

\item{tm}{time of emergence}

\item{ix}{index of aquatic habitat}

\item{dam}{id of mother}

\item{sire}{id of father}
}
\value{
a list of named elements
}
\description{
This function is a low-level utility to generate a null placeholder value for adults in ImagoQ. It is called by allocImagoQ and extendImagoQ.
}
\examples{
newAdults(N=0,tm=0,ix=0,dam=0,sire=0)
}
