% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Structures.R
\name{newEgg}
\alias{newEgg}
\title{Empty Egg Batch Slot for EggQ}
\usage{
newEgg(N = 0L, tOviposit = 0L, damID = 0L, sireID = 0L,
  genotype = NULL)
}
\arguments{
\item{N}{number of eggs in this batch}

\item{tOviposit}{time of oviposition}

\item{damID}{ID of mother}

\item{sireID}{ID of father}

\item{genotype}{placeholder}
}
\value{
a named list
}
\description{
This function is a low-level utility to generate a null placeholder value for an egg batch in EggQ.
It is called by \code{\link{allocEggQ}} and \code{\link{extendEggQ}}.
}
\section{Structure}{

\itemize{
\item N number of emerging adults
\item tOviposit time of oviposition
\item damID ID of mother
\item sireID ID of father
\item genotype integer genotype identifier
}
}

\examples{
newEgg()
}
