% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Site-AquaticEcology.R
\name{newImago}
\alias{newImago}
\title{Empty Adult Slot for ImagoQ}
\usage{
newImago(N = 0L, tm = 0, ix = 0L, dam = 0L, sire = 0L)
}
\arguments{
\item{N}{number of emerging adults from this element of queue}

\item{tm}{time of emergence}

\item{ix}{index of aquatic habitat}

\item{dam}{id of mother}

\item{sire}{id of father}
}
\value{
a list of named elements
}
\description{
This function is a low-level utility to generate a null placeholder value for adults in ImagoQ.
It is called by \code{\link{allocImagoQ}} and \code{\link{extendImagoQ}}.
}
\examples{
newImago(N=0,tm=0,ix=0,dam=0,sire=0)
}
