% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Macro-Emerge.R
\name{oneDay_emerge_MACRO}
\alias{oneDay_emerge_MACRO}
\title{Get \code{MacroPatch} Seasonal Emergence}
\usage{
oneDay_emerge_MACRO(tNow)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
some_function()
}
