% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Calibration.R
\name{oneDay_GEL4P}
\alias{oneDay_GEL4P}
\title{Run a Single Daily Time-step of Aquatic Population Dynamics}
\usage{
oneDay_GEL4P(ix, psi, pop, PAR)
}
\arguments{
\item{ix}{index of this aquatic population}

\item{psi}{value of psi}

\item{pop}{EL4P aquatic population object (see \code{link\{EL4P})}

\item{PAR}{aquatic parameters passed from \code{\link{makePAR_EL4P}}}
}
\value{
the aquatic population state variables after one day
}
\description{
Run a single aquatic population's daily difference equation dynamics for a single time step. Eggs are deposited assuming adult population at equilibrium
\eqn{M[ix]*aquaEq[ix]*\left ( \frac{G}{lifespan} \right )}, see \code{link{makePAR_EL4P}} for these parameters.
}
