% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Macro-Emerge.R
\name{oneDay_MacroEmerge}
\alias{oneDay_MacroEmerge}
\title{Get \code{MacroPatch} Seasonal Emergence}
\usage{
oneDay_MacroEmerge()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Queue the ImagoQ
}
\examples{
some_function()
}
