% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-Emerge-Methods.R
\name{oneDay_MicroEmerge}
\alias{oneDay_MicroEmerge}
\title{MICRO \code{\link{Landscape}} Method: Emerge One Day Dynamics}
\usage{
oneDay_MicroEmerge()
}
\description{
Calculate emerging adults for a single aquatic habitat and add them to that site's ImagoQ for all sites.
This method is bound to \code{Landscape$oneDay_Emerge()}.
}
