% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-Emerge-Methods.R
\name{oneDay_MicroEmergeSite}
\alias{oneDay_MicroEmergeSite}
\title{MICRO \code{\link{AquaticSite}} Method: Emerge One Day Dynamics}
\usage{
oneDay_MicroEmergeSite(tNow)
}
\arguments{
\item{tNow}{integer time to calculate emergence}
}
\description{
Calculate emerging adults for a single aquatic habitat and add them to that site's ImagoQ.
This method is bound to \code{AquaticSite$oneDay_EmergeSite()}.
}
