% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-MosquitoPop-Methods.R
\name{oneDay_RM}
\alias{oneDay_RM}
\title{MACRO \code{MacroMosquitoPop} Method: oneDay_RM}
\usage{
oneDay_RM()
}
\description{
Daily Ross-Macdonald difference/diffusion equations for mosquito dynamics.
This method is bound to \code{MacroMosquitoPop$oneDay_RM}
These equations are likely inefficient; see optimizing BLAS/LAPACK and copying less matricies for a faster down the line implementation.
}
