% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-EL4P-Methods.R
\name{oneStep_MicroEL4P}
\alias{oneStep_MicroEL4P}
\title{MICRO \code{\link{Landscape}} Method: EL4P One Day Dynamics}
\usage{
oneStep_MicroEL4P()
}
\description{
Run daily EL4P module aquatic dynamics for all sites. See \code{\link{oneStep_MicroEL4PSite}} for details.
This function fills a generic for \code{\link{simMICRO_oneStep}}.
\itemize{
\item This method is bound to \code{Landscape$oneStep_AquaticEcology()}.
}
}
