% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-EL4P-Methods.R
\name{oneStep_MicroEL4PSite}
\alias{oneStep_MicroEL4PSite}
\title{MICRO \code{\link{AquaticSite}} Method: EL4P One Day Dynamics}
\usage{
oneStep_MicroEL4PSite(tNow)
}
\arguments{
\item{tNow}{current global time of \code{\link{MicroTile}}}
}
\description{
Collect eggs from the \code{\link{EggQ}} in this site and move \code{\link{EL4P}}, and run the one step difference equation dynamics for aquatic development,
finally, move emerging adults from the pool into the \code{\link{ImagoQ}} for all genotypes.
\itemize{
\item This method is bound to \code{AquaticSite$oneStep_EL4PSite()}.
}
}
