% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-Emerge-Methods.R
\name{oneStep_MicroEmerge}
\alias{oneStep_MicroEmerge}
\title{MICRO \code{\link{Landscape}} Method: Emerge One Day Dynamics}
\usage{
oneStep_MicroEmerge()
}
\description{
Calculate emerging adults for a single aquatic habitat and add them to that site's ImagoQ for all sites.
This function fills a generic for \code{\link{simMICRO_oneStep}}.
\itemize{
\item This method is bound to \code{Landscape$oneStep_AquaticEcology()}.
}
}
