% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-Emerge-Methods.R
\name{oneStep_MicroEmergeSite}
\alias{oneStep_MicroEmergeSite}
\title{MICRO \code{\link{AquaticSite}} Method: Emerge One Day Dynamics}
\usage{
oneStep_MicroEmergeSite(tNow)
}
\arguments{
\item{tNow}{current global time of \code{\link{MicroTile}}}
}
\description{
Calculate emerging adults for a single aquatic habitat and add them to that site's ImagoQ.
This method is bound to \code{AquaticSite$oneStep_EmergeSite()}.
}
