% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Utility.R
\name{ovipositionEq_oneHistory_utility}
\alias{ovipositionEq_oneHistory_utility}
\title{M-BITES Utility: Calculate Counts of Oviposition at \code{\link{AquaticSite}}}
\usage{
ovipositionEq_oneHistory_utility(oneHistory, aquaVec)
}
\arguments{
\item{oneHistory}{single mosquito history}

\item{aquaVec}{vector of length equal to number of aquatic sites on \code{\link{Landscape}}}
}
\description{
Return vector of counts recording number of ovipositions at each site on landscape. Called by \code{\link{ovipositionEq_utility}}
}
