% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Utility.R
\name{ovipositionEq_utility}
\alias{ovipositionEq_utility}
\title{M-BITES Utility: Calculate Empirical Probability of Oviposition at \code{\link{AquaticSite}}}
\usage{
ovipositionEq_utility(history, nAqua)
}
\arguments{
\item{history}{list of mosquito histories}

\item{nAqua}{number of aquatic habitats}
}
\description{
Return vector of scaled frequencies of oviposition at aquatic habitats.
}
