% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-PATCH-R6.R
\docType{class}
\name{Patch}
\alias{Patch}
\title{MACRO Patch Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
Patch
}
\description{
This is a generic MACRO patch blah blah ...
below i describe the basic structure of the patch. methods and fields for specific COMPONENTS can be found in:
\itemize{
\item somewhere 1
\item somewhere 2
}
}
\details{
talk about me in detail!
}
\section{public}{

\itemize{
\item{\code{initialize(N)}}{
talk about me!
}
\item{\code{function()}}{
talk about me!
}
}
}

\section{private}{

\itemize{
\item{\code{function()}}{
talk about me!
}
\item{\code{function()}}{
talk about me!
}
}
}

\section{Active Bindings}{

\itemize{
\item{\code{coolActiveBinding}}{...}
}
}

\keyword{R6}
\keyword{class}
