% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-Macro-PfClass.R
\docType{class}
\name{PatchPf}
\alias{PatchPf}
\title{MACRO Pf Pathogen Object Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
PatchPf
}
\description{
This is a generic Pf Pathogen object for a patch blah blah ...
I talk about something here
\itemize{
\item somewhere 1
\item somewhere 2
}
}
\details{
talk about me in detail!
}
\section{public}{

\itemize{
\item{\code{initialize(N)}}{
talk about me!
}
\item{\code{function()}}{
talk about me!
}
}
}

\section{private}{

\itemize{
\item{\code{function()}}{
talk about me!
}
\item{\code{function()}}{
talk about me!
}
}
}

\section{Active Bindings}{

\itemize{
\item{\code{coolActiveBinding}}{...}
}
}

\keyword{R6}
\keyword{class}
