% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PFSI-PfSI.R
\name{pathOBJ_PfSI}
\alias{pathOBJ_PfSI}
\title{Pathogen Object for Humans in PfSI Module}
\usage{
pathOBJ_PfSI(b = Pf_b, c = Pf_c)
}
\arguments{
\item{b}{transmission efficiency: infected mosquito to human}

\item{c}{transmission efficiency: infected human to mosquito}
}
\value{
a named list
\itemize{
\item infected: TRUE or FALSE
\item chemoprophylaxis: is human protected by chemoprophylaxis?
}
}
\description{
This function generates the Pf object for humans Pathogen slot \code{HUMANS[[ixH]]$Pathogens$Pf <<- pathOBJ_PfSI()}
It is called during initialization by \code{PFSI.INIT()}.
}
