% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{pevaccinatePfMOI}
\alias{pevaccinatePfMOI}
\title{PfMOI \code{Human} Event: PfMOI PE Vaccination Event}
\usage{
pevaccinatePfMOI(tEvent, PAR)
}
\arguments{
\item{tEvent}{begin PE protection}

\item{PAR}{\code{NULL}}
}
\description{
Begin PfMOI PE vaccination protection.
This method is bound to \code{Human$endprophylaxisPfMOI()}
\itemize{
\item protection: infected mosquito to human transmission efficiency is modified by \code{peBlockPf}, see \code{\link{PfMOI.Parameters}}
\item waning efficacy: queue \code{\link{add2Q_pewanePfMOI}}
}
}
