% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{pevaccinatePfSI}
\alias{pevaccinatePfSI}
\title{PfSI \code{Human} Event: PfSI PE Vaccination Event}
\usage{
pevaccinatePfSI(tEvent, PAR)
}
\arguments{
\item{tEvent}{begin PE protection}

\item{PAR}{\code{NULL}}
}
\description{
Begin PfSI PE vaccination protection.
This method is bound to \code{Human$endprophylaxisPfSI()}
\itemize{
\item protection: infected mosquito to human transmission efficiency is modified by \code{peBlockPf}, see \code{\link{PfSI.Parameters}}
\item waning efficacy: queue \code{\link{add2Q_pewanePfSI}}
}
}
