% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{pewanePfMOI}
\alias{pewanePfMOI}
\title{PfMOI \code{Human} Event: PfMOI PE Waning Protection Event}
\usage{
pewanePfMOI(tEvent, PAR)
}
\arguments{
\item{tEvent}{end PE protection}

\item{PAR}{\code{NULL}}
}
\description{
End PfMOI PE protection.
This method is bound to \code{Human$pewanePfMOI()}
\itemize{
\item protection: infected mosquito to human transmission efficiency is set back to \code{Pf_b}, see \code{\link{PfMOI.Parameters}}
}
}
