% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{pewanePfSI}
\alias{pewanePfSI}
\title{PfSI \code{Human} Event: PfSI PE Waning Protection Event}
\usage{
pewanePfSI(tEvent, PAR)
}
\arguments{
\item{tEvent}{end PE protection}

\item{PAR}{\code{NULL}}
}
\description{
End PfSI PE protection.
This method is bound to \code{Human$pewanePfSI()}
\itemize{
\item protection: infected mosquito to human transmission efficiency is set back to \code{Pf_b}, see \code{\link{PfSI.Parameters}}
}
}
