% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Setup.R
\name{PfMOI.Setup}
\alias{PfMOI.Setup}
\title{Initialize PfMOI MODULE}
\usage{
PfMOI.Setup(overwrite = TRUE, interaction = FALSE, InteractionPf = 0,
  MosyMaxI = 1L, HumanMaxI = 1L, Pf_c = 0.15, Pf_b = 0.55,
  DurationPf = 200, LatentPf = 10, FeverPf = 0.3, mnFeverPf = 10,
  vrFeverPf = 0.1, TreatPf = 0.5, mnTreatPf = 3,
  mnChemoprophylaxisPf = 32, PEProtectPf = 0.99, peBlockPf = 1,
  mnPEPf = 270, vrPEPf = 50, GSProtectPf = 1, gsBlockPf = 0.9,
  mnGSPf = 180, vrGSPf = 20, rdtSensPf = 0.9, rdtSpecPf = 0.1,
  lmSensPf = 0.9, lmSpecPf = 0.1)
}
\arguments{
\item{overwrite}{overwrite existing methods and fields?}

\item{interaction}{independent clearance or pathogen interaction model for calculating infection clearance waiting times (if \code{FALSE} use \code{\link{PfMOI_ttClearPf_Independent}}, if \code{TRUE} use \code{\link{PfMOI_ttClearPf_Interaction}})}

\item{InteractionPf}{if using pathogen interaction model of clearance this parameter control strength of interaction
\itemize{
\item 0: no interaction (reduces to independent clearance)
\item <0: facilitation against host immune system (slower clearance rate than independent queueing)
\item 0<: competition for host resources (faster clearance rate than independent queueing)
\item See \url{https://doi.org/10.1186/1475-2875-8-87} for details of queueing model with pathogen interaction.
}}

\item{MosyMaxI}{maximum number of clonal variants passed in single mosquito to human transmission event (set to \code{-1L} for unlimited)}

\item{HumanMaxI}{maximum number of clonal variants passed in single human to mosquito transmission event (set to \code{-1L} for unlimited)}

\item{Pf_c}{infected human to mosquito transmission efficiency}

\item{Pf_b}{infected mosquito to human transmission efficiency}

\item{DurationPf}{duration of infection (assuming clonal variants are cleared independently)}

\item{LatentPf}{latency (How many days after the infectious bite does the infection start?)}

\item{FeverPf}{probability of fever}

\item{mnFeverPf}{mean of timing of fever incident relative to start of PfMOI infection (log-normally distributed, mean on natural scale)}

\item{vrFeverPf}{standard deviation of timing of fever incident relative to start of PfMOI infection (log-normally distributed, standard deviation on natural scale)}

\item{TreatPf}{probability of treatment after fever incident}

\item{mnTreatPf}{average waiting time from fever to treatment (exponentially distributed)}

\item{mnChemoprophylaxisPf}{constant period of proteection from chemoprophylaxis}

\item{PEProtectPf}{proportion protected by PE vaccination (probability vaccination successful)}

\item{peBlockPf}{proportion of infections blocked by PE vaccination}

\item{mnPEPf}{mean duration of protection from PE vaccination}

\item{vrPEPf}{standard deviation of protection from PE vaccination}

\item{GSProtectPf}{proportion protected by GS vaccination (probability vaccination successful)}

\item{gsBlockPf}{proportion of infections blocked by GS vaccination}

\item{mnGSPf}{mean duration of protection from GS vaccination}

\item{vrGSPf}{standard deviation of protection from GS vaccination}

\item{rdtSensPf}{RDT sensitivity}

\item{rdtSpecPf}{RDT specificity}

\item{lmSensPf}{Light Microscopy sensitivity}

\item{lmSpecPf}{Light Microscopy specificity}
}
\value{
Defines a field (list) PfSI_PAR in \code{\link{HumanPop}} and public methods in \code{\link{Human}}
}
\description{
Generate a list of parameters PfMOI_PAR in \code{\link{Human}} and public methods in \code{\link{Human}} for PfSI infection model; also defines public methods
in \code{\link{MicroMosquitoFemale}} for PfMOI infection model.
}
