% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{PfMOI_increment_PfID}
\alias{PfMOI_increment_PfID}
\title{PfMOI \code{HumanPop} Method: Increment PfID}
\usage{
PfMOI_increment_PfID()
}
\description{
Whenever a new liver-stage infection manifests in a human we increment the PfID counter and return the new PfID.
This method is bound to \code{HumanPop$increment_PfID()}
}
