% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{PfMOI_ttClearPf_Independent}
\alias{PfMOI_ttClearPf_Independent}
\title{PfMOI \code{Human} Method: Duration of Infection}
\usage{
PfMOI_ttClearPf_Independent()
}
\description{
How many days does a single PfMOI infection last? See \code{\link{PfMOI.Parameters}} parameter \code{DurationPf} for baseline clearance rate.
This function  calculates clearance rate under the assumption of independent clearance (M/M/1 queueing process for infection).
This method is called from \code{\link{}}
This method is bound to \code{Human$ttClearPf()}
}
