% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{PfMOI_ttClearPf_Interaction}
\alias{PfMOI_ttClearPf_Interaction}
\title{PfMOI \code{Human} Method: Duration of Infection with Pathogen Interaction}
\usage{
PfMOI_ttClearPf_Interaction()
}
\description{
How many days does a single PfMOI infection last? See \code{\link{PfMOI.Parameters}} parameter \code{DurationPf} for baseline clearance rate.
This function calculates clearance rate as a function of individual MOI modeling pathogen competition for host resources (faster clearance than expected under independent clearance)
or pathogen facilitation (slower clearance than expected under independent clearance).
This method is called from \code{\link{}}
This method is bound to \code{Human$ttClearPf()}
}
\details{
* See \url{https://doi.org/10.1186/1475-2875-8-87} for details of queueing model with pathogen interaction.
}
