% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{PfMOI_ttGSWanePf}
\alias{PfMOI_ttGSWanePf}
\title{PfMOI \code{Human} Method: Duration of protection by GS Vaccination}
\usage{
PfMOI_ttGSWanePf()
}
\description{
Duration of protection by GS Vaccination? See \code{\link{PfMOI.Parameters}} parameter \code{mnGSPf} and \code{vrGSPf} for normally distributed duration of protection.
This method is called from \code{\link{event_gswanePfMOI}}
This method is bound to \code{Human$ttGSWanePf()}
}
