% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{PfMOI_ttInfectionPf}
\alias{PfMOI_ttInfectionPf}
\title{PfMOI \code{Human} Method: Duration of Latency}
\usage{
PfMOI_ttInfectionPf()
}
\description{
How many days after the infectious bite does a PfMOI infection start? See \code{\link{PfMOI.Parameters}} parameter \code{LatentPf} constant latent period.
This method is called from \code{\link{}}
This method is bound to \code{Human$ttInfectPf()}
}
