% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{PfMOI_ttSusceptiblePf}
\alias{PfMOI_ttSusceptiblePf}
\title{PfMOI \code{Human} Method: Duration of Protection from Chemoprophylaxis}
\usage{
PfMOI_ttSusceptiblePf()
}
\description{
After administration of Chemoprophylaxis what is time to susceptibility? See \code{\link{PfMOI.Parameters}} parameter \code{mnChemoprophylaxisPf} constant timing period.
This method is called from \code{\link{event_endprophylaxisPfMOI}}
This method is bound to \code{Human$ttSusceptiblePf()}
}
