% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Auxiliary.R
\name{PfSI.Auxiliary.Setup}
\alias{PfSI.Auxiliary.Setup}
\title{Initialize Auxiliary PfSI Module Methods (Pathogen)}
\usage{
PfSI.Auxiliary.Setup()
}
\arguments{
\item{Pf_c}{0.15; transmission efficiency: infected human to mosquito}
}
\value{
Defines a field (list) PfSI_PAR in \code{\link{HumanPop}} and public methods in \code{\link{Human}}
}
\description{
\code{PfSI.Auxiliary} is called internally by \code{\link{PfSI.Setup}} to initialize functions auxiliary to the main PfSI module definition.
}
\examples{

}
