% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PFSI-SetupPfSI.R
\name{PFSI.INIT}
\alias{PFSI.INIT}
\title{Initialize PfSI Human Infections}
\usage{
PFSI.INIT(PfPR)
}
\arguments{
\item{PfPR}{initial parasite prevalence in the human population}
}
\value{
modify \code{HUMANS} object in global environment
}
\description{
Initialize PfSI module infections in the human population (requires \code{HUMANS} to be defined in the global environment)
\code{PFSI.SETUP()} should be called prior to initializing human infections.
}
\examples{
PFSI.INIT()
}
