% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{PfSI_increment_PfID}
\alias{PfSI_increment_PfID}
\title{PfSI \code{HumanPop} Method: Increment PfID}
\usage{
PfSI_increment_PfID()
}
\description{
Whenever a new liver-stage infection manifests in a human we increment the PfID counter and return the new PfID.
This method is bound to \code{HumanPop$increment_PfID()}
}
