% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{PfSI_ttClearPf}
\alias{PfSI_ttClearPf}
\title{PfSI \code{Human} Method: Duration of Infection}
\usage{
PfSI_ttClearPf()
}
\description{
How many days does a PfSI infection last? See \code{\link{PfSI.Parameters}} parameter \code{DurationPf} for mean of exponentially distributed infection.
This method is called from \code{\link{event_endPfSI}}
This method is bound to \code{Human$ttClearPf()}
}
