% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{PfSI_ttGSWanePf}
\alias{PfSI_ttGSWanePf}
\title{PfSI \code{Human} Method: Duration of protection by GS Vaccination}
\usage{
PfSI_ttGSWanePf()
}
\description{
Duration of protection by GS Vaccination? See \code{\link{PfSI.Parameters}} parameter \code{mnGSPf} and \code{vrGSPf} for normally distributed duration of protection.
This method is called from \code{\link{event_gswanePfSI}}
This method is bound to \code{Human$ttGSWanePf()}
}
