% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{PfSI_ttInfectionPf}
\alias{PfSI_ttInfectionPf}
\title{PfSI \code{Human} Method: Duration of Latency}
\usage{
PfSI_ttInfectionPf()
}
\description{
How many days after the infectious bite does a PfSI infection start? See \code{\link{PfSI.Parameters}} parameter \code{LatentPf} constant latent period.
This method is called from \code{\link{infectiousBite_PfSI}}
This method is bound to \code{Human$ttInfectionPf()}
}
