% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{PfSI_ttPEWanePf}
\alias{PfSI_ttPEWanePf}
\title{PfSI \code{Human} Method: Duration of protection by PE Vaccination}
\usage{
PfSI_ttPEWanePf()
}
\description{
After administration of PE Vaccination what is duration of protection (reduced infected mosquito to human transmission efficiency)? See \code{\link{PfSI.Parameters}} parameter \code{mnPEPf} and \code{vrPEPf} for normally distributed duration of protection.
This method is called from \code{\link{event_pewanePfSI}}
This method is bound to \code{Human$ttPEWanePf()}
}
