% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{PfSI_ttSusceptiblePf}
\alias{PfSI_ttSusceptiblePf}
\title{PfSI \code{Human} Method: Duration of Protection from Chemoprophylaxis}
\usage{
PfSI_ttSusceptiblePf()
}
\description{
After administration of Chemoprophylaxis what is time to susceptibility? See \code{\link{PfSI.Parameters}} parameter \code{mnChemoprophylaxisPf} constant timing period.
This method is called from \code{\link{event_endprophylaxisPfSI}}
This method is bound to \code{Human$ttSusceptiblePf()}
}
