% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{PfSI_ttTreatPf}
\alias{PfSI_ttTreatPf}
\title{PfSI \code{Human} Method: Timing of Treatment}
\usage{
PfSI_ttTreatPf()
}
\description{
What is the timing of treatment relative to the start of a fever incident? See \code{\link{PfSI.Parameters}} parameter \code{mnTreatPf} constant timing period.
This method is called from \code{\link{event_treatPfSI}}
This method is bound to \code{Human$ttTreatPf()}
}
