% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfPedigree.R
\name{PfTransmission_init}
\alias{PfTransmission_init}
\title{Initialize PfTransmission to Track Pathogen Transmission Networks}
\usage{
PfTransmission_init(directory, fileName)
}
\arguments{
\item{directory}{directory that will be created; files will be put in directory/OUTPUT/..}

\item{fileName}{name of the file to write to; directory/OUTPUT/fileName.csv}
}
\value{
a text connection
}
\description{
Generate a text connection to write .csv transmission chains. This connection \strong{must} be named \code{PfTransmissionCon}.
See \code{\link{trackPfTransmission}}.
}
\examples{
PfTransmission_init(directory, fileName)
}
