% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Utilities.R
\name{plot_PfSIOneTrajectory}
\alias{plot_PfSIOneTrajectory}
\title{plot_PfSI_oneTrajectory}
\usage{
plot_PfSIOneTrajectory(ixH, oneHistory, tMax)
}
\arguments{
\item{a}{parameter}
}
\value{
do something
}
\description{
write me
}
\examples{
plot_PfSI_oneTrajectory()
}
