% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILITIES-Mosquito.R
\name{plotAdults}
\alias{plotAdults}
\title{Plot Imported Adult Dynamics .csv Data}
\usage{
plotAdults(adults)
}
\arguments{
\item{egg}{a data frame (output of \code{\link{importAdults}})}
}
\value{
plot
}
\description{
This is a utility to plot logged adult dynamics data after it has been imported by \code{\link{importAdults}}.
}
\examples{
plotAdults(adults)
}
