% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILITIES-Human.R
\name{plotPfsiOneTrajectory}
\alias{plotPfsiOneTrajectory}
\title{Plot a Single Human Trajectory for PfSI Module}
\usage{
plotPfsiOneTrajectory(ixH, oneHistory, tMax)
}
\arguments{
\item{ixH}{index of human}

\item{oneHistory}{a single human histoy \code{history[[ixH]]}}

\item{tMax}{maximum time event observed in all histories}
}
\value{
plot
}
\description{
Plots a single trajectory through state space for PfSI module.
}
\examples{
plotPfsiOneTrajectory
}
