% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILITIES-Human.R
\name{plotPfsiTrajectory}
\alias{plotPfsiTrajectory}
\title{Plot Human Trajectories from PfSI Module Histories}
\usage{
plotPfsiTrajectory(history)
}
\arguments{
\item{history}{imported human histories from \code{\link{importHumanEvent_PfSI}}.}
}
\value{
plot
}
\description{
Plots all human trajectories through state space for PfSI module.
}
\examples{
plotPfsiOneTrajectory
}
