% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Landscape-Parameters.R
\name{pointsClustered}
\alias{pointsClustered}
\title{Generate Matern Clustering Point Pattern for Landscape Sites}
\usage{
pointsClustered(n, meanParents = 10, clusteredness = 0.25, xLim = c(0, 1),
  yLim = c(0, 1))
}
\arguments{
\item{n}{number of points to generate}

\item{meanParents}{intensity of Poisson process for cluster centers}

\item{clusteredness}{control mean scatter of child points around cluster centers}

\item{xLim}{a length 2 numeric vector of bounds for the sampling grid}

\item{yLim}{a length 2 numeric vector of bounds for the sampling grid}
}
\value{
a list with two elements x and y corresponding to the sampled points
}
\description{
This function is a low-level utility to generate spatial point patterns for the MICRO-LANDSCAPE point sets.
}
\examples{
pointsClustered(n=10, meanParents = 10, clusteredness = .25, xLim=c(0,1), yLim=c(0,1))
}
