% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Landscape-Parameters.R
\name{pointsLattice}
\alias{pointsLattice}
\title{Generate Regular Grid (Lattice) Point Pattern for Landscape Sites}
\usage{
pointsLattice(n, xLim = c(0, 1), yLim = c(0, 1))
}
\arguments{
\item{n}{number of points to generate}

\item{xLim}{a length 2 numeric vector of bounds for the sampling grid}

\item{yLim}{a length 2 numeric vector of bounds for the sampling grid}
}
\value{
a list with two elements x and y corresponding to the sampled points
}
\description{
This function is a low-level utility to generate spatial point patterns for the MICRO-LANDSCAPE point sets.
}
\examples{
pointsLattice(n=10, xLim=c(0,1), yLim=c(0,1))
}
