% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Landscape-Parameters.R
\name{pointsOverdispersed}
\alias{pointsOverdispersed}
\title{Generate Overdispersed (SSI) Point Pattern for Landscape Sites}
\usage{
pointsOverdispersed(n, inhibitionFactor = 1, xLim = c(0, 1), yLim = c(0,
  1))
}
\arguments{
\item{n}{number of points to generate}

\item{inhibitionFactor}{controls level of overdispersion (higher values correspond to a more overdispersed spatial point process)}

\item{xLim}{a length 2 numeric vector of bounds for the sampling grid}

\item{yLim}{a length 2 numeric vector of bounds for the sampling grid}
}
\value{
a list with two elements x and y corresponding to the sampled points
}
\description{
This function is a low-level utility to generate spatial point patterns for the MICRO-LANDSCAPE point sets.
}
\examples{
pointsOverdispersed(n=10, inhibitionFactor = 1, xLim=c(0,1), yLim=c(0,1))
}
