% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{probeHost_PfMOI}
\alias{probeHost_PfMOI}
\title{PfMOI \code{\link{Human}} Method: Host Probing}
\usage{
probeHost_PfMOI(tBite, EIP, mosquitoPfMOI)
}
\arguments{
\item{tBite}{time of bite}

\item{EIP}{length of entomological incubation period}

\item{mosquitoPfMOI}{\code{\link{mosquitoPfMOI}} object passed from mosquito to human}
}
\description{
This method is called by a mosquito when she probes a human host, but may also be called by \code{\link{SimBitePfSI}} as a filler.
If the biting mosquito is infectious, the method calls \code{\link{infectiousBite_PfMOI}}, otherwise does nothing.
\itemize{
\item This method is bound to \code{Human$probeHost_PfMOI()}
}
}
