% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{probeHost_PfSI}
\alias{probeHost_PfSI}
\title{PfSI \code{Human} Method: Host Probing}
\usage{
probeHost_PfSI(tBite, mosquitoPfSI)
}
\arguments{
\item{tBite}{time of bite}

\item{mosquitoPfSI}{\code{\link{mosquitoPfSI}} object passed from mosquito to human}
}
\description{
This method is called by a mosquito when she probes a human host, but may also be called by \code{\link{SimBitePfSI}} as a filler.
If the biting mosquito is infectious, the method calls \code{\link{infectiousBite_PfSI}}, otherwise does nothing.
This method is bound to \code{Human$probeHost_PfSI()}
}
