% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{probing_PfMOI}
\alias{probing_PfMOI}
\title{PfMOI \code{\link{MicroMosquitoFemale}} Method: Host Probing}
\usage{
probing_PfMOI()
}
\description{
The mosquito probes the host prior to successful feeding. Probing occurs during code{MicroMosquitoFemale$humanEncounter()}.
Mosquito to human pathogen transmission occurs during host probing.
If the PfMOI infection(s) has passed the EIP, \code{\link{probeHost_PfMOI}} is called via the mosquito's pointer to the \code{\link{HumanPop}} class
to initiate the PfMOI infection process.
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$probing()}
}
}
