% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{probing_PfSI}
\alias{probing_PfSI}
\title{PfSI \code{\link{MicroMosquitoFemale}} Method: Host Probing}
\usage{
probing_PfSI()
}
\description{
The mosquito probes the host prior to successful feeding. Probing occurs during code{MicroMosquitoFemale$humanEncounter()}.
Mosquito to human pathogen transmission occurs during host probing.
If the PfSI infection has passed the EIP, \code{\link{probeHost_PfSI}} is called via the mosquito's pointer to the \code{\link{HumanPop}} class
to initiate the PfSI infection process.
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$probing()}
}
}
