% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-EL4P-Calibrate.R
\name{psi2K_cf}
\alias{psi2K_cf}
\title{Regress K on Psi and Return Coefficients}
\usage{
psi2K_cf()
}
\value{
regression coefficients
}
\description{
Run a linear regression of K on iverse of psi and extract coefficients; optionally plot the regression. Psi and K should follow a linear relationship, and
the coefficients of the linear regression will give the parameter of the 1-dimensional response surface that relates the two parameters.
}
