% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-EL4P-Calibrate.R
\name{psi2K_cf_MicroEL4P}
\alias{psi2K_cf_MicroEL4P}
\title{Regress K on Psi and Return Coefficients}
\usage{
psi2K_cf_MicroEL4P(K, psi, plot, ...)
}
\arguments{
\item{K}{vector of K values}

\item{psi}{vector of fitted psi values from \code{\link{psiFit_MicroEL4P}}}

\item{plot}{produce plot}

\item{...}{additional named parameters passed to \code{\link{psi2K_plot_MicroEL4P}}}
}
\value{
regression coefficients
}
\description{
Run a linear regression of K on iverse of psi and extract coefficients; optionally plot the regression. Psi and K should follow a linear relationship, and
the coefficients of the linear regression will give the parameter of the 1-dimensional response surface that relates the two parameters.
}
