% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-EL4P-Calibrate.R
\name{psi2K_plot_MicroEL4P}
\alias{psi2K_plot_MicroEL4P}
\title{Plot Regression of K on Psi}
\usage{
psi2K_plot_MicroEL4P(psi2K, K, psi, label = NULL)
}
\arguments{
\item{psi2K}{regression of K on psi via \code{lm}; (see \code{\link{psi2K_cf_MicroEL4P}})}

\item{K}{vector of K values}

\item{psi}{vector of psi values}

\item{label}{title of plot}
}
\description{
Write docs.
}
