% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Calibration.R
\name{psiFit}
\alias{psiFit}
\title{Objective Function for psi Optimization}
\usage{
psiFit(x, ix, ixEL4P, ixKmesh, PAR, tMax = 150)
}
\arguments{
\item{x}{value of psi from \code{optimize(...)}}

\item{ix}{index of this aquatic population}

\item{ixEL4P}{EL4P aquatic population object corresponding to \code{ix} (see \code{link\{EL4P})}

\item{ixKmesh}{the value of K for this aquatic population}

\item{PAR}{aquatic parameters passed from \code{\link{makePAR_EL4P}}}

\item{tMax}{150; time to run \code{\link{runOne_GEL4P}} before calculating value of objective function}

\item{ixAlpha}{the value of alpha for this aquatic population}
}
\value{
value of objective function at \code{x}
}
\description{
Given a input psi, \code{x}, run a single aquatic population and output objective function; squared error of lambda around the given value of K \eqn{\left ( \lambda-K \right )^{2}}.
This will typically be called by \code{\link{meshK_EL4P}} or \code{optimize}; generally it is the objective function that will be passed to \code{optimize(...)}
}
