% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-EL4P-Calibrate.R
\name{psiFit_MicroEL4P}
\alias{psiFit_MicroEL4P}
\title{Objective Function for Fitting Psi}
\usage{
psiFit_MicroEL4P(x, EL4P, M, eqAqua, K, G, lifespan, tMax = 150)
}
\arguments{
\item{x}{value of psi from \code{optimize(...)}}

\item{EL4P}{an EL4P pool (see \code{\link{EL4P}})}

\item{M}{equilibrium mosquito density at this aquatic habitat}

\item{eqAqua}{vector of probability of oviposition at each aquatic habitat}

\item{K}{carrying capacity at this aquatic habitat}

\item{G}{mean total lifetime egg production of adult female}

\item{lifespan}{average lifespan of mosquito}

\item{tMax}{time to run EL4P pool before returning squared error of empirical lambda around value of K}
}
\value{
value of objective function at \code{x}
}
\description{
Given a input psi, \code{x}, run a single aquatic population and output objective function; squared error of lambda around the given value of K \eqn{\left ( \lambda-K \right )^{2}}.
This will typically be called by \code{\link{meshK_MicroEL4P}} or \code{optimize}; generally it is the objective function that will be passed to \code{optimize(...)}
}
