% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-SimBite.R
\name{queueBites_SimBitePfMOI}
\alias{queueBites_SimBitePfMOI}
\title{PfMOI SimBite \code{HumanPop} Method: Queue SimBites for Population}
\usage{
queueBites_SimBitePfMOI(bites)
}
\arguments{
\item{bites}{a list of length equal to number of humans where each element is a vector of biting times to add to queue with \code{\link{add2Q_SimBitePfMOI}}.}
}
\description{
Add queued bite times to a population; bite times can be generated from \code{\link{SimBite_WaitingTime}}, \code{\link{SimBite_MeanBites}}
This method is bound to \code{HumanPop$queueBites_SimBitePfMOI()}
}
