% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUEUEING-ParametricDistribution.R
\name{queueBitesNegBinom_SimBitePfSI}
\alias{queueBitesNegBinom_SimBitePfSI}
\title{PfSI SimBite \code{\link{HumanPop}} Method: Generate Negative Binomial Biting Distribution}
\usage{
queueBitesNegBinom_SimBitePfSI(tMax, meanNumberBites, shape = 5,
  plot = TRUE)
}
\arguments{
\item{tMax}{maximum time to queue bites}

\item{meanNumberBites}{population mean number of bites (over time period equal to \code{days})}

\item{shape}{shape parameter of gamma-distributed variation in individual biting rate (higher values lead to more normally distributed cumulative biting counts)}

\item{plot}{visualize biting distribution and negative binomial fit}
}
\value{
modify \code{\link{Human}} object with \code{\link{add2Q_SimBitePfSI}}
}
\description{
Wrapper method for \code{\link{SimBite_MeanBites}} to queue simulated bites on a human population.
}
\examples{
HumanPop$queueBitesNegBinom_SimBitePfSI()
}
