% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Biting.R
\name{queueInfectiousBites}
\alias{queueInfectiousBites}
\title{MACRO \code{HumanPop} Method: queueInfectiousBites}
\usage{
queueInfectiousBites()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me! a method for \code{\link{HumanPop}}
}
\examples{
some_function()
}
