% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-SimBite.R
\name{queueVaccination_SimBitePfMOI}
\alias{queueVaccination_SimBitePfMOI}
\title{PfMOI SimBite \code{HumanPop} Method: Queue Vaccination for Population}
\usage{
queueVaccination_SimBitePfMOI(tVaccine, tTreat = NULL, fracPop)
}
\arguments{
\item{tVaccine}{time of vaccination}

\item{tTreat}{time of follow up treatment (if \code{NULL} no treatment)}

\item{fracPop}{fraction of population that recieves vaccination}
}
\description{
Queue vaccination for a population, calling \code{\link{add2Q_pevaccinatePfMOI}}.
If \code{tTreat} is given, queue follow up treatment by calling \code{\link{add2Q_treatPfMOI}}.
This method is bound to \code{HumanPop$queueVaccination_SimBitePfMOI()}
}
