% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{rdtTest_PfSI}
\alias{rdtTest_PfSI}
\title{PfSI \code{Human} Method: PfSI Rapid Diagnostic Test}
\usage{
rdtTest_PfSI(tEvent, PAR)
}
\description{
Administer RDT to this human.
\itemize{
\item if infected: true positive is detected with probability \code{rdtSensPf}, see \code{\link{PfSI.Parameters}}
\item if susceptible: false positive is detected with probability \code{rdtSpecPf}, see \code{\link{PfSI.Parameters}}
}
}
