% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Movement.R
\name{returnHome}
\alias{returnHome}
\title{returnHome}
\usage{
returnHome(ixH, t, PAR = NULL)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
some_function()
}
