% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{RiskQ}
\alias{RiskQ}
\title{MICRO Risk Queue}
\usage{
RiskQ()
}
\description{
Initialize a risk queue for a feeding site object. This function is a method for \code{\link{FeedingSite}}.
}
\section{Methods}{

\itemize{
\item get_N:
\item set_N:
\item get_who:
\item push_who:
\item get_pTm:
\item set_pTm:
\item get_w:
\item push_w:
\item set_nOther:
\item get_nOther:
\item add_OtherHost:
\item get_OtherHost:
}
}

\section{Fields}{

\itemize{
\item a field:
}
}

