% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Calibration.R
\name{run2Eq_GEL4P}
\alias{run2Eq_GEL4P}
\title{Run a Single Aquatic Population to Equilibrium}
\usage{
run2Eq_GEL4P(ix, psi, pop, PAR, tMax = 800, tol = 0.1)
}
\arguments{
\item{ix}{index of this aquatic population}

\item{psi}{value of psi}

\item{pop}{EL4P aquatic population object (see \code{link\{EL4P})}

\item{PAR}{aquatic parameters passed from \code{\link{makePAR_EL4P}}}

\item{tMax}{800; time to run the model to equilibrium}

\item{tol}{target minimum variance in lambda for aquatic populations equilibrium}
}
\value{
the aquatic population
}
