% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-EL4P-Calibrate.R
\name{run2eq_MicroEL4P}
\alias{run2eq_MicroEL4P}
\title{Run a Single Aquatic Population to Equilibrium}
\usage{
run2eq_MicroEL4P(psi, EL4P, M, eqAqua, G, lifespan, tMax = 800,
  var_tol = 0.1)
}
\arguments{
\item{psi}{value of density-dependent mortality psi parameter for this EL4P pool (output of \code{optimize} called on \code{\link{psiFit_MicroEL4P}})}
}
\value{
return the EL4P pool populations at equilibrium values
}
