% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Human-Class.R
\name{runEvent}
\alias{runEvent}
\title{Run a Single Human Event}
\usage{
runEvent(tEvent, PAR, FUN, private, self)
}
\arguments{
\item{tEvent}{time the event will fire}

\item{PAR}{optional list of parameters for FUN}

\item{FUN}{the function that defines the event (FUN takes four arguments: tEvent, PAR, private, self)}

\item{private}{the private environment of the human}

\item{self}{the public environment of the human}
}
\value{
fire a single human event
}
\description{
runEvent is a wrapper than extracts event time, parameters, and passes the private env down to the event function to be called.
All event functions (FUN) must take exactly four arguments; see below.
}
\examples{
runEvent(tEvent=tEvent,PAR=PAR,FUN=FUN,private=private,self=self)
}
