% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Calibration.R
\name{runOne_GEL4P}
\alias{runOne_GEL4P}
\title{Run a Single Aquatic Population for \code{psiFit}}
\usage{
runOne_GEL4P(ix, psi, pop, PAR, tMax = 150)
}
\arguments{
\item{ix}{index of this aquatic population}

\item{psi}{value of psi}

\item{pop}{EL4P aquatic population object (see \code{link\{EL4P})}

\item{PAR}{aquatic parameters passed from \code{\link{makePAR_EL4P}}}

\item{tMax}{150; time to run the dynamic model}
}
\value{
the aquatic population
}
\description{
Run a single aquatic population's daily difference equation dynamics for \code{\link{psiFit}}
}
