% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-ChooseHost.R
\name{sampleIx_utility}
\alias{sampleIx_utility}
\title{Sample Indices}
\usage{
sampleIx_utility(x, ...)
}
\description{
Wrapper for \code{\link{sample}} that will not lead to unexpected behavior when \code{x} is of length 1.
\itemize{
\item This function is used in \code{\link{mbitesGeneric_chooseHost}}
}
}
