% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Search-MicroKernel.R
\name{SEARCH.MicroKernel.Setup}
\alias{SEARCH.MicroKernel.Setup}
\title{MICRO Search Kernels: Initialize Additional Methods & Fields in \code{MicroMosquitoPop} and \code{MicroMosquito}}
\usage{
SEARCH.MicroKernel.Setup(overwrite = TRUE)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Initialize MICRO Search Kernels module of mosquito search behavior.
}
\examples{
SEARCH.MicroKernel.Setup()
}
