% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMANS-Pf-Methods.R
\name{set_biteWeight}
\alias{set_biteWeight}
\title{Set \code{Human} Personal Biting Propensity}
\usage{
set_biteWeight(biteWeight)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
set_biteWeight()
}
