% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Tile-Methods.R
\name{set_FemalePop_MicroTile}
\alias{set_FemalePop_MicroTile}
\title{MICRO \code{\link{MicroTile}}: Re-initialize \code{\link{MicroMosquitoPopFemale}}}
\usage{
set_FemalePop_MicroTile(MosquitoPop_PAR)
}
\arguments{
\item{MosquitoPop_PAR}{output of \code{\link{MicroMosquitoPop.Setup}}}
}
\description{
Erase the old \code{private$FemalePop} object in the tile and re-allocate.
This function is needed for running MBITES-Cohort models, themselves needed to parameterize EL4P Aquatic Ecology Module.
}
\section{How to use with fitting EL4P}{

\enumerate{
\item generate a list of parameters and set methods with \code{\link{MicroMosquitoPop.Setup}} with argument \code{cohort=TRUE}
\item initialize a \code{MicroTile} as normal
\item run the cohort simulation and output data
\item use \code{\link{EL4P.Mesh.Fit}} or \code{\link{EL4P.Landscape.Fit}} to fit EL4P; see those functions for details
\item make a new list of parameters and overwrite cohort methods with \code{\link{MicroMosquitoPop.Setup}} with argument \code{cohort=FALSE}
\item use this method to re-allocate the females using the parameters and methods from step 4.
}
\itemize{
\item This method is bound to \code{MicroTile$set_FemalePop}
}
}

