% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-Queues.R
\name{set_MicroEggQ}
\alias{set_MicroEggQ}
\title{MICRO \code{\link{AquaticSite}} Method: Set the EggQ}
\usage{
set_MicroEggQ(EggQ, ixQ = NULL)
}
\arguments{
\item{EggQ}{the object to insert; if \code{ixQ = NULL} then it should be a full EggQ object, see \code{\link{allocEggQ}} for details, else see \code{\link{eggBatch}} for the structure of a single batch.}

\item{ixQ}{if \code{NULL} set the entire EggQ, else, set the slot \code{ixQ}}
}
\description{
Set either a single slot or the entire EggQ.
This method is bound to \code{AquaticSite$set_EggQ()}.
}
