% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Human-ActivitySpace.R
\name{set_MicroHuman_ActivitySpace}
\alias{set_MicroHuman_ActivitySpace}
\title{MICRO: Set \code{\link{Human}} ActivitySpace}
\usage{
set_MicroHuman_ActivitySpace(nDaily, Nplaces, p, loc)
}
\arguments{
\item{nDaily}{average daily number of other sites visited}

\item{Nplaces}{number of places visited}

\item{p}{weight on proportion of time spent at home}

\item{loc}{vector of other sites visited}
}
\description{
This function is bound to \code{Human$set_ActivitySpace()}
}
