% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Human-Setup.R
\name{set_MicroHuman_FemalePopPointer}
\alias{set_MicroHuman_FemalePopPointer}
\title{MICRO: Set \code{\link{Human}} \code{\link{MicroMosquitoPopFemale}} Pointer}
\usage{
set_MicroHuman_FemalePopPointer(FemalePopPointer)
}
\arguments{
\item{LandscapePointer}{the R6 \code{\link{MicroMosquitoPopFemale}} object to point to}
}
\description{
This function is bound to \code{Human$set_FemalePopPointer()}
}
