% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Human-Setup.R
\name{set_MicroHuman_LandscapePointer}
\alias{set_MicroHuman_LandscapePointer}
\title{MICRO: Set \code{\link{Human}} \code{\link{Landscape}} Pointer}
\usage{
set_MicroHuman_LandscapePointer(LandscapePointer)
}
\arguments{
\item{LandscapePointer}{the R6 \code{\link{Landscape}} object to point to}
}
\description{
This function is bound to \code{Human$set_LandscapePointer()}
}
