% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Human-Setup.R
\name{set_MicroHumanPop_TilePointer}
\alias{set_MicroHumanPop_TilePointer}
\alias{set_MicroHumanPop_TilePointer}
\title{MICRO: Set \code{\link{HumanPop}} \code{\link{MicroTile}} Pointer}
\usage{
set_MicroHumanPop_TilePointer(LandscapePointer)

set_MicroHumanPop_TilePointer(LandscapePointer)
}
\arguments{
\item{LandscapePointer}{the R6 \code{\link{Landscape}} object to point to}

\item{TilePointer}{the R6 \code{\link{MicroTile}} object to point to}
}
\description{
This function is bound to \code{HumanPop$set_TilePointer()}

This function is bound to \code{HumanPop$set_LandscapePointer()}
}
