% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-Queues.R
\name{set_MicroImagoQ}
\alias{set_MicroImagoQ}
\title{MICRO \code{\link{AquaticSite}} Method: Set the ImagoQ}
\usage{
set_MicroImagoQ(ImagoQ, ixQ = NULL)
}
\arguments{
\item{ImagoQ}{the object to insert; if \code{ixQ = NULL} then it should be a full ImagoQ object, see \code{\link{allocImagoQ}} for details, else see \code{\link{eggBatch}} for the structure of a single batch.}

\item{ixQ}{if \code{NULL} set the entire ImagoQ, else, set the slot \code{ixQ}}
}
\description{
Set either a single slot or the entire ImagoQ.
This method is bound to \code{AquaticSite$set_ImagoQ()}.
}
