% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Methods.R
\name{set_MosquitoPointer}
\alias{set_MosquitoPointer}
\title{Set \code{Human} \code{MacroMosquitoPop} Pointer}
\usage{
set_MosquitoPointer(MosquitoPointer)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
some_function()
}
