% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Biting.R
\name{set_myEIR}
\alias{set_myEIR}
\title{Set \code{Human} myEIR}
\usage{
set_myEIR(myEIR)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
some_function()
}
