% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Macro-Emerge.R
\name{set_PatchesEggQ}
\alias{set_PatchesEggQ}
\title{Set \code{MacroPatch} EggQ (Egg Batches)}
\usage{
set_PatchesEggQ(PatchesEggQ, ixP = NULL)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
some_function()
}
