% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Macro-Emerge.R
\name{set_PatchesImagoQ}
\alias{set_PatchesImagoQ}
\title{Set \code{MacroPatch} ImagoQ (Emerging Adults Queue)}
\usage{
set_PatchesImagoQ(PatchesImagoQ, ixP = NULL)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
some_function()
}
