% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Methods.R
\name{set_PatchesPointer}
\alias{set_PatchesPointer}
\title{Set \code{Human} patchID}
\usage{
set_PatchesPointer(PatchesPointer)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
some_function()
}
