% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Macro-Emerge.R
\name{set_season}
\alias{set_season}
\title{Set \code{MacroPatch} Seasonal Emergence}
\usage{
set_season(season, ixP = NULL)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
some_function()
}
