% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Methods.R
\name{set_TilePointer}
\alias{set_TilePointer}
\title{Set \code{Human} \code{MacroTile} Pointer}
\usage{
set_TilePointer(TilePointer)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
some_function()
}
