% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Calibration.R
\name{setupAquaPop_EL4PsamplePoints}
\alias{setupAquaPop_EL4PsamplePoints}
\title{Fit EL4P Aquatic Ecology Model on Sample Grid}
\usage{
setupAquaPop_EL4PsamplePoints(PAR, gridN, tol = 0.1, plot = FALSE)
}
\arguments{
\item{PAR}{named list of parameters calculated from \code{\link{makePAR_EL4P}}}

\item{gridN}{number of points to sample in K}

\item{tol}{target minimum variance in lambda for aquatic populations equilibrium}

\item{plot}{produce diagnostic plots?}
}
\value{
named list of fitted parameters
\itemize{
\item EL4P_pops: fitted EL4P aquatic populations
\item PAR: input PAR with fitted values of \code{psi} and \code{alpha} appended
}
}
\description{
Fit the EL4P Aquatic Ecology module on a sampling grid of K values to match a desired level of daily emergence at equilibrium.
This means that the site-specific density dependent mortality parameter \code{psi} will be fit
to K based on a sampling grid of values for K in log-space. If \code{plot = TRUE}, the linear regression of \code{psi} against logged values of K
should show exact linear dependence, indicating fitted \code{psi} will produce desired level of lambda at equilibrium. Compare with \code{\link{setupAquaPop_EL4Pexact}} which will
fit \code{psi} based on an exact LANDSCAPE. This will also return the coefficients of a linear regression of K on psi (see \code{\link{psi2K_cf}}) to give the functional relationship between K and psi.
}
\examples{
setupAquaPop_EL4PsamplePoints(PAR, gridN, tol = 0.1, plot = FALSE)
}
