% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Human-ActivitySpace.R
\name{sim_MicroHuman_ActivitySpace}
\alias{sim_MicroHuman_ActivitySpace}
\title{MICRO \code{\link{Human}} Method: Initialize Activity Space}
\usage{
sim_MicroHuman_ActivitySpace()
}
\description{
This function is bound to \code{Human$sim_ActivitySpace()}
}
