% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Human-ActivitySpace.R
\name{sim_MicroHumanPop_ActivitySpace}
\alias{sim_MicroHumanPop_ActivitySpace}
\title{MICRO \code{\link{HumanPop}} Method: Simulate Activity Space}
\usage{
sim_MicroHumanPop_ActivitySpace()
}
\description{
This function is bound to \code{HumanPop$sim_ActivitySpace()}
}
