% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PFSI-PfSI.R
\name{simbite_PfSI}
\alias{simbite_PfSI}
\title{\code{XX} Family: Simulated Bite}
\usage{
simbite_PfSI(ixH, t, PAR)
}
\arguments{
\item{ixH}{index of human}

\item{t}{time the event will occur}

\item{PAR}{passed from Pf0}
}
\description{
This function runs a simulated bite on a human. It is called by \code{\link{event_simbitePfSI}} and simulates \code{\link{probeHost_PfSI}}.
For more information on how this family of functions works, see \url{http://smitdave.github.io/MASH-Development/docHumans.html#add2q_xx}
}
\examples{
simbite_PfSI(ixH, t, PAR)
}
