% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUEUEING-ParametricDistribution.R
\name{SimBite_WaitingTime}
\alias{SimBite_WaitingTime}
\title{Simulate Biting Process from Population Mean Waiting Time Between Bites}
\usage{
SimBite_WaitingTime(nH, meanWaitingTime, days = 365, shape = 5,
  plot = TRUE)
}
\arguments{
\item{nH}{number of humans}

\item{meanWaitingTime}{population mean waiting time between bites}

\item{days}{number of days}

\item{shape}{shape parameter of gamma-distributed variation in individual biting rate (higher values lead to more normally distributed cumulative biting counts)}

\item{plot}{visualize biting distribution and negative binomial fit}
}
\value{
a list of length \code{nH}; each element is the arrivial times of bites on that individual
}
\description{
Simulate a biting process to result in a negative-binomially distributed cumulative number of bites per person over \code{days}.
The biting process for each person is a Poisson process with individual gamma-distributed rates.
Because the negative binomial distribution arises as a compound Poisson-Gamma distribution, it will also arise from independent Poisson processes each with gamma-distributed rates,
because the expected value of a Poisson process is also Poisson.
This calls \code{\link{gammaRates}}.
}
