% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-SimBite.R
\name{SimBitePfMOI}
\alias{SimBitePfMOI}
\title{PfMOI SimBite \code{Human} Event: PfMOI Simulated Bite Event}
\usage{
SimBitePfMOI(tEvent, PAR)
}
\arguments{
\item{tEvent}{time of bite}

\item{PAR}{named list
\itemize{
\item mosquitoPfSI: object of class \code{\link{mosquitoPfMOI}}
}}
}
\description{
Simulate a PfMOI Simulated Bite. This calls \code{\link{probeHost_PfMOI}} to simulate host probing by an infectious mosquito.
This method is bound to \code{Human$SimBitePfMOI()}
}
