% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Tile-Methods.R
\name{simMacro}
\alias{simMacro}
\title{MACRO \code{MacroTile} Simulate MACRO Dynamics}
\usage{
simMacro(tMax)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
write more stuff!!!!!!!!
}
\examples{
some_function()
}
