% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Tile-Methods.R
\name{simMICRO_oneStep}
\alias{simMICRO_oneStep}
\title{MICRO \code{\link{MicroTile}}: Run Simulation one Time Step}
\usage{
simMICRO_oneStep(timeStep = 1, print = FALSE)
}
\arguments{
\item{timeStep}{the time step of the model}
}
\description{
Run MICRO simulation for one time step, the length of which defines the temporal window for indifference to contingent events.
}
