% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMANS-HumanPop-Parameters.R
\name{siteAges}
\alias{siteAges}
\title{Generate Ages for \code{HumanPop}}
\usage{
siteAges(N)
}
\arguments{
\item{N}{size of population (may refer to household \code{\link{FeedingSite}} or an individual \code{\link{MacroPatch}}) patch}
}
\value{
numeric vector
}
\description{
Generate numeric vector of ages. This function is called in \code{\link{sitePops}}
}
\examples{
hhAges(N = 10)
}
