% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMANS-HumanPop-Parameters.R
\name{sitePops}
\alias{sitePops}
\title{Generate Site Sizes for \code{HumanPop}}
\usage{
sitePops(N, siteSize = 10, siteMin = 2)
}
\arguments{
\item{N}{number of sites (may refer to household or a \code{\link{MacroPatch}} patch)}

\item{siteSize}{average size of population at sites}

\item{siteMin}{minimum size of population at sites}
}
\value{
list:
\itemize{
\item nHumans: total human population size
\item sitePops: vector of population size at each site
\item siteHumanID: list of human IDs at each site
\item homeHumanID: vector of home site ID for each human
\item siteAges: list of human ages at each site; calculated from \code{\link{siteAges}}
}
}
\description{
Site may be defined as a MICRO \code{\link{FeedingSite}} or a MACRO \code{\link{MacroPatch}}.
}
\examples{
sitePops(N = 10)
}
