% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Patch-Methods.R
\name{sumKappa}
\alias{sumKappa}
\title{Set \code{MacroPatch} sumKappa}
\usage{
sumKappa()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
generate local human biting propensity kappa
}
\examples{
some_function()
}
