% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-Queues.R
\name{track_MicroEggQ}
\alias{track_MicroEggQ}
\title{MICRO \code{AquaticSite} Method: Track the EggQ}
\usage{
track_MicroEggQ()
}
\description{
Return the total number of eggs in this EggQ
This method is bound to \code{AquaticSite$track_EggQ()}.
DEV_NOTE: eventually may want to add argument time or something and only accumulate those values tm <= time or something...
}
