% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-Queues.R
\name{track_MicroImagoQ}
\alias{track_MicroImagoQ}
\title{MICRO \code{\link{AquaticSite}} Method: Track the ImagoQ}
\usage{
track_MicroImagoQ()
}
\description{
Return the total number of emerging adults in this ImagoQ
This method is bound to \code{AquaticSite$track_ImagoQ()}.
DEV_NOTE: eventually may want to add argument time or something and only accumulate those values tm <= time or something..., also might want to filter genotype
}
