% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Travel.R
\name{track_travel}
\alias{track_travel}
\title{Track \code{Human} Travel History}
\usage{
track_travel(tTravel, locationH)
}
\arguments{
\item{tTravel}{time of trip}

\item{location}{destination of trip}
}
\value{
does stuff
}
\description{
Write me! Defined in MACRO-Human-Travel.R
}
\examples{
self$trackTravel(tTravel = 1, location = 1L)
}
