% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUATIC-ECOLOGY.R
\name{trackEggQ}
\alias{trackEggQ}
\title{Write EggQ Values to .csv}
\usage{
trackEggQ(con)
}
\arguments{
\item{con}{text connection to write to}
}
\value{
none
}
\description{
Generate daily output of EggQ tracking to .csv file through text connection created with \code{trackEggQ_init()}.
}
\examples{
trackEggQ(con)
}
