% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUATIC-ECOLOGY.R
\name{trackEggQ_init}
\alias{trackEggQ_init}
\title{Initialize Text Connection to Track EggQ}
\usage{
trackEggQ_init(directory, fileName)
}
\arguments{
\item{directory}{directory that will be created; files will be put in directory/OUTPUT/..}

\item{fileName}{name of the file to write to; directory/OUTPUT/fileName.csv}
}
\value{
a text connection
}
\description{
Generate a text connection to write .csv EggQ values to. This records daily egg laying in each aquatic habitat.
The text connection generated must be named \code{EggQCon}; this function also defines a global flag \code{EggQ_TRACK} that is
used by \code{oneDay_EL4P()} or \code{oneDay_emerge()} to determine whether or not to write to output.
}
\examples{
trackEggQ_init(directory, fileName)
}
