% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfPedigree.R
\name{trackPfTransmission}
\alias{trackPfTransmission}
\title{Add New Transmisson Event to PfTransmission}
\usage{
trackPfTransmission(M2H)
}
\arguments{
\item{M2H}{direction of transmission}

\item{tBite}{time of infectious bite (mosquito to human)}

\item{ixH}{index of human host}

\item{ixM}{index of mosquito vector}

\item{ixS}{site of infectious bite}

\item{PfM}{Pf object from infectious mosquito}
}
\value{
Write the following Pf pedigree to file
\itemize{
\item tStart: time of start of infection
\item tBite: time of infectious bite
\item tMosy: time mosquito initially infected
\item ixH: index of human host
\item ixS: site of vector to human transmission
\item ixM: index of mosquito vector
\item damID: female gametocyte ID
\item sireID: male gametocyte ID
\item PfID: the Pf ID; every infection that makes it to bloodstream stage is considered a new clonal variant
}
}
\description{
Add a new transmission event PfTransmission .csv for data logging.
}
