% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Movement.R
\name{travelHabit}
\alias{travelHabit}
\title{MACRO \code{Human} Method: travelHabit}
\usage{
travelHabit(n, freqMean = 7, freqSd = 2, lengthMean = 2, lengthSd = 1,
  tNow = 0)
}
\arguments{
\item{n}{number of other patches to visit}

\item{freqMean}{meanlog}

\item{freqSd}{sdlog}
}
\value{
does stuff
}
\description{
Write me! a method for \code{\link{Human}}
}
\examples{
some_function()
}
