% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{treatPfSI}
\alias{treatPfSI}
\title{PfSI \code{Human} Event: PfSI Treatment Event}
\usage{
treatPfSI(tEvent, PAR)
}
\arguments{
\item{tEvent}{time of treatment}

\item{PAR}{\code{NULL}}
}
\description{
Simulate a PfSI treatment event. If the human is infected, set susceptible and track history; also initiate period of chemoprophylaxis, see \code{\link{add2Q_endprophylaxisPfSI}}
This method is bound to \code{Human$treatPfSI()}
}
