% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{ttFeverPfMOI}
\alias{ttFeverPfMOI}
\title{PfMOI \code{Human} Method: Timing of Fever Incident}
\usage{
ttFeverPfMOI()
}
\description{
What is the timing of fever relative to the start of a PfMOI infection? See \code{\link{PfMOI.Parameters}} parameter \code{Pf_ttF} mean (on natural scale) and parameter \code{Pf_ttFvar} standard deviation (on natural scale) of log-normally distributed time to fever.
This method is called from \code{\link{}}
This method is bound to \code{Human$ttFeverPfMOI()}
}
