% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{ttInfectPfMOI}
\alias{ttInfectPfMOI}
\title{PfMOI \code{Human} Method: Duration of Latency}
\usage{
ttInfectPfMOI()
}
\description{
How many days after the infectious bite does a PfMOI infection start? See \code{\link{PfMOI.Parameters}} parameter \code{Pf_latent} constant latent period.
This method is called from \code{\link{}}
This method is bound to \code{Human$ttInfectPfMOI()}
}
