% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{ttSusceptiblePfMOI}
\alias{ttSusceptiblePfMOI}
\title{PfMOI \code{Human} Method: Duration of Protection from Chemoprophylaxis}
\usage{
ttSusceptiblePfMOI()
}
\description{
After administration of treatment what is time to susceptibility? See \code{\link{PfMOI.Parameters}} parameter \code{Pf_ttS} constant timing period.
This method is called from \code{\link{}}
This method is bound to \code{Human$ttSusceptiblePfMOI()}
}
