% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Patch-Methods.R
\name{updateKappa}
\alias{updateKappa}
\title{Set \code{MacroPatch} updateKappa}
\usage{
updateKappa()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Update normalized biting propensities for all patches
}
\examples{
some_function()
}
