% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-EL4P-Methods.R
\name{updateLandscape_MicroEL4P}
\alias{updateLandscape_MicroEL4P}
\title{Update \code{\link{Landscape}} and \code{\link{AquaticSite}} after Fitting EL4P Aquatic Ecology Module}
\usage{
updateLandscape_MicroEL4P(EL4P_fit, EL4P_PAR)
}
\arguments{
\item{EL4P_fit}{output of \code{\link{EL4P.Mesh.Fit}}}

\item{EL4P_PAR}{output of \code{\link{EL4P.Parameters}}}
}
\description{
This function updates parameters psi, alpha, and p in a \code{\link{Landscape}} and \code{\link{AquaticSite}}
It modifies \code{\link{AquaticSite}} and \code{\link{Landscape}} classes.
}
\details{
\itemize{
\item This method is bound to \code{Landscape$updateLandscapeEL4P}
}
}
