% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASH-Utilities.R
\name{util_List2Df}
\alias{util_List2Df}
\title{Utility: Convert List to Data Frame}
\usage{
util_List2Df(list)
}
\arguments{
\item{list}{a list, each element of which will be transformed into a row of a data frame.}
}
\value{
data frame
}
\description{
Convert a list to a data frame.
}
\examples{
util_List2Df(list = lapply(1:100,MASH:::util_PfSISlice))
}
