% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Utilities.R
\name{util_PfSIHistory}
\alias{util_PfSIHistory}
\title{PfSI: Calculate State Space Occupancy Vector}
\usage{
util_PfSIHistory(history, cpp = TRUE)
}
\arguments{
\item{history}{human event histories from \code{HumanPop$get_History()}}

\item{cpp}{use Rcpp helper function \code{\link{util_PfSIoneHistory}}}
}
\value{
list
}
\description{
From history output of a \code{\link{HumanPop}}, convert individual histories in to a continuous time occupany vector.
The PfSI module has the following state space:
\itemize{
\item S: susceptible
\item I: infected
\item F: fever
\item P: chemoprophylactic protection
\item PEvaxx: begin PE vaccination protection
\item PEwane: end PE vaccination protection
\item GSvaxx: begin GS vaccination protection
\item GSwane: end GS vaccination protection
}
}
\examples{
util_PfSIHistory()
}
