% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{util_PfSIoneHistory}
\alias{util_PfSIoneHistory}
\title{PfSI: Write One Human History}
\usage{
util_PfSIoneHistory(historyIxH, timeBins, timeSeries)
}
\arguments{
\item{historyIxH}{a single human history from \code{Human$get_History()}}

\item{timeBins}{vector of time bins (state transition times)}

\item{timeSeries}{matrix of occupancy vectors}
}
\description{
This is a C++ helper function for \code{\link{util_PfSIHistory}}
}
