% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Utilities.R
\name{util_PfSISlice}
\alias{util_PfSISlice}
\title{PfSI: Empty Occupancy Vector}
\usage{
util_PfSISlice(time)
}
\arguments{
\item{time}{current time slice}
}
\value{
list
}
\description{
Make an empty occupancy vector at a single time slice for the \code{PfSI} module. This is a helper function called by \code{\link{util_PfSIHistory}}
}
\examples{
util_PfSISlice(time = 0)
}
