% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILITIES-Human.R
\name{writeHumanEvent}
\alias{writeHumanEvent}
\title{Write eventT and events from HUMANS to .json}
\usage{
writeHumanEvent(directory, fileName)
}
\arguments{
\item{directory}{directory; files will be put in directory/OUTPUT/..}

\item{fileName}{name of the file to write to; directory/OUTPUT/fileName.csv}
}
\value{
nothing
}
\description{
Write HUMANS histories to .json file.
}
\examples{
writeHumanEvent(directory, fileName)
}
