% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILITIES-Human.R
\name{writeHumanEvent_PfSI}
\alias{writeHumanEvent_PfSI}
\title{Write eventT and events from HUMANS to .json}
\usage{
writeHumanEvent_PfSI(directory, fileName, overWrite = FALSE)
}
\arguments{
\item{directory}{directory; files will be put in directory/OUTPUT/..}

\item{fileName}{name of the file to write to; directory/OUTPUT/fileName.csv}

\item{overWrite}{allow overwriting of files in directory/OUTPUT/..? Use with caution. Generally it is reccomended to use \code{\link{clearOutput}} to clear the OUTPUT/.. folder after moving data to a permanant location and prior to running the simulation.}
}
\value{
nothing
}
\description{
Write HUMANS histories from PfSI module to .json file. This should typically be called at the end of a model run, assuming demography is not implemented.
}
\examples{
writeHumanEvent(directory, fileName overWrite = FALSE)
}
