% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-Queues.R
\name{zero_MicroEggQ}
\alias{zero_MicroEggQ}
\title{MICRO \code{AquaticSite} Method: Zero out a slot in the EggQ}
\usage{
zero_MicroEggQ(ixQ)
}
\arguments{
\item{ixQ}{the slot in the EggQ to zero out}
}
\description{
This method zeros out a slot in the EggQ.
This method is bound to \code{AquaticSite$zero_EggQ()}.
}
