% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-Queues.R
\name{zero_MicroImagoQ}
\alias{zero_MicroImagoQ}
\title{MICRO \code{\link{AquaticSite}} Method: Zero out a slot in the ImagoQ}
\usage{
zero_MicroImagoQ(ixQ)
}
\arguments{
\item{ixQ}{the slot in the ImagoQ to zero out}
}
\description{
This method zeros out a slot in the ImagoQ.
This method is bound to \code{AquaticSite$zero_ImagoQ()}.
}
