% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUATIC-ECOLOGY.R
\name{zeroAdult}
\alias{zeroAdult}
\title{Zero out a space in ImagoQ}
\usage{
zeroAdult(ixQ, ix)
}
\arguments{
\item{ixQ}{index of position in ImagoQ to zero out}

\item{ix}{index of aquatic habitat of emergence}
}
\value{
modifies global LANDSCAPE$aquaSites[[ix]]$ImagoQ[[ixQ]]
}
\description{
This function is a low-level utility to zero out a position in ImagoQ after emergence. It is called by addCohort.
}
\examples{
zeroAdult(ixQ=1,ix=1)
}
