% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Macro-Emerge.R
\name{addCohort_MacroEmerge}
\alias{addCohort_MacroEmerge}
\title{MACRO: Calculate Emerging Adults for \code{MacroPatch}}
\usage{
addCohort_MacroEmerge()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me! does this for all patches
}
\examples{
some_function()
}
