% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-Emerge-Methods.R
\name{addCohort_MicroEmerge}
\alias{addCohort_MicroEmerge}
\title{MICRO \code{\link{Landscape}} Method: Get Emerging Adults from ImagoQ and Zero out ImagoQ}
\usage{
addCohort_MicroEmerge()
}
\value{
does stuff
}
\description{
Grab emerging adult batches where tEmerge <= tNow and add to the \code{\link{MicroMosquitoPopFemale}}.
This method is bound to \code{Landscape$addCohort_MicroEmerge()}
}
\examples{
some_function()
}
