% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-EL4P-Calibrate.R
\name{EL4P.Parameters}
\alias{EL4P.Parameters}
\title{Initialize EL4P Aquatic Ecoogy Module Parameters}
\usage{
EL4P.Parameters(nAqua, nHumans, R0, M, eqAqua, EIP, lifespan, S,
  alpha.m = 0.8, p = 0.9, K.a = 1, K.b = 1)
}
\arguments{
\item{nAqua}{number of aquatic habitats on landscape}

\item{nHumans}{number of humans on landscape}

\item{R0}{desired pathogen transmission level at equilibrium}

\item{M}{vector of densities of mosquitoes at each aquatic habitat}

\item{eqAqua}{vector of probability of oviposition at each aquatic habitat (see \code{\link{aquaIx_equilibrium}})}

\item{EIP}{length of the entomological incubation period}

\item{lifespan}{average lifespan of mosquito}

\item{S}{stability index; number of bites on humans over the average mosquito lifespan}

\item{alpha.m}{mean of density independent survival parameter \code{alpha} from pupae to adult}

\item{p}{expected fraction of cohort that advances to next life stage (1/p is expected time spent in each stage L1,L2,L3,L4,P)}

\item{K.a}{shape parameter for gamma distributed weights on K}

\item{K.b}{scale parameter for gamma distributed weights on K}
}
\description{
Generate named list of EL4P parameters.
}
