% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Macro-Emerge.R
\name{emergingAdults_MacroEmerge}
\alias{emergingAdults_MacroEmerge}
\title{MACRO: Calculate Emerging Adults from ImagoQ for \code{MacroPatch}}
\usage{
emergingAdults_MacroEmerge()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me! does this for all patches. generates a vector of emerging adults, and zeros out the ImagoQ
}
\examples{
some_function()
}
