% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-Emerge.R
\name{emergingAdults_MicroEmerge}
\alias{emergingAdults_MicroEmerge}
\title{MICRO \code{\link{Landscape}} Method: Get Emerging Adults from ImagoQ and Zero out ImagoQ}
\usage{
emergingAdults_MicroEmerge()
}
\description{
Grab emerging adult batches where tEmerge <= tNow and return as a list.
This is a helper function for \code{\link{addCohort_MicroEmerge}}.
This method is bound to \code{Landscape$emergingAdults_Emerge()}.
}
