% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-Queues.R
\name{extend_MicroEggQ}
\alias{extend_MicroEggQ}
\title{MICRO \code{\link{AquaticSite}} Method: Extend the EggQ}
\usage{
extend_MicroEggQ()
}
\description{
This method extends the EggQ by 2 times its current length of empty egg batch objects, see \code{\link{newEgg}} for the egg batch structure.
This method is bound to \code{AquaticSite$extend_EggQ()}.
}
