% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Search-MicroKernel.R
\name{get_MicroKernel_movement_BRO}
\alias{get_MicroKernel_movement_BRO}
\title{MICRO Search Kernels: \code{\link{MicroMosquitoPopFemale}} Access MvAll Object for M-BITES BRO Lifecycle Model}
\usage{
get_MicroKernel_movement_BRO(ixS, state, inPointSet)
}
\description{
Replace generic \code{MicroMosquitoPopFemale$get_movement()} method for MicroKernel module; it will be bound to \code{MicroMosquitoPopFemale$get_movement()}
}
