% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEARCH-MicroKernel-Move.R
\name{get_MicroKernel_movement_BROM}
\alias{get_MicroKernel_movement_BROM}
\title{MICRO Search Kernels: \code{\link{MicroMosquitoPopFemale}} Access MvAll Object for M-BITES BROM Lifecycle Model}
\usage{
get_MicroKernel_movement_BROM(ixS, state, inPointSet)
}
\description{
Replace generic \code{MicroMosquitoPopFemale$get_movement()} method for MicroKernel module; it will be bound to \code{MicroMosquitoPopFemale$get_movement()}
}
