% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEARCH-MICRO-KernelSetup.R
\name{get_MicroKernel_movement_FULL}
\alias{get_MicroKernel_movement_FULL}
\title{MICRO Search Kernels: \code{\link{MicroMosquitoPopFemale}} Access MvAll Object for Full M-BITES Lifecycle Model}
\usage{
get_MicroKernel_movement_FULL(ixS, state, inPointSet)
}
\description{
Replace generic \code{MicroMosquitoPopFemale$get_movement()} method for MicroKernel module; it will be bound to \code{MicroMosquitoPopFemale$get_movement()}
}
