% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{HumanEventQ}
\alias{HumanEventQ}
\title{Human Event Queue}
\usage{
HumanEventQ(initQ = 100)
}
\arguments{
\item{initQ}{reserve memory for \code{initQ} number of events}
}
\description{
A class for queueing human events. Header file defined \href{https://github.com/slwu89/RcppQueues/blob/master/inst/include/RcppQueues/HumanEventQ.hpp}{here} and \code{RcppR6} export handled \href{https://github.com/slwu89/RcppQueues/blob/master/inst/HumanEventQ.yml}{here}.
}
\section{Methods}{

\itemize{
\item hi
\itemize{
\item nope.
}
}
}

