% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Class.R
\docType{class}
\name{humanPfMOI}
\alias{humanPfMOI}
\title{PfMOI Human-stage Pathogen Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
humanPfMOI
}
\description{
This is the human-stage PfMOI module pathogen object defined in \code{\link{Human}}.
}
\section{Methods}{

\itemize{
\item \strong{Constructor}
\itemize{
\item new: initialize a new \code{Landscape} object
\itemize{
\item Arguments:
\itemize{
\item \code{Landscape_PAR}: see \code{\link{Landscape.Parameters}} for parameter generation function.
}
}
}
\item \strong{Getters & Setters}
\itemize{
\item get_XX: get \code{private$XX}
\itemize{
\item Arguments:
\itemize{
\item \code{ixS}: if \code{NULL} return all sites, otherwise return site indexed by \code{ixS}
}
}
\item set_FeedingSites: set \code{private$FeedingSites}
\itemize{
\item Arguments:
\itemize{
\item \code{FeedingSites}: replacement object (see below)
\item \code{ixS}: if \code{NULL} replace all sites, otherwise replace site indexed by \code{ixS}
}
}
}
\item \strong{Pointers}
\itemize{
\item get_TilePointer: get \code{\link{MicroTile}} pointer
\item set_TilePointer: set \code{\link{MicroTile}} pointer
}
}
}

\section{Fields}{

\itemize{
\item \strong{History}
\itemize{
\item tEvent: time of the event (jump time for embedded CTMC)
\item events: the new state (state after the jump)
\item MOI: current multiplicity of infection
}
}
}

\keyword{R6}
\keyword{class}
