% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMANS-HumanPop-Parameters.R
\name{HumanPop.Parameters}
\alias{HumanPop.Parameters}
\title{Initialize HumanPop Parameters for \code{HumanPop}}
\usage{
HumanPop.Parameters(nSite, bWeight = NULL, siteSize = 10, siteMin = 2)
}
\arguments{
\item{nSite}{number of sites (either \code{\link{FeedingSite}} or \code{link{MacroPatch}})}

\item{bWeight}{numeric value (if \code{NULL} biting weights are Gamma(1,1) distributed)}

\item{siteSize}{average size of population at sites (passed to \code{\link{sitePops}})}

\item{siteMin}{minimum size of population at sites (passed to \code{\link{sitePops}})}
}
\value{
return a list \code{HumanPop_PAR}
}
\description{
make a list of pars for \code{\link{HumanPop}}
}
\examples{
HumanPop.Parameters(nSite = 5)
 * nHumans: total human population size
 * tStart: time to begin simulation
 * humanIDs: human IDs
 * homeIDs: site IDs of each human
 * bDay: birthdays of each human
 * bWeight: biting weight on humans
}
