% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMANS-HumanPop-Parameters.R
\name{HumanPop.Parameters}
\alias{HumanPop.Parameters}
\title{Initialize HumanPop Parameters for \code{HumanPop}}
\usage{
HumanPop.Parameters(nSite, demographics = NULL, ...)
}
\arguments{
\item{nSite}{number of sites (either \code{\link{FeedingSite}} or \code{link{MacroPatch}})}

\item{demographics}{either a list or \code{NULL}; if \code{NULL} then \code{\link{sitePops}} is used to generate these parameters, see that function for details on the list structure that is expected.}

\item{...}{additional named arguments for \code{\link{sitePops}}}
}
\value{
return a list \code{HumanPop_PAR}
}
\description{
make a list of pars for \code{\link{HumanPop}}
}
\examples{
HumanPop.Parameters(nSite = 5)
}
