% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PFSI-PfSI.R
\name{infectiousBite_PfSI}
\alias{infectiousBite_PfSI}
\title{Infectious Bite (Vector to Human Transmission)}
\usage{
infectiousBite_PfSI(tBite, ixH, ixS, ixM, PfM)
}
\arguments{
\item{ixH}{index of human}
}
\value{
a single pfid (integer)
}
\description{
This is the routine that handles a bite from an infectious vector on a human for PfSI module. It calls \code{link{add2Q_startPfSI}}
to begin queuing human infection events if tranmission of sporozoites into the bloodstream is successful (probability \code{b}).
Depending on the global flags \code{PfPedigree_TRACK} and \code{PfTransmission_TRACK} it may call \code{\link{addPf2Pedigree}} or \code{\link{trackPfTransmission}} to log data.
}
\examples{
getPfParent_SI(ixH)
}
