% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-PATHOGEN-PfSI.R
\name{init.PfSI}
\alias{init.PfSI}
\title{Initialize PfSI Module Methods}
\usage{
init.PfSI()
}
\arguments{
\item{write}{me}
}
\value{
write me
}
\description{
Generate a list of parameters and define functions in the global environment for PfSI. All arguments have default values which are listed below before the definition.
}
\examples{
init.PfSI()
}
