% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-PATHOGEN-PfSI.R
\name{init.simbitePfSI}
\alias{init.simbitePfSI}
\title{Initialize PfSI-SimBite Module Methods}
\usage{
init.simbitePfSI()
}
\arguments{
\item{write}{me}
}
\value{
write me
}
\description{
Initialize methods in \code{\link{Human}} and \code{\link{HumanPop}} classes for simulated biting.
This should be run after calling \code{\link{init.PfSI}} if humans are being simulated as a stand-alone component.
}
\examples{
init.simbitePfSI()
}
