% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{init_MACRO_PfSI}
\alias{init_MACRO_PfSI}
\title{PfSI \code{HumanPop} Method: Initialize MACRO PfSI Infections}
\usage{
init_MACRO_PfSI(PfPR, tStart = 0)
}
\arguments{
\item{PfPR}{a vector of parasite prevalence at each patch}

\item{tStart}{time to start simulation}
}
\description{
Initialize PfSI infections with parasite prevalence PfPR for each patch.
This method is bound to \code{HumanPop$init_MACRO_PfSI()}
}
