% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Landscape-Class.R
\docType{class}
\name{Landscape}
\alias{Landscape}
\title{MICRO Landscape Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
Landscape
}
\description{
MICRO Landscape Class Definition
}
\details{

}
\section{Methods}{

\describe{
  \item{\code{new(reg, DataStorageClass.g0, ...)}}{...}
  \item{\code{fit(data)}}{...}
  \item{\code{predict(newdata)}}{...}
  \item{\code{predictAeqa(newdata)}}{...}
}
}

\section{Active Bindings}{

\describe{
  \item{\code{cats}}{...}
}
}

\keyword{R6}
\keyword{class}
