% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Methods.R
\name{MACRO.Humans.Setup}
\alias{MACRO.Humans.Setup}
\title{Initialize Additional Methods & Fields in \code{Human} and \code{HumanPop} for MACRO}
\usage{
MACRO.Humans.Setup()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
MACRO.Humans.Setup()
}
