% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Setup.R
\name{MACRO.Humans.Setup}
\alias{MACRO.Humans.Setup}
\title{MACRO: Initialize Additional Methods & Fields in \code{\link{Human}} and \code{\link{HumanPop}}}
\usage{
MACRO.Humans.Setup(pathogenModule = "PfSI", overwrite = TRUE)
}
\arguments{
\item{pathogenModule}{what PATHOGEN module to use (must be character in "PfSI","PfMOI")}

\item{overwrite}{overwrite existing fields}
}
\description{
Initialize methods and fields for biting and inter-patch travel for a macrosimulation tile \code{\link{MacroTile}}.
}
\examples{
MACRO.Humans.Setup()
}
