% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-MosquitoPop-Parameters.R
\name{MACRO.MosquitoPop.Parameters}
\alias{MACRO.MosquitoPop.Parameters}
\title{Initialize MosquitoPop Parameters for \code{MacroMosquitoPop}}
\usage{
MACRO.MosquitoPop.Parameters(N, M_density = 10, p = 0.9, f = 0.3,
  Q = 0.9, v = 20, maxEIP = 30, psi = NULL)
}
\arguments{
\item{some}{stuff}
}
\value{
return a list \code{MacroMosquitoPop_PAR}
}
\description{
make a list of pars for \code{\link{MacroMosquitoPop}}
}
\examples{
MACRO.MosquitoPop.Parameters(N = 5)
}
