% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Tile-Parameters.R
\name{MACRO.Tile.Parameters}
\alias{MACRO.Tile.Parameters}
\title{Initialize MACRO Tile Parameters for \code{MacroTile}}
\usage{
MACRO.Tile.Parameters(N = 10, patchSize = 20, patchMin = 10,
  aquaEcology = list(aquaModel = "emerge", lambda = rep(10, 10)))
}
\arguments{
\item{N}{number of patches}

\item{patchSize}{passed to \code{\link{sitePops}}}

\item{patchMin}{passed to \code{\link{sitePops}}}

\item{more}{eventually want to pass more params to \code{\link{MACRO.Patch.Parameters}}}
}
\value{
return a list
}
\description{
This is used to generate a list of parameters for \code{\link{MacroTile}} and should be used during its initialization.
\code{MACRO.Tile.Parameters} will generate parameters that are passed to the \code{\link[R6]{initialize}} method of
the component objects in a tile:
\itemize{
\item \code{\link{HumanPop}}: creates parameters by calling \code{\{link{HumanPop.Parameters}}
\item \code{\link{MacroPatch}}: creates parameters by calling \code{\link{MACRO.Patch.Parameters}}
\item \code{\link{MacroMosquitoPop}}: directly initializes parameters
}
}
\examples{
MACRO.Tile.Parameters()
}
