% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO-Parameters.R
\name{MBITES.BRO.Parameters}
\alias{MBITES.BRO.Parameters}
\title{Generate Parameters for M-BITES 'BRO' Module}
\usage{
MBITES.BRO.Parameters(B_wts = rep(1, 5), R_wts = rep(1, 5), O_wts = rep(1,
  5), InAndOut = matrix(c(c(4, 2, 1, 0, 1), c(2, 1, 1, 0, 1), c(2, 1, 1, 0,
  2), c(1, 1, 1, 0, 1), c(6, 4, 2, 1, 0)), byrow = FALSE, ncol = 5),
  B_surv = 0.85, B_succeed = 0.8, B_time = 3/4, surviveH = 1,
  probeH = 1, surviveprobeH = 1, feedH = 1, surviveZ = 1, feedZ = 1,
  R_surv = 0.9, R_time = 0.8, O_surv = 0.8, O_succeed = 0.7,
  O_time = 1.3, bm.a = 7.5, bm.b = 2.5, REFEED = FALSE, rf.a = 60,
  rf.b = 5000, OVERFEED = FALSE, of.a = 5, of.b = 5000,
  SENESCE = FALSE, sns.a = 0.1, sns.b = 100, TATTER = FALSE,
  ttsz.p = 0.5, ttsz.a = 5, ttsz.b = 95, ttr.a = 15, ttr.b = 500,
  bs.m = 30, bs.v = 5, maxBatch = 30, emt.m = 3, emt.v = 1,
  eggT = 0, eggP = 0, energyPreG = 0, PfEIP = 12, Q = 0.9)
}
\arguments{
\item{B_wts}{landing spot weights}

\item{R_wts}{landing spot weights}

\item{O_wts}{landing spot weights}

\item{InAndOut}{matrix of landing probabilities}

\item{B_surv}{Blood feeding bout baseline survival probability \code{\link{mbitesGeneric_surviveFlight}}}

\item{B_succeed}{Blood feeding bout probability of success (if succeed call \code{\link{mbitesGeneric_chooseHost}})}

\item{B_time}{Blood feeding bout mean time elapsed (in days); alternatively the waiting time until next state transition}

\item{surviveH}{human host encounter survival probability of initial encounter (proceed to probe)}

\item{probeH}{human host encounter probability undeterred and begin probing}

\item{surviveprobeH}{human host encounter survival probability of probing}

\item{feedH}{human host encounter probability to successfully blood feed}

\item{surviveZ}{non-human host encounter survival probability of initial encounter (proceed to feed)}

\item{feedZ}{non-human host encounter probability to successfully blood feed}

\item{R_surv}{Post Prandial Resting bout baseline survival probability \code{\link{mbitesGeneric_surviveFlight}}}

\item{R_time}{Post Prandial Resting bout mean time elapsed (in days); alternatively the waiting time until next state transition}

\item{O_surv}{Egg Laying bout baseline survival probability \code{\link{mbitesGeneric_surviveFlight}}}

\item{O_succeed}{Egg Laying bout probability of success (if successfull call \code{layEggs})}

\item{O_time}{Egg Laying bout mean time elapsed (in days); alternatively the waiting time until next state transition}

\item{bm.a}{shape param for bloodmeal size (see \code{\link{mbitesGeneric_rBloodMealSize}})}

\item{bm.b}{shape param for bloodmeal size (see \code{\link{mbitesGeneric_rBloodMealSize}})}

\item{REFEED}{control boolean for refeeding behavior during resting bout}

\item{rf.a}{exp param for refeeding as function of bmSize (see \code{\link{mbitesGeneric_pReFeed}})}

\item{rf.b}{exp param for refeeding as function of bmSize (see \code{\link{mbitesGeneric_pReFeed}})}

\item{OVERFEED}{control boolean for overfeeding during \code{\link{mbitesBRO_BloodMeal}}}

\item{of.a}{exp param for overfeeding as function of bmSize (see \code{\link{mbitesGeneric_pOverFeed}})}

\item{of.b}{exp param for overfeeding as function of bmSize (see \code{\link{mbitesGeneric_pOverFeed}})}

\item{SENESCE}{control boolean for senesce mortality during \code{\link{mbitesGeneric_surviveFlight}}}

\item{sns.a}{exp param for senescence (see \code{\link{mbitesGeneric_pSenesce}})}

\item{sns.b}{exp param for senescence (see \code{\link{mbitesGeneric_pSenesce}})}

\item{TATTER}{control boolean for wing tattering during \code{\link{mbitesGeneric_surviveFlight}}}

\item{ttsz.p}{zero-inflation for tattering damage (see \code{\link{mbitesGeneric_rTatterSize}})}

\item{ttsz.a}{shape param for tattering damage (see \code{\link{mbitesGeneric_rTatterSize}})}

\item{ttsz.b}{shape param for tattering damage (see \code{\link{mbitesGeneric_rTatterSize}})}

\item{ttr.a}{exp param for tattering survival (see \code{\link{mbitesGeneric_pTatter}})}

\item{ttr.b}{exp param for tattering survival (see \code{\link{mbitesGeneric_pTatter}})}

\item{bs.m}{mean of normally-distributed egg batch size (used in \code{\link{mbitesGeneric_rBatchSizeNorm}})}

\item{bs.v}{standard deviation of normally-distributed egg batch size (used in \code{\link{mbitesGeneric_rBatchSizeNorm}})}

\item{maxBatch}{maximum egg batch size (used in \code{\link{mbitesGeneric_rBatchSizeBms}})}

\item{emt.m}{mean of normally-distributed egg batch maturation time (used in \code{\link{mbitesGeneric_rEggMaturationTimeNorm}})}

\item{emt.v}{standard deviation of normally-distributed egg batch maturation time (used in \code{\link{mbitesGeneric_rEggMaturationTimeNorm}})}

\item{eggT}{minimum time to egg maturation}

\item{eggP}{minimum provision to produce eggs}

\item{energyPreG}{pre-gonotrophic energy requirement}

\item{PfEIP}{entomological inoculation period for Plasmodium falciparum during \code{MicroMosquitoFemale$probing()}}

\item{Q}{human blood index (used in \code{\link{mbitesCohort_chooseHost}})}
}
\value{
a named list of parameters
}
\description{
Generate named list of parameters used throughout MICRO/M-BITES. All arguments have default values which are listed below before the definition.
}
\examples{
MBITES.BRO.Parameters()
}
