% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO.R
\name{MbitesBRO.Setup}
\alias{MbitesBRO.Setup}
\title{MBITES-BRO: Initialize Additional Methods & Fields in \code{MicroMosquitoPop} and \code{MicroMosquito}}
\usage{
MbitesBRO.Setup(overwrite = TRUE)
}
\arguments{
\item{overwrite}{overwrite methods}
}
\description{
Initialize M-BITES BRO (Blood Feeding, Resting, Oviposition) lifecycle model.
}
\examples{
MbitesBRO.Setup()
}
