% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO.R
\name{MbitesBRO_getWTS}
\alias{MbitesBRO_getWTS}
\title{MBITES-BRO: Return Landing Spot Weights for \code{MicroMosquitoFemale}}
\usage{
MbitesBRO_getWTS()
}
\value{
vector of landing spot weights
}
\description{
Method for return landing spot weights based on behavioral state of mosquito.
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$getWTS()}.
}
}
