% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO-Bouts.R
\name{mbitesBRO_landingSpot}
\alias{mbitesBRO_landingSpot}
\title{MBITES-BRO: Land After Flight \code{MicroMosquitoFemale}}
\usage{
mbitesBRO_landingSpot()
}
\description{
Mosquito lands after a flight, which may cause various events.
This function always calls \code{\link{mbitesBro_newSpot}} and may call \code{\link{mbitesBro_enterHouse}}
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$landingSpot()}.
}
}
