% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO-Bouts.R
\name{mbitesBro_Pop_MBITES}
\alias{mbitesBro_Pop_MBITES}
\title{MBITES-BRO: Run Simulation for \code{\link{MicroMosquitoPopFemale}}}
\usage{
mbitesBro_Pop_MBITES()
}
\description{
Run the M-BITES life cycle simulation algorithm while alive and has not overrun time in enclosing \code{\link{MicroTile}}.
This method calls \code{\link{mbitesBro_oneMosquito_MBITES}} one each living mosquito sequentially.
\itemize{
\item This method is bound to \code{MicroMosquitoPopFemale$MBITES()}.
}
}
