% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Auxiliary.R
\name{mbitesGeneric_isAlive}
\alias{mbitesGeneric_isAlive}
\title{MBITES-Generic: Alive Check for \code{\link{MicroMosquitoFemale}} and \code{\link{MicroMosquitoMale}}}
\usage{
mbitesGeneric_isAlive()
}
\description{
Check if this mosquito is alive and return a logical value.
\itemize{
\item This method is bound to \code{MicroMosquitoFemale$isAlive()} or \code{MicroMosquitoMale$isAlive()}.
}
}
