% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-EL4P-Calibrate.R
\name{meshK_MicroEL4P}
\alias{meshK_MicroEL4P}
\title{Fit Psi on Mesh of K Values}
\usage{
meshK_MicroEL4P(K.l, K.u, AquaPops, psi.min = 0, psi.max = 10, tMax = 500,
  var.tol = 0.1)
}
\arguments{
\item{K.l}{lower bound of K mesh}

\item{K.u}{upper bound of K mesh}

\item{AquaPops}{list of EL4P objects (see \code{\link{EL4P}})}

\item{psi.min}{lower bound for fitting psi via one dimensional optimization}

\item{psi.max}{upper bound for fitting psi via one dimensional optimization}

\item{tMax}{maximum time to to run aquatic populations to equilibrium}

\item{var.tol}{target minimum variance in lambda for aquatic populations equilibrium}
}
\value{
named list
\itemize{
\item EL4P_pops: aquatic populations run to equilibrium values
\item psi.hat: fitted values of psi
\item K.mesh: sampled values of K
}
}
\description{
Generate a mesh of K values in log space and fit psi to mesh via numerical optization, then run input EL4P populations to equilibrium.
This sets values of psi so that lambda \eqn{\lambda=K} at given parameter values.
}
