% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICRO-Tile-Class.R
\docType{class}
\name{MicroTile}
\alias{MicroTile}
\title{MICRO Tile Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
MicroTile
}
\description{
This is a generic MICRO microsimulation tile blah blah ...
I talk about something here
\itemize{
\item somewhere 1
\item somewhere 2
}
}
\section{Methods}{

\itemize{
\item \strong{Constructor}
\itemize{
\item new: initialize a new \code{MicroTile} object
\itemize{
\item Arguments:
\itemize{
\item \code{MicroTile_PAR}: see \code{\link{MICRO.Tile.Parameters}} for parameter generation function.
}
}
}
\item \strong{Getters & Setters}
\itemize{
\item get_tNow:
\item set_tNow:
\item get_MicroTile_PAR:
\item set_MicroTile_PAR:
\item get_HumanPop:
\item get_Landscape:
\item get_FemalePop:
\item get_MalePop:
}
}
}

\keyword{R6}
\keyword{class}
