% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Class.R
\name{MosquitoFemaleHistory}
\alias{MosquitoFemaleHistory}
\title{\code{\link{MosquitoFemale}} History Object}
\usage{
MosquitoFemaleHistory(stateH = NULL, timeH = NULL, ixH, pSetH,
  feedAllH = 0L, feedAllT = NULL, feedHumanH = 0L, feedHumanT = NULL,
  feedIxH = NULL, bmSizeH = NULL, batchH = NULL)
}
\arguments{
\item{stateH}{state trajectory}

\item{timeH}{transition times}

\item{ixH}{sites visited (initialized to site of emergence)}

\item{pSetH}{point sets visited (initialized to site of emergence)}

\item{feedAllH}{number of blood meals}

\item{feedAllT}{times of blood meals}

\item{feedHumanH}{number of blood meals on human hosts}

\item{feedHumanT}{times of blood meals on human hosts}

\item{feedIxH}{ids of all blood hosts}

\item{bmSizeH}{blood meal sizes}

\item{batchH}{size of egg batch}
}
\value{
list
}
\description{
Generate the history object for \code{\link{MosquitoFemale}}
}
\examples{
MosquitoFemaleHistory(ixH = 0L, pSetH = "l")
}
