% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{MosquitoFemaleHistory}
\alias{MosquitoFemaleHistory}
\title{Mosquito Female History Class}
\usage{
MosquitoFemaleHistory()
}
\description{
Class to store female mosquito histories; \code{\link{MicroMosquitoFemale}} and \code{\link{MicroMosquito}}
}
\section{Methods}{

\itemize{
\item clear_ImagoQ:
\item add_ImagoQ:
\item track_ImagoQ:
\item get_N:
\item set_N:
\item get_ImagoQ:
}
}

\section{Fields}{

\itemize{
\item a field:
}
}

