% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{mosquitoPfMOI}
\alias{mosquitoPfMOI}
\title{PfMOI Mosquito Stage Pathogen Class}
\usage{
mosquitoPfMOI(PfID_init, tInf_init = -1, MOI_init = 0, damID_init = -1,
  sireID_init = -1)
}
\arguments{
\item{PfID_init}{PfID derived from the infecting human's gametocyte ID}

\item{tInf_init}{time of current infection (successful human to mosquito transmission event)}

\item{MOI_init}{initial multiplicity of infection}

\item{damID_init}{female gametocyte 'mother'}

\item{sireID_init}{male gametocyte 'father'}
}
\description{
A class for storing parameters and state variables of mosquito-stage PfMOI infections.
}
\section{Methods}{

\itemize{
\item get_PfID:
\item push_PfID:
\item get_tInf:
\item push_tInf:
\item get_MOI:
\item set_MOI:
\item get_damID:
\item set_damID:
\item get_sireID:
\item set_sireID:
}
}

\section{Fields}{

\itemize{
\item a field: talk about me
}
}

