% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Structures.R
\name{newImago}
\alias{newImago}
\title{Generate a ImagoQ Slot}
\usage{
newImago(N = 0L, tEmerge = 0L, damID = 0L, sireID = 0L,
  genotype = NULL)
}
\arguments{
\item{N}{number of emerging adults}

\item{tEmerge}{time this batch will emerge}

\item{damID}{ID of mother}

\item{sireID}{ID of father}

\item{genotype}{integer genotype identifier}
}
\value{
a named list
}
\description{
Generate a new batch of emerging mosquitoes for a slot in an ImagoQ. It can be called from \code{\link{allocImagoQ}}.
}
\section{Structure}{

\itemize{
\item N number of emerging adults
\item tEmerge time of emergence
\item damID ID of mother
\item sireID ID of father
\item genotype integer genotype identifier
}
}

\examples{
newImago()
}
