% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Micro-Emerge.R
\name{oneDay_MicroEmerge}
\alias{oneDay_MicroEmerge}
\title{MICRO \code{Landscape} Seasonal Emergence}
\usage{
oneDay_MicroEmerge()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Queue the ImagoQ for MICRO Emerge aquatic ecology module.
}
\examples{
some_function()
}
