% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-MosquitoPop-Methods.R
\name{oneDay_RM}
\alias{oneDay_RM}
\title{Set \code{MacroMosquitoPop} oneDay_RM}
\usage{
oneDay_RM()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Daily Ross-Macdonald difference/diffusion equations for mosquito dynamics.
These equations are likely inefficient; see optimizing BLAS/LAPACK and copying less matricies for a faster down the line implementation.
}
\examples{
some_function()
}
