% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfPedigree.R
\name{PfPedigree_init}
\alias{PfPedigree_init}
\title{Initialize PfPedigree to Generation of New Clonal Variants}
\usage{
PfPedigree_init(directory, fileName)
}
\arguments{
\item{directory}{directory that will be created; files will be put in directory/OUTPUT/..}

\item{fileName}{name of the file to write to; directory/OUTPUT/fileName.csv}
}
\value{
a text connection
}
\description{
Generate a text connection to write .csv pedigrees. This connection \strong{must} be named \code{PfPedigreeCon}.
}
\examples{
PfPedigree_init(directory, fileName)
}
