% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Auxiliary.R
\name{plot_PfSI_oneTrajectory}
\alias{plot_PfSI_oneTrajectory}
\title{plot_PfSI_oneTrajectory}
\usage{
plot_PfSI_oneTrajectory(ixH, oneHistory, tMax)
}
\arguments{
\item{Pf_c}{0.15; transmission efficiency: infected human to mosquito}
}
\description{
write me
}
\examples{
plot_PfSI_oneTrajectory()
}
