% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILITIES-Aqua.R
\name{plotEL4P}
\alias{plotEL4P}
\title{Plot Imported EL4P .csv Data}
\usage{
plotEL4P(el4p)
}
\arguments{
\item{el4p}{a data frame (output of importEL4P)}
}
\value{
formatted data object; a list of sites, each site has named elements L1, L2, L3, L4, P
}
\description{
This is a utility to format logged EL4P data after it has been imported by importEL4P
}
\examples{
formatEL4P(el4p)
}
