% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILITIES-Aqua.R
\name{plotEL4P}
\alias{plotEL4P}
\title{Plot Imported EL4P .csv Data}
\usage{
plotEL4P(el4p, egg = NULL)
}
\arguments{
\item{el4p}{a data frame (output of \code{\link{importEL4P}})}

\item{egg}{a data frame (output of \code{\link{importEggQ}}; may be set to NULL)}
}
\value{
plot
}
\description{
This is a utility to plot logged EL4P and EggQ data after it has been imported by importEL4P
}
\examples{
formatEL4P(el4p)
}
