% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-SimBite.R
\name{probeHost_SimBitePfMOI}
\alias{probeHost_SimBitePfMOI}
\title{PfMOI \code{Human} Method: Simulated Host Probing}
\usage{
probeHost_SimBitePfMOI(tBite, mosquitoPfMOI)
}
\arguments{
\item{tBite}{time of bite}

\item{mosquitoPfMOI}{\code{\link{mosquitoPfMOI}} object passed from mosquito to human}
}
\description{
This method is a filler for a host probing event in called in \code{\link{SimBitePfMOI}}.
If the bite is infectious, the method calls \code{\link{infectiousBite_PfMOI}}, otherwise does nothing.
This method is bound to \code{Human$probeHost_PfMOI()} after correct initialization with \code{\link{SimBitePfMOI.Setup}}.
}
