% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Calibration.R
\name{psi2K_cf}
\alias{psi2K_cf}
\title{Regress K on Psi and Return Coefficients}
\usage{
psi2K_cf(meshK, psiHat, plot = FALSE, ...)
}
\arguments{
\item{meshK}{vector of sampled K values}

\item{psiHat}{vector of psi calculated from \link{\code{psiFit}} which provides reasonable initial starting values}

\item{plot}{plot the regression line?}
}
\value{
coefficients of regression
}
\description{
Run a linear regression of K on iverse of psi and extract coefficients; optionally plot the regression. Psi and K should follow a linear relationship, and
the coefficients of the linear regression will give the parameter of the 1-dimensional response surface that relates the two parameters.
}
