% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Calibration.R
\name{psi2K_plot}
\alias{psi2K_plot}
\title{Plot Regression of K on Psi}
\usage{
psi2K_plot(lmFit, K, psi, main = NULL)
}
\arguments{
\item{lmFit}{NULL}

\item{K}{NULL}

\item{psi}{NULL}

\item{main}{NULL}
}
\value{
nothing
}
\description{
Write docs.
}
