% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEARCH-MicroKernel-Setup.R
\name{SEARCH.MicroKernel.Setup}
\alias{SEARCH.MicroKernel.Setup}
\title{MICRO Search Kernels: Initialize Additional Methods & Fields in \code{\link{MicroMosquitoPopFemale}} and \code{\link{MicroMosquitoFemale}}}
\usage{
SEARCH.MicroKernel.Setup(MBITES = "BRO", overwrite = TRUE)
}
\arguments{
\item{MBITES}{what M-BITES module to run?
\itemize{
\item BRO: Blood Feeding, Resting, Oviposition module
\item BROM: Blood Feeding, Resting, Oviposition, Mating module
\item FULL: Full life cycle M-BITES module
}}
}
\value{
does stuff
}
\description{
Initialize MICRO Search Kernels module of mosquito search behavior.
}
\examples{
SEARCH.MicroKernel.Setup()
}
