% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-Macro-PfMethods.R
\name{set_PatchPf}
\alias{set_PatchPf}
\title{Pf: Set \code{MacroPatch} Patch Pf Population}
\usage{
set_PatchPf(PatchPf, ix = NULL)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
get_PatchPf()
}
