% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-SimBite.R
\name{SimBitePfSI.Setup}
\alias{SimBitePfSI.Setup}
\title{Initialize SimBite PfSI Module}
\usage{
SimBitePfSI.Setup(overwrite = TRUE)
}
\arguments{
\item{Pf_c}{0.15; transmission efficiency: infected human to mosquito}

\item{Pf_b}{0.55; transmission efficiency: infected mosquito to human}

\item{DurationPf}{200; duration of infection (How many days does the infection last?)}
}
\value{
Defines a field (list) PfSI_PAR in \code{\link{HumanPop}} and public methods in \code{\link{Human}}
}
\description{
Generate a list of parameters PfSI_PAR in \code{\link{HumanPop}} and public methods in \code{\link{Human}}
}
\examples{
SimBitePfSI.Setup()
}
