% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Biting.R
\name{sumKappa}
\alias{sumKappa}
\title{MACRO: Update \code{Human} kappa For a Patch}
\usage{
sumKappa()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Add my contribution to kappa to current patch (current value of \code{location}).
}
\examples{
some_function()
}
