% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Methods.R
\name{travelHabit}
\alias{travelHabit}
\title{MACRO \code{Human} Method: travelHabit}
\usage{
travelHabit(here, N, n, t = 0)
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Write me!
}
\examples{
some_function()
}
