% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Biting.R
\name{updateKappa}
\alias{updateKappa}
\title{MACRO: Update \code{HumanPop} kappa For all Patches}
\usage{
updateKappa()
}
\arguments{
\item{a}{parameter}
}
\value{
does stuff
}
\description{
Update normalized biting propensities (kappa) for all patches.
}
\examples{
some_function()
}
